% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-elements.r
\name{a_element_blank}
\alias{a_element_blank}
\title{Theme element: blank.
This theme element draws nothing, and assigns no space}
\usage{
a_element_blank()
}
\description{
Theme element: blank.
This theme element draws nothing, and assigns no space
}
