\name{absoluteGrob}
\alias{absoluteGrob}
\alias{grobHeight.absoluteGrob}
\alias{grobWidth.absoluteGrob}
\alias{grobX.absoluteGrob}
\alias{grobY.absoluteGrob}
\alias{grid.draw.absoluteGrob}
\title{Absolute grob}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
This grob has fixed dimesions and position.
}
\usage{absoluteGrob(grob, width = NULL, height = NULL, xmin = NULL, ymin = NULL)}
\arguments{
\item{grob}{}
\item{width}{}
\item{height}{}
\item{xmin}{}
\item{ymin}{}
}

\details{It's still experimental}

\examples{}

