% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/grob-absolute.r
\name{absoluteGrob}
\alias{absoluteGrob}
\title{Absolute grob}
\usage{
absoluteGrob(grob, width = NULL, height = NULL, xmin = NULL,
  ymin = NULL, vp = NULL)
}
\description{
This grob has fixed dimesions and position.
}
\details{
It's still experimental
}
\keyword{internal}

