\name{accessors_print}
\alias{accessors_print}
\title{Write out all convenience accessor functions to R file.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
This calls the \code{accessor_print} method for each of the main
}
\usage{accessors_print(file = "")}
\arguments{
\item{file}{}
}

\details{}

\examples{}
\keyword{internal}
