\name{add_missing_levels}
\alias{add_missing_levels}
\title{Adding missing levels}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Ensure all data frames in list have same levels for selected variables
}
\usage{add_missing_levels(dfs, levels)}
\arguments{
\item{dfs}{}
\item{levels}{}
}




\keyword{internal}
