% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-construction.r
\name{add_ggplot}
\alias{add_ggplot}
\title{Add ggplot-like objects to custom objects}
\usage{
add_ggplot(plot, object, object_name)
}
\arguments{
\item{plot}{The ggplot-like object to add \code{object} to}

\item{object}{An object to add to the plot}

\item{object_name}{The name of the object to add}
}
\value{
A modified ggplot object
}
\description{
This generic allows you to add your own methods for adding objects to
a ggplot-like with \link{+.gg}.
}
\keyword{internal}
