% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.r
\name{add_theme}
\alias{add_theme}
\title{Modify properties of an element in a theme object}
\usage{
add_theme(t1, t2, t2name)
}
\arguments{
\item{t1}{A theme object}

\item{t2}{A theme object that is to be added to \code{t1}}

\item{t2name}{A name of the t2 object. This is used for printing
informative error messages.}
}
\description{
Modify properties of an element in a theme object
}
\seealso{
+.gg
}

