% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-construction.r
\name{add_to_ggplot}
\alias{add_to_ggplot}
\title{Add custom objects to ggplot}
\usage{
add_to_ggplot(object, p, objectname)
}
\arguments{
\item{object}{An object to add to the plot}

\item{p}{The ggplot object to add \code{object} to}

\item{objectname}{The name of the object to add}
}
\value{
A modified ggplot object
}
\description{
This generic allows you to add your own methods for adding custom objects to
a ggplot with \link{+.gg}.
}
\keyword{internal}
