\name{adjust}
\alias{adjust}
\title{Position adjustment occurs over all groups within a geom}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
They work only with discrete x scales and may affect x and y position.
}
\usage{adjust(., data, scales, ...)}
\arguments{
\item{.}{}
\item{data}{}
\item{scales}{}
\item{...}{}
}

\details{}

\examples{}

