\name{aes_all}
\alias{aes_all}
\title{Generate identity mappings}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Given a character vector, create a set of identity mappings
}
\usage{aes_all(vars)}
\arguments{
\item{vars}{vector of variable names}
}

\details{}

\examples{aes_all(names(mtcars))
aes_all(c("x", "y", "col", "pch"))}
\keyword{internal}
