\name{aes}
\alias{aes}
\title{Generate aesthetic mappings}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Aesthetic mappings describe how variables in the data are mapped to visual properties (aesthetics) of geoms.
}
\usage{aes(...)}
\arguments{
\item{...}{List of name value pairs}
}

\details{This function also performs partial name matching, converts color to colour,
and old style R names to new ggplot names (eg. pch to shape, cex to size)}

\examples{}
\keyword{internal}
