\name{aes_all}
\alias{aes_all}
\title{Given a character vector, create a set of identity mappings...}
\usage{aes_all(vars)}

\description{
  Given a character vector, create a set of identity
  mappings
}
\arguments{
  \item{vars}{vector of variable names}
}
\examples{aes_all(names(mtcars))
aes_all(c("x", "y", "col", "pch"))}
