% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aes.r
\name{aes_auto}
\alias{aes_auto}
\title{Automatic aesthetic mapping}
\usage{
aes_auto(data = NULL, ...)
}
\arguments{
\item{data}{data.frame or names of variables}

\item{...}{aesthetics that need to be explicitly mapped.}
}
\description{
Automatic aesthetic mapping
}
\examples{
df <- data.frame(x = 1, y = 1, colour = 1, label = 1, pch = 1)
aes_auto(df)
aes_auto(names(df))

df <- data.frame(xp = 1, y = 1, colour = 1, txt = 1, foo = 1)
aes_auto(df, x = xp, label = txt)
aes_auto(names(df), x = xp, label = txt)

df <- data.frame(foo = 1:3)
aes_auto(df, x = xp, y = yp)
aes_auto(df)
}

