\name{aesdefaults}
\alias{aesdefaults}
\title{Aesthetic defaults}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Convenience method for setting aesthetic defaults
}
\usage{aesdefaults(data, y., params.)}
\arguments{
\item{data}{values from aesthetic mappings}
\item{y.}{defaults}
\item{params.}{user specified values}
}
\value{a data.frame, with all factors converted to character strings}



\keyword{internal}
