\name{all_examples_run}
\alias{all_examples_run}
\title{Run all examples}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Runs all ggplot2 examples
}
\usage{all_examples_run(path=NULL, verbose = TRUE)}
\arguments{
\item{path}{path to save file, if non-NULL}
\item{verbose}{if TRUE, report progress during run}
}

\details{}

\examples{}
\keyword{internal}
