\name{all_rdoc_pages_create}
\alias{all_rdoc_pages_create}
\title{Rebuild all rdoc documentation}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Builds rdoc pages for all ggplot2 objects
}
\usage{all_rdoc_pages_create(path="web/")}
\arguments{
\item{path}{path to save rd files}
}

\details{}

\examples{}
\keyword{internal}
