% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alucard_palette.R
\docType{data}
\name{alucard_tibble}
\alias{alucard_tibble}
\title{Alucard Tibble}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 4 columns.
}
\usage{
alucard_tibble
}
\value{
Alucard Tibble
}
\description{
A Tibble of Alucard data that includes the palette specification.
See https://spec.draculatheme.com for details.
}
\keyword{datasets}
