\name{annotate}
\alias{annotate}
\title{annotate("text", x = 0, y = 0, label = "title")}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{

}
\usage{annotate(geom, x, y, xmin = NULL, xmax = NULL, ymin = NULL, ymax = NULL, ...)}
\arguments{
\item{geom}{}
\item{x}{}
\item{y}{}
\item{xmin}{}
\item{xmax}{}
\item{ymin}{}
\item{ymax}{}
\item{...}{}
}

\details{}

\examples{}

