% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-ticks.R
\name{annotation_ticks}
\alias{annotation_ticks}
\title{Annotation: tick marks}
\usage{
annotation_ticks(sides = "b", base = rep(10, length(sides)),
  scaled = TRUE, short = grid::unit(0.1, "cm"), mid = grid::unit(0.2,
  "cm"), long = grid::unit(0.3, "cm"), colour = "black", size = 0.5,
  linetype = 1, alpha = 1, color = NULL, ticks_per_base = rep(base - 1,
  length(sides)), delog = rep(TRUE, length(sides)), ...)
}
\arguments{
\item{sides}{a string that controls which sides of the plot the log ticks appear on.
It can be set to a string containing any of \code{"trbl"}, for top, right,
bottom, and left.}

\item{base}{the base of the log (default 10)}

\item{scaled}{is the data already log-scaled? This should be \code{TRUE}
(default) when the data is already transformed with \code{log10()} or when
using \code{scale_y_log10}. It should be \code{FALSE} when using
\code{coord_trans(y = "log10")}.}

\item{short}{a \code{\link[grid:unit]{grid::unit()}} object specifying the length of the
short tick marks}

\item{mid}{a \code{\link[grid:unit]{grid::unit()}} object specifying the length of the
middle tick marks. In base 10, these are the "5" ticks.}

\item{long}{a \code{\link[grid:unit]{grid::unit()}} object specifying the length of the
long tick marks. In base 10, these are the "1" (or "10") ticks.}

\item{colour}{Colour of the tick marks.}

\item{size}{Thickness of tick marks, in mm.}

\item{linetype}{Linetype of tick marks (\code{solid}, \code{dashed}, etc.)}

\item{alpha}{The transparency of the tick marks.}

\item{color}{An alias for \code{colour}.}

\item{ticks_per_base}{integer, number of minor ticks between each pair of major ticks, Default: base-1}

\item{delog}{boolean, if an idenity transformation is needed use set to TRUE, Default: FALSE}

\item{...}{Other parameters passed on to the layer}
}
\description{
This annotation adds tick marks to an axis
}
\examples{

p<-ggplot(msleep, aes(bodywt, brainwt)) + geom_point()

#Default behavior
p+annotation_ticks(sides=c('l')) #add identity scale minor ticks on y axis
p+annotation_ticks(sides=c('lb')) #add identity scale minor ticks on x,y axis

#Control number of minor ticks of each side independently
p+annotation_ticks(sides=c('l','b'),ticks_per_base=c(10,5)) #add identity scale minor ticks on xy axis

#log10 scale
p1 <- p + scale_x_log10()
p1+annotation_ticks(sides=c('b'),delog=FALSE) #add minor ticks on log10 scale
p1+annotation_ticks(sides=c('l','b'),base=c(10,10),delog=c(TRUE,FALSE)) #add minor ticks on both scales
p1+annotation_ticks(sides=c('l','b'),base=c(10,10),delog=c(TRUE,TRUE)) #add minor ticks on both scales, but force x axis to be identity

#log scale
p2 <- p + scale_x_continuous(trans='log')
p2+annotation_ticks(sides=c('b'),base=exp(1),delog=FALSE) #add minor ticks on log scale
p2+annotation_ticks(sides=c('l','b'),base=c(10,exp(1)),delog=c(TRUE,FALSE)) #add minor ticks on both scales
p2+annotation_ticks(sides=c('l','b'),base=c(10,exp(1)),delog=c(TRUE,TRUE)) #add minor ticks on both scales, but force x axis to be identity

}
