% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggproto.r
\name{as.list.ggproto}
\alias{as.list.ggproto}
\title{Convert a ggproto object to a list}
\usage{
\method{as.list}{ggproto}(x, inherit = TRUE, ...)
}
\arguments{
\item{x}{A ggproto object to convert to a list.}

\item{inherit}{If \code{TRUE} (the default), flatten all inherited items into
the returned list. If \code{FALSE}, do not include any inherited items.}

\item{...}{Arguments passed on to \code{base::as.list.environment}
\describe{
  \item{all.names}{a logical indicating whether to copy all values or
    (default) only those whose names do not begin with a dot.}
  \item{sorted}{a logical indicating whether the \code{\link{names}} of
    the resulting list should be sorted (increasingly).  Note that this
    is somewhat costly, but may be useful for comparison of environments.}
}}
}
\description{
This will not include the object's \code{super} member.
}
\keyword{internal}
