% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-plyr.R
\name{as.quoted}
\alias{as.quoted}
\title{Create a quoted version of x}
\usage{
as.quoted(x, env = parent.frame())
}
\arguments{
\item{x}{A formula, string, or call to be quoted}

\item{env}{The environment to a attach to the quoted expression.}
}
\description{
This function captures the special meaning of formulas in the context of
facets in ggplot2, where \code{+} have special meaning. It works as
\code{plyr::as.quoted} but only for the special cases of \code{character}, \code{call}, and
\code{formula} input as these are the only situations relevant for ggplot2.
}
\keyword{internal}
