% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/facet-labels.r
\name{as_labeller}
\alias{as_labeller}
\title{Coerce to labeller function}
\usage{
as_labeller(x, default = label_value, multi_line = TRUE)
}
\arguments{
\item{x}{Object to coerce to a labeller function. If a named
character vector, it is used as a lookup table before being
passed on to \code{default}. If a non-labeller function, it is
assumed it takes and returns character vectors and is applied to
the labels. If a labeller, it is simply applied to the labels.}

\item{default}{Default labeller to process the labels produced by
lookup tables or modified by non-labeller functions.}

\item{multi_line}{Whether to display the labels of multiple factors
on separate lines. This is passed to the labeller function.}
}
\description{
This transforms objects to labeller functions. Used internally by
\code{\link{labeller}()}.
}

