\name{assign_viewports}
\alias{assign_viewports}
\title{Assign viewports}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Assign viewports to a matrix of grobs
}
\usage{assign_viewports(grobs)}
\arguments{
\item{grobs}{named matrix of grobs}
}

\details{Uses the structure (and names) of the matrix of grobs, to automatically
assign each grob to the appropriate viewport}


\keyword{internal}
