\name{bin}
\alias{bin}
\title{Bin data}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
This function powers \code{\link{stat_bin}}
}
\usage{bin(x, weight=NULL, binwidth=NULL, origin=NULL, breaks=NULL, range=NULL, width=0.9, drop = FALSE, right = TRUE)}
\arguments{
\item{x}{}
\item{weight}{}
\item{binwidth}{}
\item{origin}{}
\item{breaks}{}
\item{range}{}
\item{width}{}
\item{drop}{}
\item{right}{}
}




\keyword{internal}
