\name{breaks}
\alias{breaks}
\title{Discretise continuous vector}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Method that powers \code{\link{cut_number}} and \code{\link{cut_interval}}
}
\usage{breaks(x, equal, nbins = NULL, binwidth = NULL)}
\arguments{
\item{x}{}
\item{equal}{}
\item{nbins}{}
\item{binwidth}{}
}

\details{}

\examples{}

