% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-contour.r
\name{breaks_default}
\alias{breaks_default}
\title{Default breaks}
\usage{
breaks_default(range, bins, binwidth)
}
\arguments{
\item{range}{The range of the data}

\item{bins}{Number of evenly spaced breaks.}

\item{binwidth}{Distance between breaks.}
}
\description{
Default behaviour for computing breaks in \code{\link[=stat_contour]{stat_contour()}}.
}
