% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guides-.r
\name{build_guides}
\alias{build_guides}
\title{build_guides function}
\usage{
build_guides(scales, layers, default_mapping, position, theme, guides, labels)
}
\arguments{
\item{scales}{...}

\item{layers}{...}

\item{default_mapping}{...}

\item{position}{..}

\item{theme}{...}

\item{guides}{..}

\item{labels}{...}
}
\description{
Describtion Here
}
