\name{calc_element}
\alias{calc_element}
\title{Calculate the element properties, by inheriting properties from its parents}
\usage{
  calc_element(element, theme)
}
\arguments{
  \item{element}{The name of the theme element to
  calculate}

  \item{theme}{A theme object (like theme_grey())}
}
\description{
  Calculate the element properties, by inheriting
  properties from its parents
}

