\name{ceiling_date}
\alias{ceiling_date}
\alias{ceiling_time}
\title{Ceiling for dates and times}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Round date down to nearest multiple of time
}
\usage{ceiling_date(date, time)}
\arguments{
\item{date}{date to round}
\item{time}{unit of time to round to (see \code{\link{cut.Date}}) for valid values}
}




\keyword{internal}
