\name{chop.breaks}
\alias{chop.breaks}
\title{Chop breaks}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Calculate breakpoints for chop function
}
\usage{chop.breaks(x, n, method, midpoint=NA)}
\arguments{
\item{x}{continuous variable}
\item{n}{number of bins to chop into}
\item{method}{method to use: quantiles (approximately equal numbers), cut (equal lengths) or pretty}
\item{midpoint}{mid point for diverging factors}
}

\details{}

\examples{}
\keyword{manip}
\keyword{internal}
