% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\name{class_ggplot}
\alias{class_ggplot}
\title{The ggplot class}
\usage{
class_ggplot(
  data = waiver(),
  layers = list(),
  scales = NULL,
  guides = NULL,
  mapping = aes(),
  theme = NULL,
  coordinates = coord_cartesian(default = TRUE),
  facet = facet_null(),
  layout = NULL,
  labels = labs(),
  meta = list(),
  plot_env = parent.frame()
)
}
\arguments{
\item{data}{A property containing any data coerced by \code{\link[=fortify]{fortify()}}.}

\item{layers}{A list of layer instances created by \code{\link[=layer]{layer()}}.}

\item{scales}{A ScalesList ggproto object.}

\item{guides}{A Guides ggproto object created by \code{\link[=guides]{guides()}}.}

\item{mapping}{A mapping class object created by \code{\link[=aes]{aes()}}.}

\item{theme}{A theme class object created by \code{\link[=theme]{theme()}}.}

\item{coordinates}{A Coord ggproto object created by \verb{coord_*()} family of
functions.}

\item{facet}{A Facet ggproto object created by \verb{facet_*()} family of
functions.}

\item{layout}{A Layout ggproto object.}

\item{labels}{A labels object created by \code{\link[=labs]{labs()}}.}

\item{meta}{A list for additional metadata. This will be deprecated in the
future.}

\item{plot_env}{An environment.}
}
\description{
The ggplot class collects the needed information to render a plot.
This class can be constructed using the \code{\link[=ggplot]{ggplot()}} function.
}
\keyword{internal}
