% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-build.R
\name{class_ggplot_built}
\alias{class_ggplot_built}
\title{The ggplot built class}
\usage{
class_ggplot_built(data = NULL, layout = NULL, plot = NULL)
}
\arguments{
\item{data}{A list of plain data frames; one for each layer.}

\item{layout}{A Layout ggproto object.}

\item{plot}{A completed ggplot class object.}
}
\description{
The ggplot built class is an intermediate class and represents a processed
ggplot object ready for rendering. It is constructed by calling
\code{\link[=ggplot_build]{ggplot_build()}} on a \link[=class_ggplot]{ggplot} object and is not meant to be
instantiated directly. The class can be rendered to a gtable object by
calling the \code{\link[=ggplot_gtable]{ggplot_gtable()}} function on a ggplot built class object.
}
