% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\name{class_labels}
\alias{class_labels}
\title{The labels class}
\usage{
class_labels(labels)
}
\arguments{
\item{labels}{A named list.}
}
\description{
The labels class holds a list with label information to display as titles
of plot components. The preferred way to construct an object of the labels
class is to use the \code{\link[=labs]{labs()}} function.
}
\keyword{internal}
