% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\name{class_theme}
\alias{class_theme}
\title{The theme class}
\usage{
class_theme(elements = list(), ..., complete = FALSE, validate = TRUE)
}
\arguments{
\item{elements}{A named list containing theme elements.}

\item{...}{Reserved for future expansion.}

\item{complete}{A boolean value stating whether a theme is complete.}

\item{validate}{A boolean value stating whether a theme should still be
validated.}
}
\description{
The theme class holds information on how non-data elements of the plot
should be rendered. The preferred way to construct an object of this class
is through the \code{\link[=theme]{theme()}} function.
}
\keyword{internal}
