\name{col2hcl}
\alias{col2hcl}
\title{Modify standard R colour in hcl colour space}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{

}
\usage{col2hcl(colour, h = 0, c = 35, l = 85, alpha = 1)}
\arguments{
\item{colour}{}
\item{h}{}
\item{c}{}
\item{l}{}
\item{alpha}{}
}

\details{}

\examples{}
\keyword{internal}
