\name{collide}
\alias{collide}
\title{Collide}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Detect and prevent collisions
}
\usage{collide(data, width = NULL, name, strategy, check.width = TRUE)}
\arguments{
\item{data}{}
\item{width}{}
\item{name}{}
\item{strategy}{}
\item{check.width}{}
}

\details{Powers dodging, stacking and filling}

\examples{}
\keyword{internal}
