% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{combine_elements}
\alias{combine_elements}
\title{Combine the properties of two elements}
\usage{
combine_elements(child, parent)
}
\arguments{
\item{child}{A child element object, or a property value such as a \code{\link[=rel]{rel()}} or
\code{\link[=margin]{margin()}} object. Can also be \code{NULL} to inherit everything from \code{parent}.}

\item{parent}{A parent element object or property value from which \code{child}
inherits. Can be \code{NULL} if there is nothing to inherit from.}
}
\description{
Resolves an element's properties against a parent element, applying
inheritance for any unspecified properties and resolving deferred values like
\code{\link[=rel]{rel()}} and \code{\link[=margin_part]{margin_part()}}. This is the mechanism used internally by
\code{\link[=calc_element]{calc_element()}} at each step of the element tree traversal.
}
\details{
Note that the similarly named \code{\link[=merge_element]{merge_element()}} behaves differently: If
passed two element objects, it does not resolve properties with deferred
values like \code{\link[=rel]{rel()}} or \code{\link[=margin_part]{margin_part()}}. See the examples.
}
\examples{
child <- element_text(
  colour = "red",
  size = rel(1.5),
  margin = margin_part(t = 10)
)
parent <- element_text(
  family = "mono",
  colour = "blue",
  size = 10,
  margin = margin(1, 2, 3, 4)
)

# compare combine_elements() with merge_element()
combine_elements(child, parent)
merge_element(child, parent)

# pass properties directly
combine_elements(child@size, parent@size)
merge_element(child@size, parent@size)

combine_elements(child@margin, parent@margin)
merge_element(child@margin, parent@margin)
}
\keyword{internal}
