% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-.R
\name{combine_vars}
\alias{combine_vars}
\title{Take input data and define a mapping between faceting variables and ROW,
COL and PANEL keys}
\usage{
combine_vars(data, env = emptyenv(), vars = NULL, drop = TRUE)
}
\arguments{
\item{data}{A list of data.frames, the first being the plot data and the
subsequent individual layer data}

\item{env}{The environment the vars should be evaluated in}

\item{vars}{A list of quoted symbols matching columns in data}

\item{drop}{should missing combinations/levels be dropped}
}
\value{
A data.frame with columns for PANEL, ROW, COL, and faceting vars
}
\description{
Take input data and define a mapping between faceting variables and ROW,
COL and PANEL keys
}
\keyword{internal}
