\name{comma}
\alias{comma}
\title{Comma formatter}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Format number with commas separating thousands
}
\usage{comma(x, ...)}
\arguments{
\item{x}{numeric vector to format}
\item{...}{other arguments passed on to \code{\link{format}}}
}

\details{}

\examples{}

