\name{compare_img}
\alias{compare_img}
\title{Compare two images}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Saves image displaying differences
}
\usage{compare_img(file, path_a, path_b, path_out)}
\arguments{
\item{file}{name of file}
\item{path_a}{location of image a}
\item{path_b}{location of image b}
\item{path_out}{location where output should be saved}
}




\keyword{internal}
