\name{compute_intercept}
\alias{compute_intercept}
\title{Compute intercept from data}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Compute intercept for vline and hline from data and parameters
}
\usage{compute_intercept(data, intercept, var = "x")}
\arguments{
\item{data}{}
\item{intercept}{}
\item{var}{}
}




\keyword{internal}
