% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coord-.r
\name{coord_expand_defaults}
\alias{coord_expand_defaults}
\title{Set the default expand values for the scale, if NA}
\usage{
coord_expand_defaults(coord, scale, aesthetic = NULL)
}
\description{
Set the default expand values for the scale, if NA
}
\keyword{internal}

