% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-zflex.r
\name{coord_flex_cart}
\alias{coord_flex_cart}
\alias{coord_flex_fixed}
\alias{coord_flex_flip}
\title{Cartesian coordinates with flexible options for drawing axes}
\usage{
coord_flex_cart(xlim = NULL, ylim = NULL, expand = TRUE, top = waiver(),
  left = waiver(), bottom = waiver(), right = waiver())

coord_flex_flip(xlim = NULL, ylim = NULL, expand = TRUE, top = waiver(),
  left = waiver(), bottom = waiver(), right = waiver())

coord_flex_fixed(ratio = 1, xlim = NULL, ylim = NULL, expand = TRUE,
  top = waiver(), left = waiver(), bottom = waiver(), right = waiver())
}
\arguments{
\item{xlim}{Limits for the x and y axes.}

\item{ylim}{Limits for the x and y axes.}

\item{expand}{If \code{TRUE}, the default, adds a small expansion factor to
the limits to ensure that data and axes don't overlap. If \code{FALSE},
limits are taken exactly from the data or \code{xlim}/\code{ylim}.}

\item{top, left, bottom, right}{Function for drawing axis lines, ticks, and labels,
use e.g. \code{\link{capped_horisontal}} or \code{\link{brackets_horisontal}}.}

\item{ratio}{aspect ratio, expressed as \code{y / x}.}
}
\description{
Allows user to inject a function for drawing axes, such as
\code{\link{capped_horisontal}} or \code{\link{brackets_horisontal}}.
}
\details{
NB! A panel-border is typically drawn on top such that it covers tick marks,
grid lines, and axis lines.
Many themes also do not draw axis lines.
To ensure the modified axis lines are visible, use
\code{theme(panel.border=element_blank(), axis.lines=element_line())}.
}
\section{User defined functions}{

The provided function in \code{top}, \code{right}, \code{bottom}, and \code{left}
defaults to \code{render_axis} which is defined in \file{coord-.r}, which in
turns calls \code{guide_axis} (see \file{guides-axis.r}).

The provided function is with the arguments
\code{scale_details}, \code{axis}, \code{scale}, \code{position}, and \code{theme},
and the function should return an \code{\link{absoluteGrob}} object.

For examples of modifying the drawn object, see e.g.
\code{\link{capped_horisontal}} or \code{\link{brackets_horisontal}}.
}
\examples{
# ensures that the ranges of axes are equal to the specified ratio by
# adjusting the plot aspect ratio

p <- ggplot(mtcars, aes(mpg, wt)) + geom_point()
p + coord_fixed(ratio = 1)
p + coord_fixed(ratio = 5)
p + coord_fixed(ratio = 1/5)
}

