% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-plyr.R
\name{count}
\alias{count}
\title{Count number of occurences for each unique combination of variables}
\usage{
count(df, vars = NULL, wt_var = NULL)
}
\arguments{
\item{df}{A data.frame}

\item{vars}{A vector of column names. If \code{NULL} all columns in \code{df} will be
used}

\item{wt_var}{The name of a column to use as weight}
}
\value{
A data.frame with the unique combinations counted along with a \code{n}
column giving the counts
}
\description{
Each unique combination of the variables in \code{df} given by \code{vars} will be
identified and their occurences counted. If \code{wt_var} is given the counts will
be weighted by the values in this column.
}
\keyword{internal}
