% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{create_layout}
\alias{create_layout}
\title{create_layout}
\usage{
create_layout(plot)
}
\arguments{
\item{plot}{plotting object}
}
\value{
Layout ggproto object
}
\description{
This S3 method is called within \code{ggplot_build.ggplot} method and
allows package developers to have control over how the Layout object
is made by exporting their own \code{create_layout} method for their subclassed
ggplot object.
}
