\name{cunion}
\alias{cunion}
\title{Col union}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Form the union of columns in a and b.  If there are columns of the same name in both a and b, take the column from a.
}
\usage{cunion(a, b)}
\arguments{
\item{a}{data frame a}
\item{b}{data frame b}
}





