% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cweave}
\alias{cweave}
\title{Col weave}
\usage{
cweave(...)
}
\arguments{
  \item{...}{matrices to weave together}
}
\description{
Weave together two (or more) matrices by column
}
\details{
Matrices must have same dimensions
}
\keyword{internal}

