% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-plyr.R
\name{dapply}
\alias{dapply}
\title{Apply function to unique subsets of a data.frame}
\usage{
dapply(df, by, fun, ..., drop = TRUE)
}
\arguments{
\item{df}{A data.frame}

\item{by}{A character vector of column names to split by}

\item{fun}{A function to apply to each split}

\item{...}{Further arguments to \code{fun}}

\item{drop}{Should unused factor levels in the columns given in \code{by} be
dropped.}
}
\value{
A data.frame if the result of \code{fun} does not include the columns
given in \code{by} these will be prepended to the result.
}
\description{
This function is akin to \code{plyr::ddply}. It takes a single data.frame,
splits it by the unique combinations of the columns given in \code{by}, apply a
function to each split, and then reassembles the results into a sigle
data.frame again.
}
\keyword{internal}
