% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-plyr.R
\name{defaults}
\alias{defaults}
\title{Adds missing elements to a vector from a default vector}
\usage{
defaults(x, y)
}
\arguments{
\item{x, y}{Named vectors or lists}
}
\value{
\code{x} with missing values from \code{y} appended
}
\description{
This function appends a given named vector or list with additional elements
from a default vector, only adding those that does not already exist in the
first.
}
\keyword{internal}
