\name{discrete_range}
\alias{discrete_range}
\title{Calculate range for discrete position variables}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
This is the equivalent of range for discrete variables
}
\usage{discrete_range(..., drop = TRUE)}
\arguments{
\item{...}{}
\item{drop}{}
}

\details{}

\examples{}

