\name{discrete_scale}
\alias{discrete_scale}
\title{Discrete scale constructor.}
\usage{
  discrete_scale(aesthetics, scale_name, palette, name =
  NULL, breaks = NULL, labels = NULL, legend = TRUE, limits
  = NULL, expand = c(0, 0), na.value = NA, drop = TRUE)
}
\description{
  Discrete scale constructor.
}
\keyword{internal}

