\name{dist_central_angle}
\alias{dist_central_angle}
\title{Compute central angle between two points.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Multiple by radius of sphere to get great circle distance
}
\usage{dist_central_angle(lon, lat)}
\arguments{
\item{lon}{}
\item{lat}{}
}





