\name{dollar}
\alias{dollar}
\title{Currency formatter}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Round to nearest cent and display dollar sign
}
\usage{dollar(x, ...)}
\arguments{
\item{x}{numeric vector to format}
\item{...}{other arguments passed on to \code{\link{format}}}
}





