% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legend-draw.r
\name{draw_key}
\alias{draw_key}
\alias{draw_key_abline}
\alias{draw_key_blank}
\alias{draw_key_boxplot}
\alias{draw_key_crossbar}
\alias{draw_key_dotplot}
\alias{draw_key_path}
\alias{draw_key_point}
\alias{draw_key_pointrange}
\alias{draw_key_polygon}
\alias{draw_key_smooth}
\alias{draw_key_text}
\alias{draw_key_vline}
\title{Key drawing functions}
\usage{
draw_key_point(self, data, params)

draw_key_abline(self, data, params)

draw_key_polygon(self, data, params)

draw_key_blank(self, data, params)

draw_key_boxplot(self, data, params)

draw_key_crossbar(self, data, params)

draw_key_path(self, data, params)

draw_key_dotplot(self, data, params)

draw_key_pointrange(self, data, params)

draw_key_smooth(self, data, params)

draw_key_text(self, data, params)

draw_key_vline(self, data, params)
}
\arguments{
\item{self}{The Geom object}

\item{data}{A single row data frame containing the scaled aesthetics to
display in this key}

\item{params}{A list of additional parameters supplied to the geom.}
}
\value{
A grid grob.
}
\description{
Each Geom has an associated function that draws the key when the geom needs
to be displayed in a legend. These are the options built into ggplot2.
}
\keyword{internal}

