\name{drop_calculated_aes}
\alias{drop_calculated_aes}
\title{Remove aesthetics mapped to variables created by statistics}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{

}
\usage{drop_calculated_aes(aesthetics)}
\arguments{
\item{aesthetics}{}
}

\details{}

\examples{}
\keyword{internal}
