\name{economics}
\alias{economics}
\title{US economic time series.}
\usage{data(economics)}

\description{
  US economic time series.
}

\details{
  This dataset was produced from US economic time series
  data available from
  \url{http://research.stlouisfed.org/fred2}.

  \itemize{ \item date.  Month of data collection

  \item psavert, personal savings rate,
  \url{http://research.stlouisfed.org/fred2/series/PSAVERT/}
  \item pce, personal consumption expenditures, in billions
  of dollars,
  \url{http://research.stlouisfed.org/fred2/series/PCE}
  \item unemploy, number of unemployed in thousands,
  \url{http://research.stlouisfed.org/fred2/series/UNEMPLOY}
  \item uempmed, median duration of unemployment, in week,
  \url{http://research.stlouisfed.org/fred2/series/UEMPMED}
  \item pop, total population, in thousands,
  \url{http://research.stlouisfed.org/fred2/series/POP}

  }
}
\docType{data}
\keyword{datasets}

