% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-elements.r
\name{element}
\alias{element}
\alias{element_blank}
\alias{element_line}
\alias{element_rect}
\alias{element_text}
\title{Theme elements}
\usage{
element_blank()

element_rect(fill = NULL, colour = NULL, size = NULL, linetype = NULL,
  color = NULL, inherit.blank = FALSE)

element_line(colour = NULL, size = NULL, linetype = NULL,
  lineend = NULL, color = NULL, arrow = NULL, inherit.blank = FALSE)

element_text(family = NULL, face = NULL, colour = NULL, size = NULL,
  hjust = NULL, vjust = NULL, angle = NULL, lineheight = NULL,
  color = NULL, margin = NULL, debug = NULL, inherit.blank = FALSE)
}
\arguments{
\item{fill}{Fill colour.}

\item{colour, color}{Line/border colour. Color is an alias for colour.}

\item{size}{Line/border size in mm; text size in pts.}

\item{linetype}{Line type. An integer (0:8), a name (blank, solid,
dashed, dotted, dotdash, longdash, twodash), or a string with
an even number (up to eight) of hexadecimal digits which give the
lengths in consecutive positions in the string.}

\item{inherit.blank}{Should this element inherit the existence of an
element_blank among its parents?}

\item{lineend}{Line end Line end style (round, butt, square)}

\item{arrow}{Arrow specification, as created by \code{\link[grid]{arrow}}}

\item{family}{Font family}

\item{face}{Font face ("plain", "italic", "bold", "bold.italic")}

\item{hjust}{Horizontal justification (in [0, 1])}

\item{vjust}{Vertical justification (in [0, 1])}

\item{angle}{Angle (in [0, 360])}

\item{lineheight}{Line height}

\item{margin}{Margins around the text. See \code{\link{margin}} for more
details. When creating a theme, the margins should be placed on the
side of the text facing towards the center of the plot.}

\item{debug}{If \code{TRUE}, aids visual debugging by drawing a solid
rectangle behind the complete text area, and a point where each label
is anchored.}
}
\value{
An S3 object of class \code{element}.
}
\description{
In conjunction with the \link{theme} system, these objects specify the display of
how non-data components of the plot are a drawn.
}
\details{
\itemize{
  \item \code{element_blank}: draws nothing, and assigns no space.
  \item \code{element_rect}: borders and backgrounds.
  \item \code{element_line}: lines.
  \item \code{element_text}: text.
}
}
\examples{
plot <- ggplot(mpg, aes(displ, hwy)) + geom_point()

plot + theme(
  panel.background = element_blank(),
  axis.text = element_blank()
)

plot + theme(
  axis.text = element_text(colour = "red", size = 12)
)

plot + theme(
  axis.line.x = element_line()
)

plot + theme(
  panel.background = element_rect(fill = "grey95"),
  panel.border = element_rect(colour = "grey70", fill = NA, size = 1)
)
}

