% Generated by roxygen2 (4.0.0): do not edit by hand
\name{element_blank}
\alias{element_blank}
\title{Theme element: blank.
This theme element draws nothing, and assigns no space}
\usage{
element_blank()
}
\description{
Theme element: blank.
This theme element draws nothing, and assigns no space
}

