% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-elements.r
\name{element_grob}
\alias{element_grob}
\title{Generate grid grob from theme element}
\usage{
element_grob(element, ...)
}
\arguments{
\item{element}{Theme element, i.e. \code{element_rect} or similar.}

\item{...}{Other arguments to control specific of rendering. This is
usually at least position. See the source code for individual methods.}
}
\description{
Generate grid grob from theme element
}
\keyword{internal}
