\name{element_line}
\alias{element_line}
\title{Theme element: line.}
\usage{
  element_line(colour = NULL, size = NULL, linetype = NULL)
}
\arguments{
  \item{colour}{line color}

  \item{size}{line size}

  \item{linetype}{line type}
}
\description{
  This element draws a line between two (or more) points
}

