\name{element_rect}
\alias{element_rect}
\title{Theme element: rectangle.}
\usage{
  element_rect(fill = NULL, colour = NULL, size = NULL,
    linetype = NULL)
}
\arguments{
  \item{fill}{fill colour}

  \item{colour}{border color}

  \item{size}{border size}

  \item{linetype}{border linetype}
}
\description{
  Most often used for backgrounds and borders.
}

