% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-elements.R
\name{element_render}
\alias{element_render}
\title{Render a specified theme element into a grob}
\usage{
element_render(theme, element, ..., name = NULL)
}
\arguments{
\item{theme}{The theme object}

\item{element}{The element name given as character vector}

\item{...}{Other arguments provided to \code{\link[=element_grob]{element_grob()}}}

\item{name}{Character vector added to the name of the grob}
}
\description{
Given a theme object and element name, returns a grob for the element.
Uses \code{\link[=element_grob]{element_grob()}} to generate the grob.
}
\keyword{internal}
