\name{element_segment}
\alias{element_segment}
\title{Theme element: segments.}
\usage{
  element_segment(colour = NULL, size = NULL,
    linetype = NULL)
}
\arguments{
  \item{colour}{line color}

  \item{size}{line size}

  \item{linetype}{line type}
}
\description{
  This element draws segments between a set of points
}
\seealso{
  \code{link{theme_line}}
}

