\name{element_text}
\alias{element_text}
\title{Theme element: text.}
\usage{
  element_text(family = NULL, face = NULL, colour = NULL,
    size = NULL, hjust = NULL, vjust = NULL, angle = NULL,
    lineheight = NULL)
}
\arguments{
  \item{family}{font family}

  \item{face}{font face ("plain", "italic", "bold")}

  \item{colour}{text colour}

  \item{size}{text size (in pts)}

  \item{hjust}{horizontal justification (in [0, 1])}

  \item{vjust}{vertical justification (in [0, 1])}

  \item{angle}{angle (in [0, 360])}

  \item{lineheight}{line height}
}
\description{
  Theme element: text.
}

