% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{empty}
\alias{empty}
\title{Is a data.frame empty}
\usage{
empty(df)
}
\arguments{
\item{df}{A data.frame or \code{NULL}}
}
\description{
An empty data.frame is defined as either \code{NULL} or a data.frame with zero
rows or columns
}
\keyword{internal}
