\name{equal_ranges}
\alias{equal_ranges}
\title{Compute equal scale ranges}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Used by \code{\link{coord_equal}} to compute equal ranges for equal scale axes
}
\usage{equal_ranges(xlim, ylim, ratio = 1)}
\arguments{
\item{xlim}{x limits}
\item{ylim}{y limits}
\item{ratio}{desired ratio between x and y ranges}
}





