\name{expand_range}
\alias{expand_range}
\title{Expand range}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Convenience function for expanding a range with a multiplicative or additive constant.
}
\usage{expand_range(range, mul=0, add=0)}
\arguments{
\item{range}{range of data}
\item{mul}{multiplicative constract}
\item{add}{additive constant}
}

\details{}

\examples{}
\keyword{manip}
