% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{expand_range4}
\alias{expand_range4}
\title{Similar to expand_range(), but taking a vector ‘expand’
of \emph{four} expansion values, where the 1st and 2nd
elements are used for the lower limit, and the 3rd and
4th elements are used for the upper limit).}
\usage{
expand_range4(limits, expand)
}
\description{
The ‘expand’ argument can also be of length 2,
and the expansion values for the lower limit
are then reused for the upper limit.
}
\keyword{internal}
