% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-absolute.R
\name{facet_absolute}
\alias{facet_absolute}
\title{Facet specification: a single panel.}
\usage{
facet_absolute(shrink = TRUE, y_size = unit(1, "null"),
  x_size = unit(1, "null"), debug = FALSE)
}
\arguments{
\item{shrink}{If \code{TRUE}, will shrink scales to fit output of
statistics, not raw data. If \code{FALSE}, will be range of raw data
before statistical summary.}
}
\description{
Facet specification: a single panel.
}
\examples{
# facet_null is the default faceting specification if you
# don't override it with facet_grid or facet_wrap
ggplot(mtcars, aes(mpg, wt)) + geom_point()
}
\keyword{internal}
