% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{faithfuld}
\alias{faithfuld}
\title{2d density estimate of Old Faithful data}
\format{A data frame with 5,625 observations and 3 variables.}
\usage{
faithfuld
}
\description{
A 2d density estimate of the waiting and eruptions variables data
\link{faithful}.
}
\keyword{datasets}

