% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.r
\name{finalize_plot_theme}
\alias{finalize_plot_theme}
\title{Combine plot defaults with current theme to get complete theme for a plot}
\usage{
finalize_plot_theme(x, default = theme_get(), cache = FALSE)
}
\arguments{
\item{x}{Plot object}

\item{default}{The default theme}

\item{cache}{If \code{TRUE}, precalculates all theme elements based
on their inheritance tree. This can speed up subsequent repeated
lookup of the same theme elements.}
}
\description{
Combine plot defaults with current theme to get complete theme for a plot.
}
\keyword{internal}
