\name{floor_date}
\alias{floor_date}
\alias{floor_time}
\title{Floor for dates and times}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Round date up to nearest multiple of time
}
\usage{floor_date(date, time)}
\arguments{
\item{date}{date to round}
\item{time}{unit of time to round to (see \code{\link{cut.Date}}) for valid values}
}





