% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggproto.r
\name{format.ggproto}
\alias{format.ggproto}
\title{Format a ggproto object}
\usage{
\method{format}{ggproto}(x, ..., flat = TRUE)
}
\arguments{
\item{x}{A ggproto object to print.}

\item{...}{If the ggproto object has a \code{print} method, further arguments
will be passed to it. Otherwise, these arguments are unused.}

\item{flat}{If \code{TRUE} (the default), show a flattened list of all local
and inherited members. If \code{FALSE}, show the inheritance hierarchy.}
}
\description{
Format a ggproto object
}

