% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/proto2.r
\name{format.proto2}
\alias{format.proto2}
\title{Format a proto2 object}
\usage{
\method{format}{proto2}(x, ..., flat = TRUE)
}
\arguments{
\item{flat}{If \code{TRUE} (the default), show a flattened list of all local
  and inherited members. If \code{FALSE}, show the inheritance hierarchy.}
}
\description{
Format a proto2 object
}

