\name{fortify}
\alias{fortify}
\alias{fortify.data.frame}
\alias{fortify.NULL}
\alias{fortify.default}
\title{Fortify a model with data}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Generic method to supplement the original data with model fit statistics
}
\usage{fortify(model, data, ...)}
\arguments{
\item{model}{model}
\item{data}{dataset}
\item{...}{other arguments passed to methods}
}


\seealso{\code{\link{fortify.lm}}}


