% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fortify.list}
\alias{fortify.list}
\title{Fortify method for x, y, z lists.}
\usage{
\method{fortify}{list}(model, data, ...)
}
\arguments{
\item{model}{list with components x, y and z}

\item{data}{not used by this method}

\item{...}{not used by this method}
}
\description{
This function turns a list with components x, y, z, such as those used
by \code{\link[graphics]{image}} and \code{\link[graphics]{image}} into
a data frame that can more easily be plotted with ggplot2.
}
\examples{
x <- seq(0, 1, by=0.1)
y <- seq(0, 2, by=0.1)
z <- outer(x, y, "+")
d <- list(x=x, y=y, z=z)
image(d)
persp(d)
head(fortify(d))
ggplot(d) + geom_tile(aes(x=x, y=y, fill=z))
}

