% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/fortify-raster.r
\name{fortify.raster}
\alias{fortify.RasterBrick}
\alias{fortify.RasterLayer}
\alias{fortify.RasterStack}
\alias{fortify.raster}
\title{Fortify method for classes from the raster package.}
\usage{
\method{fortify}{RasterLayer}(x, maxpixels = 5e+05)

\method{fortify}{RasterBrick}(...)

\method{fortify}{RasterStack}(...)
}
\arguments{
\item{x}{\code{Raster*} object to convert into a dataframe.}

\item{maxpixels}{maximal number of pixels to display}

\item{...}{not used by this method}

\item{data}{not used by this method}
}
\value{
Returns a data.frame with coordinates (x,y) and corresponding raster values.
}
\description{
Fortify method for classes from the raster package.
}
\examples{
if (require("raster")) {
br <- brick(system.file("external/rlogo.grd", package="raster"))
br_df <- fortify(br)
}
}

