\name{fullseq}
\alias{fullseq}
\title{Generate sequence of fixed size intervals covering range}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
All locations are multiples of size
}
\usage{fullseq(range, size)}
\arguments{
\item{range}{range}
\item{size}{interval size}
}

\details{}
\seealso{\code{\link{round_any}}}
\examples{}
\keyword{internal}
