\name{fullseq_date}
\alias{fullseq_date}
\title{Fullseq for dates}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Analog of \code{link{fullseq}}, but for dates
}
\usage{fullseq_date(range, time)}
\arguments{
\item{range}{range of dates}
\item{time}{}
}

\details{Use in \code{\link{scale_date}}}

\examples{}

