% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-abline-hline-vline.R
\name{geom_abline}
\alias{geom_abline}
\alias{geom_hline}
\alias{geom_vline}
\title{Reference lines: horizontal, vertical, and diagonal}
\usage{
geom_abline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  ...,
  slope,
  intercept,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = FALSE
)

geom_hline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  yintercept,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = FALSE
)

geom_vline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  xintercept,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = FALSE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}}.}

\item{data}{The data to be displayed in this layer. There are three options:
\itemize{
\item \code{NULL} (default): the data is inherited from the plot data as specified
in the call to \code{\link[=ggplot]{ggplot()}}.
\item A \code{data.frame}, or other object, will override the plot data. All objects
will be fortified to produce a data frame. See \code{\link[=fortify]{fortify()}} for which
variables will be created.
\item A \code{function} will be called with a single argument, the plot data. The return
value must be a \code{data.frame}, and will be used as the layer data. A
\code{function} can be created from a \code{formula} (e.g. \code{~ head(.x, 10)}).
}}

\item{stat}{The statistical transformation to use on the data for this layer. When using
a \verb{geom_*()} function to construct a layer, the \code{stat} argument can be used
to override the default coupling between geoms and stats. The \code{stat} argument
accepts the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}.
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, to use \code{stat_count()},
give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[=layer_stats]{layer stat} documentation.
}}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}'s \code{params} argument. These arguments
broadly fall into one of 4 categories below. Notably, further arguments to
the \code{position} argument, or aesthetics that are required can
\emph{not} be passed through \code{...}. Unknown arguments that are not part of the 4
categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"} or
\code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics} section
that lists the available options. The 'required' aesthetics cannot be
passed on to the \code{params}. Please note that while passing unmapped
aesthetics as vectors is technically possible, the order and required
length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation lists
which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[=layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[=draw_key]{key glyphs}, to change the display of the layer in the legend.
\item A \code{clip} argument passed on to \code{geom}, one of \code{"on"}, \code{"off"} or
\code{"inherit"} (default). Can be used to escape coord clipping settings. The
default (\code{"inherit"}) lets layer retain the coord's setting.
}}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with a warning. If
\code{TRUE}, missing values are silently removed.}

\item{show.legend}{Logical. Should this layer be included in the legends? \code{NA}, the default,
includes if any aesthetics are mapped. \code{FALSE} never includes, and \code{TRUE}
always includes. It can also be a named logical vector to finely select the
aesthetics to display. To include legend keys for all levels, even when no
data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend, but
unobserved levels are omitted.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather than combining with
them. This is most useful for helper functions that define both data and
aesthetics and shouldn't inherit behaviour from the default plot
specification, e.g. \code{\link[=annotation_borders]{annotation_borders()}}.}

\item{position}{A position adjustment to use on the data for this layer. This can be used in
various ways, including to prevent overplotting and improving the display.
The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example, to
use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[=layer_positions]{layer position} documentation.
}}

\item{xintercept, yintercept, slope, intercept}{Parameters that control the
position of the line. If these are set, \code{data}, \code{mapping} and
\code{show.legend} are overridden.}
}
\description{
These geoms add reference lines (sometimes called rules) to a plot, either
horizontal, vertical, or diagonal (specified by slope and intercept).
These are useful for annotating plots.
}
\details{
These geoms act slightly differently from other geoms. You can supply the
parameters in two ways: either as arguments to the layer function,
or via aesthetics. If you use arguments, e.g.
\code{geom_abline(intercept = 0, slope = 1)}, then behind the scenes
the geom makes a new data frame containing just the data you've supplied.
That means that the lines will be the same in all facets; if you want them
to vary across facets, construct the data frame yourself and use aesthetics.

Unlike most other geoms, these geoms do not inherit aesthetics from the plot
default, because they do not understand x and y aesthetics which are
commonly set in the plot.

Unlike \code{\link[=geom_abline]{geom_abline()}}, \code{\link[=geom_vline]{geom_vline()}} and \code{\link[=geom_hline]{geom_hline()}} include their
intercepts in the position aesthetics. As a result, \code{xintercept} and
\code{yintercept} values can affect the x and y scales, particularly when
intercepts fall outside the range of the data.
}
\section{Aesthetics}{

These geoms are drawn using \code{\link[=geom_line]{geom_line()}} so they support the
same aesthetics: \code{alpha}, \code{colour}, \code{linetype} and
\code{linewidth}. They also each have aesthetics that control the position of
the line:
\itemize{
\item \code{geom_vline()}: \code{xintercept}
\item \code{geom_hline()}: \code{yintercept}
\item \code{geom_abline()}: \code{slope} and \code{intercept}
}
}

\examples{
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()

# Fixed values
p + geom_vline(xintercept = 5)
p + geom_vline(xintercept = 1:5)
p + geom_hline(yintercept = 20)

p + geom_abline() # Can't see it - outside the range of the data
p + geom_abline(intercept = 20)

# Calculate slope and intercept of line of best fit
coef(lm(mpg ~ wt, data = mtcars))
p + geom_abline(intercept = 37, slope = -5)
# But this is easier to do with geom_smooth:
p + geom_smooth(method = "lm", se = FALSE)

# To show different lines in different facets, use aesthetics
p <- ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  facet_wrap(~ cyl)

mean_wt <- data.frame(cyl = c(4, 6, 8), wt = c(2.28, 3.11, 4.00))
p + geom_hline(aes(yintercept = wt), mean_wt)

# You can also control other aesthetics
ggplot(mtcars, aes(mpg, wt, colour = wt)) +
  geom_point() +
  geom_hline(aes(yintercept = wt, colour = wt), mean_wt) +
  facet_wrap(~ cyl)
}
\seealso{
See \code{\link[=geom_segment]{geom_segment()}} for a more general approach to
adding straight line segments to a plot.
}
