% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geom_arc.R
\name{geom_arc}
\alias{geom_arc}
\title{Curved lines.}
\usage{
geom_arc(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", arrow = NULL, shape = 1, lineend = "butt",
  na.rm = FALSE, ...)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link{aes}} or \code{\link{aes_string}}. Only needs to be set
at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override
the plot defaults.}

\item{stat}{The statistical transformation to use on the data for this
layer.}

\item{position}{The position adjustment to use for overlapping points
on this layer}

\item{arrow}{specification for arrow heads, as created by arrow()}

\item{lineend}{Line end style (round, butt, square)}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with
a warning.  If \code{TRUE} silently removes missing values.}

\item{...}{other arguments passed on to \code{\link{layer}}. This can
include aesthetics whose values you want to set, not map. See
\code{\link{layer}} for more details.}
}
\description{
Curved lines.
}
\section{Aesthetics}{

\Sexpr[results=rd,stage=build]{ggplot2:::rd_aesthetics("geom", "arc")}
}
\examples{
library(grid) # needed for arrow function

# Adding arcs
library(grid) # needed for arrow function
b <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
b + geom_arc(aes(x = 2, y = 15, xend = 2, yend = 25))
b + geom_arc(aes(x = 2, y = 15, xend = 3, yend = 15))
b + geom_arc(aes(x = 5, y = 30, xend = 3.5, yend = 25),
   arrow = arrow(length = unit(0.5, "cm")))
}

