% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-col.r
\name{geom_col}
\alias{geom_col}
\title{Bars, rectangles with bases on x-axis}
\usage{
geom_col(mapping = NULL, data = NULL, position = "stack", ...,
  width = NULL, na.rm = FALSE, show.legend = NA, inherit.aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{other arguments passed on to \code{\link{layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{width}{Bar width. By default, set to 90\% of the resolution of the data.}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with
a warning.  If \code{TRUE} silently removes missing values.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
Can also be a named logical vector to specify which aesthetics to include.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}
}
\description{
This is an alternate version of \code{geom_bar} that uses
\code{stat="identity"} (i.e. the heights of the bars represent values in the
data) by default, requiring the use of a \code{y} aesthetic.
If you need \code{stat="count"} functionality, use \code{geom_bar}.
}
\details{
A bar chart maps the height of the bar to a variable, and so the base of the
bar must always be shown to produce a valid visual comparison. Naomi Robbins
has a nice
\href{http://www.b-eye-network.com/view/index.php?cid=2468}{article on this
topic}. This is why it doesn't make sense to use a log-scaled y axis with a
bar chart.

By default, multiple x's occurring in the same place will be stacked atop one
another by \code{\link{position_stack}}. If you want them to be dodged
side-to-side, see \code{\link{position_dodge}}. Finally,
\code{\link{position_fill}} shows relative proportions at each x by stacking
the bars and then stretching or squashing to the same height.
}
\section{Aesthetics}{

  \Sexpr[results=rd,stage=build]{ggplot2:::rd_aesthetics("geom", "col")}
}
\examples{
df <- data.frame(trt = c("a", "b", "c"), outcome = c(2.3, 1.9, 3.2))
ggplot(df, aes(trt, outcome)) +
  geom_col()
# But geom_point() displays exactly the same information and doesn't
# require the y-axis to touch zero.
ggplot(df, aes(trt, outcome)) +
  geom_point()
}
\seealso{
\code{\link{geom_bar}} to make the height of the bar proportional to
 the number of cases in each group or sum of weights,
  \code{\link{geom_histogram}} for continuous data,
  \code{\link{position_dodge}} for creating side-by-side barcharts.
}

