% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom-curve.r
\name{geom_curve}
\alias{geom_curve}
\title{Single curved line segments.}
\usage{
geom_curve(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", arrow = NULL, lineend = "butt", na.rm = FALSE,
  ...)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link{aes}} or \code{\link{aes_string}}. Only needs to be set
at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override
the plot defaults.}

\item{stat}{The statistical transformation to use on the data for this
layer.}

\item{position}{The position adjustment to use for overlapping points
on this layer}

\item{arrow}{specification for arrow heads, as created by arrow()}

\item{lineend}{Line end style (round, butt, square)}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with
a warning.  If \code{TRUE} silently removes missing values.}

\item{...}{other arguments passed on to \code{\link{layer}}. This can
include aesthetics whose values you want to set, not map. See
\code{\link{layer}} for more details.}

\item{curvature}{see curveGrob}

\item{angle}{see curveGrob}

\item{ncp}{see curveGrob}
}
\description{
Single curved line segments.
}
\section{Aesthetics}{

\Sexpr[results=rd,stage=build]{ggplot2:::rd_aesthetics("geom", "curve")}
}
\examples{
# Adding curve segments
library(grid) # needed for arrow function
b <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
b + geom_curve(aes(x = 2, y = 15, xend = 2, yend = 25, curvature = 0.2))
b + geom_curve(aes(x = 2, y = 15, xend = 3, yend = 15, ncp = 2))
b + geom_curve(aes(x = 5, y = 30, xend = 3.5, yend = 25), arrow = arrow(length = unit(0.5, "cm")))
}
\seealso{
\code{\link{geom_segment}}, \code{\link{geom_path}} and \code{\link{geom_line}} for multi-
segment lines and paths.
}

