\name{geom_errorbarh}
\alias{geom_errorbarh}
\alias{GeomErrorbarh}
\title{geom\_errorbarh}
\description{Horizontal error bars}
\details{
This page describes geom\_errorbarh, see \code{\link{layer}} and \code{\link{qplot}} for how to create a complete plot from individual components.
}
\section{Aesthetics}{
The following aesthetics can be used with geom\_errorbarh.  Aesthetics are mapped to variables in the data with the aes function: \code{geom\_errorbarh(aes(x = var))}
\itemize{
  \item \code{x}: x position (\strong{required}) 
  \item \code{xmin}: left (hortizontal minimum) (\strong{required}) 
  \item \code{xmax}: right (hortizontal maximum) (\strong{required}) 
  \item \code{colour}: border colour 
  \item \code{size}: size 
  \item \code{linetype}: line type 
  \item \code{width}: width 
  \item \code{alpha}: transparency 
}
}
\usage{geom_errorbarh(mapping=NULL, data=NULL, stat="identity", position="identity", ...)}
\arguments{
 \item{mapping}{mapping between variables and aesthetics generated by aes}
 \item{data}{dataset used in this layer, if not specified uses plot dataset}
 \item{stat}{statistic used by this layer}
 \item{position}{position adjustment used by this layer}
 \item{...}{ignored }
}
\seealso{\itemize{
  \item \code{\link{geom_errorbar}}: vertical error bars
  \item \url{http://had.co.nz/ggplot2/geom_errorbarh.html}
}}
\value{A \code{\link{layer}}}
\examples{\dontrun{
df <- data.frame(
  trt = factor(c(1, 1, 2, 2)), 
  resp = c(1, 5, 3, 4), 
  group = factor(c(1, 2, 1, 2)), 
  se = c(0.1, 0.3, 0.3, 0.2)
)

# Define the top and bottom of the errorbars

p <- ggplot(df, aes(resp, trt, fill = group))
p + geom_point() +
  geom_errorbarh(aes(xmax = resp + se, xmin = resp - se))
  
last_plot() + aes(colour = group)
  
}}
\author{Hadley Wickham, \url{http://had.co.nz/}}
\keyword{hplot}
