% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/geom-km.r
\name{geom_km}
\alias{geom_km}
\title{Add a Kaplan-Meier survival curve}
\usage{
geom_km(mapping = NULL, data = NULL, stat = "km", position = "identity",
  ...)
}
\description{
Add a Kaplan-Meier survival curve
}
\section{Aesthetics}{

\Sexpr[results=rd,stage=build]{ggplot2:::rd_aesthetics("geom", "km")}
}
\examples{
df <- survival::lung
ggplot(df, aes(x = time, status = status, color = factor(sex))) + geom_km()
qplot(time, status = status, data = df, geom = "km")
}
\seealso{
The default stat for this geom is \code{\link{stat_km}} see
  that documentation for more options to control the underlying statistical transformation.
}

