% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-shadowpath.r
\name{geom_shadowpath}
\alias{geom_shadowpath}
\title{Connect observations, plot a shadow beneath the connected lines to make it easier to read
a chart with several overlaping observations.}
\usage{
geom_shadowpath(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  lineend = "butt",
  linejoin = "round",
  linemitre = 10,
  arrow = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{lineend}{Line end style (round, butt, square).}

\item{linejoin}{Line join style (round, mitre, bevel).}

\item{linemitre}{Line mitre limit (number greater than 1).}

\item{arrow}{Arrow specification, as created by \code{\link[grid:arrow]{grid::arrow()}}.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}
}
\description{
\code{geom_shadowpath()} connects the observations in the order in which they appear
in the data. \code{geom_shadowline()} connects them in order of the variable on the
x axis. \code{geom_shadowstep()} creates a stairstep plot, highlighting exactly
when changes occur. The \code{group} aesthetic determines which cases are
connected together. These functions are designed as a straight replacement
to the \code{\link[=geom_path]{geom_path()}}, \code{\link[=geom_line]{geom_line()}} and \code{\link[=geom_step]{geom_step()}} functions.
}
\details{
To set the order of drawing, make the \code{colour} aesthetic a factor, and set the order
from bottom to top.
}
\section{Orientation}{

This geom treats each axis differently and, thus, can thus have two orientations. Often the orientation is easy to deduce from a combination of the given mappings and the types of positional scales in use. Thus, ggplot2 will by default try to guess which orientation the layer should have. Under rare circumstances, the orientation is ambiguous and guessing may fail. In that case the orientation can be specified directly using the \code{orientation} parameter, which can be either \code{"x"} or \code{"y"}. The value gives the axis that the geom should run along, \code{"x"} being the default orientation you would expect for the geom.
}

\section{Aesthetics}{

\code{geom_path()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
\item \strong{\code{x}}
\item \strong{\code{y}}
\item \code{alpha}
\item \code{colour}
\item \code{group}
\item \code{linetype}
\item \code{size}
}
Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.
}

\section{Missing value handling}{

\code{geom_path()}, \code{geom_line()}, and \code{geom_step} handle \code{NA} as follows:
\itemize{
\item If an \code{NA} occurs in the middle of a line, it breaks the line. No warning
is shown, regardless of whether \code{na.rm} is \code{TRUE} or \code{FALSE}.
\item If an \code{NA} occurs at the start or the end of the line and \code{na.rm} is \code{FALSE}
(default), the \code{NA} is removed with a warning.
\item If an \code{NA} occurs at the start or the end of the line and \code{na.rm} is \code{TRUE},
the \code{NA} is removed silently, without warning.
}
}

\examples{
# geom_shadowline() is suitable for time series
ggplot(economics, aes(date, unemploy)) + geom_shadowline()
ggplot(economics_long, aes(date, value01, colour = variable)) +
  geom_shadowline()

}
\seealso{
\code{\link[=geom_path]{geom_path()}}, \code{\link[=geom_line]{geom_line()}}, \code{\link[=geom_step]{geom_step()}}: Filled paths (polygons);
}
