% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-raster.R, R/geom-rect.R, R/geom-tile.R
\name{geom_raster}
\alias{geom_raster}
\alias{geom_rect}
\alias{geom_tile}
\title{Rectangles}
\usage{
geom_raster(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  interpolate = FALSE,
  hjust = 0.5,
  vjust = 0.5,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_rect(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  lineend = "butt",
  linejoin = "mitre",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_tile(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  lineend = "butt",
  linejoin = "mitre",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}}. If specified and \code{inherit.aes = TRUE} (the default), it is combined with the default mapping at the top level
of the plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three options:
\itemize{
\item \code{NULL} (default): the data is inherited from the plot data as specified
in the call to \code{\link[=ggplot]{ggplot()}}.
\item A \code{data.frame}, or other object, will override the plot data. All objects
will be fortified to produce a data frame. See \code{\link[=fortify]{fortify()}} for which
variables will be created.
\item A \code{function} will be called with a single argument, the plot data. The return
value must be a \code{data.frame}, and will be used as the layer data. A
\code{function} can be created from a \code{formula} (e.g. \code{~ head(.x, 10)}).
}}

\item{stat}{The statistical transformation to use on the data for this layer. When using
a \verb{geom_*()} function to construct a layer, the \code{stat} argument can be used
to override the default coupling between geoms and stats. The \code{stat} argument
accepts the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}.
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, to use \code{stat_count()},
give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[=layer_stats]{layer stat} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This can be used in
various ways, including to prevent overplotting and improving the display.
The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example, to
use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[=layer_positions]{layer position} documentation.
}}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}'s \code{params} argument. These arguments
broadly fall into one of 4 categories below. Notably, further arguments to
the \code{position} argument, or aesthetics that are required can
\emph{not} be passed through \code{...}. Unknown arguments that are not part of the 4
categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"} or
\code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics} section
that lists the available options. The 'required' aesthetics cannot be
passed on to the \code{params}. Please note that while passing unmapped
aesthetics as vectors is technically possible, the order and required
length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation lists
which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[=layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[=draw_key]{key glyphs}, to change the display of the layer in the legend.
\item A \code{clip} argument passed on to \code{geom}, one of \code{"on"}, \code{"off"} or
\code{"inherit"} (default). Can be used to escape coord clipping settings. The
default (\code{"inherit"}) lets layer retain the coord's setting.
}}

\item{interpolate}{If \code{TRUE} interpolate linearly, if \code{FALSE}
(the default) don't interpolate.}

\item{hjust, vjust}{horizontal and vertical justification of the grob.  Each
justification value should be a number between 0 and 1.  Defaults to 0.5
for both, centering each pixel over its data location.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with a warning. If
\code{TRUE}, missing values are silently removed.}

\item{show.legend}{Logical. Should this layer be included in the legends? \code{NA}, the default,
includes if any aesthetics are mapped. \code{FALSE} never includes, and \code{TRUE}
always includes. It can also be a named logical vector to finely select the
aesthetics to display. To include legend keys for all levels, even when no
data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend, but
unobserved levels are omitted.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather than combining with
them. This is most useful for helper functions that define both data and
aesthetics and shouldn't inherit behaviour from the default plot
specification, e.g. \code{\link[=annotation_borders]{annotation_borders()}}.}

\item{lineend}{Line end style, one of \code{"round"}, \code{"butt"} or \code{"square"}.}

\item{linejoin}{Line join style, one of \code{"round"}, \code{"mitre"} or \code{"bevel"}.}
}
\description{
\code{geom_rect()} and \code{geom_tile()} do the same thing, but are
parameterised differently: \code{geom_tile()} uses the center of the tile and its
size (\code{x}, \code{y}, \code{width}, \code{height}), while \code{geom_rect()} can use those or the
locations of the corners (\code{xmin}, \code{xmax}, \code{ymin} and \code{ymax}).
\code{geom_raster()} is a high performance special case for when all the tiles
are the same size, and no pattern fills are applied.
}
\details{
Please note that the \code{width} and \code{height} aesthetics are not true position
aesthetics and therefore are not subject to scale transformation. It is
only after transformation that these aesthetics are applied.
}
\examples{
# The most common use for rectangles is to draw a surface. You always want
# to use geom_raster here because it's so much faster, and produces
# smaller output when saving to PDF
ggplot(faithfuld, aes(waiting, eruptions)) +
 geom_raster(aes(fill = density))

# Interpolation smooths the surface & is most helpful when rendering images.
ggplot(faithfuld, aes(waiting, eruptions)) +
 geom_raster(aes(fill = density), interpolate = TRUE)

# If you want to draw arbitrary rectangles, use geom_tile() or geom_rect()
df <- data.frame(
  x = rep(c(2, 5, 7, 9, 12), 2),
  y = rep(c(1, 2), each = 5),
  z = factor(rep(1:5, each = 2)),
  w = rep(diff(c(0, 4, 6, 8, 10, 14)), 2)
)
ggplot(df, aes(x, y)) +
  geom_tile(aes(fill = z), colour = "grey50")
ggplot(df, aes(x, y, width = w)) +
  geom_tile(aes(fill = z), colour = "grey50")
ggplot(df, aes(xmin = x - w / 2, xmax = x + w / 2, ymin = y, ymax = y + 1)) +
  geom_rect(aes(fill = z), colour = "grey50")

\donttest{
# Justification controls where the cells are anchored
df <- expand.grid(x = 0:5, y = 0:5)
set.seed(1)
df$z <- runif(nrow(df))
# default is compatible with geom_tile()
ggplot(df, aes(x, y, fill = z)) +
  geom_raster()
# zero padding
ggplot(df, aes(x, y, fill = z)) +
  geom_raster(hjust = 0, vjust = 0)

# Inspired by the image-density plots of Ken Knoblauch
cars <- ggplot(mtcars, aes(mpg, factor(cyl)))
cars + geom_point()
cars + stat_bin_2d(aes(fill = after_stat(count)), binwidth = c(3,1))
cars + stat_bin_2d(aes(fill = after_stat(density)), binwidth = c(3,1))

cars +
  stat_density(
    aes(fill = after_stat(density)),
    geom = "raster",
    position = "identity"
   )
cars +
  stat_density(
    aes(fill = after_stat(count)),
    geom = "raster",
    position = "identity"
  )
}
}
\section{Aesthetics}{

\code{geom_rect()} understands the following aesthetics. Required aesthetics are displayed in bold and defaults are displayed for optional aesthetics:
\tabular{rll}{
 • \tab \strong{\code{\link[=aes_position]{x}} \emph{or} \code{width} \emph{or} \code{\link[=aes_position]{xmin}} \emph{or} \code{\link[=aes_position]{xmax}}} \tab   \cr
 • \tab \strong{\code{\link[=aes_position]{y}} \emph{or} \code{height} \emph{or} \code{\link[=aes_position]{ymin}} \emph{or} \code{\link[=aes_position]{ymax}}} \tab   \cr
 • \tab \code{\link[=aes_colour_fill_alpha]{alpha}} \tab → \code{NA} \cr
 • \tab \code{\link[=aes_colour_fill_alpha]{colour}} \tab → via \code{theme()} \cr
 • \tab \code{\link[=aes_colour_fill_alpha]{fill}} \tab → via \code{theme()} \cr
 • \tab \code{\link[=aes_group_order]{group}} \tab → inferred \cr
 • \tab \code{\link[=aes_linetype_size_shape]{linetype}} \tab → via \code{theme()} \cr
 • \tab \code{\link[=aes_linetype_size_shape]{linewidth}} \tab → via \code{theme()} \cr
}
\code{geom_tile()} understands only the \code{x}/\code{width} and \code{y}/\code{height} combinations.
Note that \code{geom_raster()} ignores \code{colour}.

Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.
}
