% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.r
\name{get_alt_text}
\alias{get_alt_text}
\title{Extract an alt text from a plot}
\usage{
get_alt_text(p)
}
\arguments{
\item{p}{a ggplot object}
}
\value{
A text string
}
\description{
This function returns a text that can be used as alt-text in webpages etc.
It will either uses the \code{alt} label, added with \verb{+ labels(alt = <...>)}, or
synthesize one from the information in the plot itself.
}
\keyword{internal}
