% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-last.R
\name{get_last_plot}
\alias{get_last_plot}
\alias{last_plot}
\title{Retrieve the last plot to be modified or created.}
\usage{
get_last_plot()

last_plot()
}
\description{
Retrieve the last plot to be modified or created.
}
\seealso{
\code{\link[=ggsave]{ggsave()}}
}
\keyword{internal}
