% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-construction.r, R/theme.r
\name{+.gg}
\alias{+.gg}
\alias{\%+\%}
\alias{\%+replace\%}
\title{Add a new component to a ggplot or theme object.}
\usage{
\method{+}{gg}(e1, e2)

e1 \%+\% e2

e1 \%+replace\% e2
}
\arguments{
\item{e1}{An object of class \code{\link{ggplot}} or a \code{\link{theme}}.}

\item{e2}{A component, or list of components to add to \code{e1}.}
}
\description{
This operator allows you to add objects to a ggplot or theme object.
}
\section{Adding on to a ggplot object}{

You can add any of the following types of objects:

\itemize{
  \item \code{uneval}: replace current aesthetics
  \item \code{layer}: add new layer
  \item \code{theme}: update plot theme
  \item \code{scale}: replace current scale
  \item \code{coord}: override current coordinate system
  \item \code{facet}: override current coordinate faceting
  \item \code{list}: a list of any of the above.
}

To replace the current default data frame, you must use \code{\%+\%},
due to S3 method precedence issues.
}

\section{Adding on to a theme}{


\code{+} and \code{\%+replace\%} can be used to modify elements in themes.

\code{+} updates the elements of e1 that differ from elements specified (not
NULL) in e2. Thus this operator can be used to incrementally add or modify
attributes of a ggplot theme.

In contrast, \code{\%+replace\%} replaces the entire element; any element of
a theme not specified in e2 will not be present in the resulting theme (i.e.
NULL). Thus this operator can be used to overwrite an entire theme.
}
\examples{
# Adding on to a plot -----------------------------------------------

base <- ggplot(mpg, aes(displ, hwy)) + geom_point()
base + geom_smooth()

# To override the data, you must use \%+\%
base \%+\% subset(mpg, fl == "p")

# Alternatively, you can add multiple components with a list.
# This can be useful to return from a list.
base + list(subset(mpg, fl == "p"), geom_smooth())

# Adding on to a theme ----------------------------------------------

# Compare these results of adding theme objects to other theme objects
add_el <- theme_grey() + theme(text = element_text(family = "Times"))
rep_el <- theme_grey() \%+replace\% theme(text = element_text(family = "Times"))

add_el$text
rep_el$text
}
\seealso{
\code{\link{theme}}
}

