% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{gg_dep}
\alias{gg_dep}
\title{Give a deprecation error, warning, or message, depending on version number.}
\usage{
gg_dep(version, msg)
}
\arguments{
\item{version}{The last version of ggplot2 where this function was good
(in other words, the last version where it was not deprecated).}

\item{msg}{The message to print.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
