\name{ggaxis}
\alias{ggaxis}
\title{Grob axis}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Grob for axes
}
\usage{ggaxis(at, labels, position="right", scale=c(0,1))}
\arguments{
\item{at}{position of ticks}
\item{labels}{labels at ticks}
\item{position}{position of axis (top, bottom, left or right)}
\item{scale}{range of data values}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
