\name{ggaxis_labels}
\alias{ggaxis_labels}
\title{Grob axis labels}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Grob for axis lables
}
\usage{ggaxis_labels(at, labels, position)}
\arguments{
\item{at}{position of ticks}
\item{labels}{grob labels at ticks}
\item{position}{position of axis (top, bottom, left or right)}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
