\name{ggaxis_line}
\alias{ggaxis_line}
\title{Grob axis line}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Grob for axis baseline
}
\usage{ggaxis_line(at, position)}
\arguments{
\item{at}{position of ticks}
\item{position}{position of axis (top, bottom, left or right)}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
