\name{ggaxis_ticks}
\alias{ggaxis_ticks}
\title{Grob axis ticks}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Grob for axis ticks
}
\usage{ggaxis_ticks(at, position)}
\arguments{
\item{at}{position of ticks}
\item{position}{position of axis (top, bottom, left or right)}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
