\name{ggaxis_vp}
\alias{ggaxis_vp}
\title{Grob axis viewport}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Generate viewport for axis grobs
}
\usage{ggaxis_vp(position, labels, scale=c(0,1))}
\arguments{
\item{position}{position of axis (top, bottom, left or right)}
\item{labels}{labels at ticks}
\item{scale}{range of data values}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
