\name{gglegend}
\alias{gglegend}
\title{Build a legend grob}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Build the grob for a single legend.
}
\usage{gglegend(legend, usage=usage)}
\arguments{
\item{legend}{}
\item{usage}{}
}
\value{A grid grob}
\details{}

\examples{}
\keyword{internal}
