% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities-grid.r
\name{ggname}
\alias{ggname}
\title{Name ggplot grid object}
\usage{
ggname(prefix, grob)
}
\arguments{
\item{prefix}{New name for \code{grob}.}

\item{grob}{A grob to name.}
}
\description{
Convenience function to name grid objects
}
\keyword{internal}

