\name{ggorder}
\alias{ggorder}
\title{Order plot}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
A plot to investigate the order in which observations were recorded.
}
\usage{ggorder(data, scale="rank")}
\arguments{
\item{data}{data set to plot}
\item{scale}{type of scaling to use.  See \code{\link[reshape]{rescaler}} for options}
}




\keyword{hplot}
