\name{ggplot_build}
\alias{ggplot_build}
\title{Build ggplot for rendering}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
This function is the powerhouse that converts the plot specification into something that's ready to be rendered on screen
}
\usage{ggplot_build(plot)}
\arguments{
\item{plot}{}
}





