\name{ggplot.data.frame}
\alias{ggplot.data.frame}
\alias{package-ggplot}
\alias{package-ggplot}
\title{Create a new plot}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Create a new ggplot plot
}
\usage{ggplot.data.frame(data, mapping=aes(), ..., environment = globalenv())}
\arguments{
\item{data}{default data frame}
\item{mapping}{default list of aesthetic mappings (these can be colour, size, shape, line type -- see individual geom functions for more details)}
\item{...}{}
\item{environment}{}
}

\details{}
\seealso{\url{http://had.co.nz/ggplot2}}
\examples{}
\keyword{hplot}
