\name{ggplot}
\alias{ggplot}
\alias{ggplot.default}
\title{Create a new plot}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Create a new ggplot plot
}
\usage{ggplot(data = NULL, ...)}
\arguments{
\item{data}{default data set}
\item{...}{other arguments passed to specific methods}
}


\seealso{\url{http://had.co.nz/ggplot2}}

\keyword{hplot}
