% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ggplot.data.frame}
\alias{ggplot.data.frame}
\title{Create a new ggplot plot from a data frame}
\usage{
\method{ggplot}{data.frame}(data, mapping = aes(), ...,
  environment = parent.frame())
}
\arguments{
\item{data}{default data frame for plot}

\item{mapping}{default list of aesthetic mappings (these can be colour,
size, shape, line type -- see individual geom functions for more details)}

\item{...}{ignored}

\item{environment}{in which evaluation of aesthetics should occur}
}
\description{
Create a new ggplot plot from a data frame
}
\seealso{
\url{http://had.co.nz/ggplot2}
}

