% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities-grid.r
\name{ggplot2-grid}
\alias{arrow}
\alias{ggplot2-grid}
\alias{unit}
\title{unit & arrow functions from grid.}
\description{
See \code{\link[grid]{arrow}} and \code{\link[grid]{unit}} for more details.
These functions are re-exported from grid by ggplot2 since they are
so commonly used.
}
\keyword{internal}

