% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guides-.r
\name{ggplot2-guide}
\alias{ggplot2-guide}
\alias{guide_gengrob}
\alias{guide_geom}
\alias{guide_merge}
\alias{guide_train}
\title{Guide generic functions for ggplot2}
\usage{
guide_train(guide, scale)

guide_merge(guide, new_guide)

guide_geom(guide, layers, default_mapping)

guide_gengrob(guide, theme)
}
\arguments{
\item{guide}{A guide object.}

\item{scale}{A \code{\link{Scale}} ggproto defining how aesthetics are mapped
to layers.}

\item{new_guide}{A guide object to merge with \code{guide}.}

\item{layers}{List of layers in the plot}

\item{default_mapping}{Top-level aesthetic mappings for the plot.}

\item{theme}{The \code{\link{theme}} object for the plot.}
}
\description{
Extending ggplot2 with new guides requires the implementation of an S3 class
with methods for these generic functions.
}
\details{
\code{guide_train}: Configure the guide with aesthetic mapping,
  range, breaks, and other information provided by the Scale. Returns the
  modified guide object including a \code{hash} of the relevant properties.

\code{guide_merge}: When two guides have identical \code{hash}
  objects, merge them into a single guide. Returns the merged guide object.

\code{guide_geom}: Gather geom information from layers needed to
  later render the guide. Returns the modified guide object.

\code{guide_gengrob}: Render the guide for display. Returns a
  \code{\link[gtable]{gtable}}.
}
\seealso{
\code{\link[digest]{digest}} to create a \code{hash}.
}
\keyword{internal}

