% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-package.R
\docType{package}
\name{ggplot2-package}
\alias{ggplot2}
\alias{ggplot2-package}
\title{ggplot2: Create Elegant Data Visualisations Using the Grammar of Graphics}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A system for 'declaratively' creating graphics, based on "The Grammar of Graphics". You provide the data, tell 'ggplot2' how to map variables to aesthetics, what graphical primitives to use, and it takes care of the details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ggplot2.tidyverse.org}
  \item \url{https://github.com/tidyverse/ggplot2}
  \item Report bugs at \url{https://github.com/tidyverse/ggplot2/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomas.pedersen@posit.co} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

Authors:
\itemize{
  \item Thomas Lin Pedersen \email{thomas.pedersen@posit.co} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})
  \item Hadley Wickham \email{hadley@posit.co} (\href{https://orcid.org/0000-0003-4757-117X}{ORCID})
  \item Winston Chang (\href{https://orcid.org/0000-0002-1576-2126}{ORCID})
  \item Lionel Henry
  \item Kohske Takahashi
  \item Claus Wilke (\href{https://orcid.org/0000-0002-7470-9261}{ORCID})
  \item Kara Woo (\href{https://orcid.org/0000-0002-5125-4188}{ORCID})
  \item Hiroaki Yutani (\href{https://orcid.org/0000-0002-3385-7233}{ORCID})
  \item Dewey Dunnington (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})
  \item Teun van den Brand (\href{https://orcid.org/0000-0002-9335-7468}{ORCID})
}

Other contributors:
\itemize{
  \item Posit, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
