% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aaa-.r
\name{ggplot2-r6}
\alias{Geom}
\alias{Position}
\alias{Stat}
\alias{ggplot2-r6}
\title{Base R6 classes for ggplot2}
\description{
If you are creating a new geom, stat or positions in another package, you'll
need to extend from \code{ggplot2::Geom}, \code{ggplot2::Stat} or
\code{ggplot2::Position}.
}
\keyword{internal}

