% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.r
\docType{package}
\name{ggplot2}
\alias{ggplot2}
\alias{ggplot2-package}
\title{ggplot2: An Implementation of the Grammar of Graphics}
\description{
An implementation of the grammar of graphics in R.
}
\details{
ggplot2 combines the advantages of both base and lattice graphics:
conditioning and shared axes are handled automatically, and you can still
build up a plot step by step from multiple data sources.
It also implements a sophisticated multidimensional conditioning system
and a consistent interface to map data to aesthetic attributes.

See http://ggplot2.org for more information, documentation and examples.

To learn more about ggplot2, start with the vignettes:
\code{browseVignettes(package = "ggplot2")}
}

