% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-construction.R
\name{ggplot_add}
\alias{ggplot_add}
\title{Add custom objects to ggplot}
\usage{
ggplot_add(object, plot, ...)
}
\arguments{
\item{object}{An object to add to the plot}

\item{plot}{The ggplot object to add \code{object} to}

\item{...}{Additional arguments to pass to the methods. Typically, an
\code{object_name} argument that gives a display name for \code{object} to use
in error messages.}
}
\value{
A modified ggplot object
}
\description{
This generic allows you to add your own methods for adding custom objects to
a ggplot with \link{+.gg}.
}
\keyword{internal}
