\name{ggplot_gtable}
\alias{ggplot_gtable}
\title{Build a plot with all the usual bits and pieces.}
\usage{ggplot_gtable(plot, data=ggplot_build(plot),
    drop=plot$options$drop, keep=plot$options$keep)}

\description{
  Build a plot with all the usual bits and pieces.
}

\details{
  This function builds all grobs necessary for displaying
  the plot, and stores them in a special data structure
  called a \code{\link{gtable}}. This object is amenable to
  programmatic manipulation, should you want to (e.g.) make
  the legend box 2 cm wide, or combine multiple plots into
  a single display, preserving aspect ratios across the
  plots.
}
\seealso{\code{\link{print.ggplot}} and \code{link{benchplot}} for 
for functions that contain the complete set of steps for generating
a ggplot2 plot.}
\value{a \code{\link{gtable}} object}
\keyword{internal}
\arguments{
  \item{plot}{plot object}
  \item{data}{plot data generated by \code{\link{ggplot_build}}}
  \item{drop}{optional, elements to drop out of the final display}
  \item{keep}{optional, elements to keep in the final display}
}
