% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-render.r
\name{ggplotGrob}
\alias{ggplotGrob}
\title{Generate a ggplot2 plot grob.}
\usage{
ggplotGrob(x)
}
\arguments{
\item{x}{ggplot2 object}
}
\description{
Generate a ggplot2 plot grob.
}
\keyword{internal}

