\name{ggstrip}
\alias{ggstrip}
\title{Grob strip}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Grob for strip labels
}
\usage{ggstrip(text, horizontal=TRUE, strip.gp=ggopt()$strip.gp, text.gp=ggopt()$strip.text.gp)}
\arguments{
\item{text}{text to display}
\item{horizontal}{orientation, horizontal or vertical}
\item{strip.gp}{}
\item{text.gp}{}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
