\name{ggtheme}
\alias{ggtheme}
\title{Set ggplot theme.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
A theme is a list of options for \code{\link{ggopt}}.
}
\usage{ggtheme(theme)}
\arguments{
\item{theme}{theme, a list of options for \code{\link{ggopt}}}
}

\details{Use \code{ggtheme(defaulttheme)} to reset back to the
default theme.}

\examples{}
\keyword{manip}
