% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-.R
\name{graphical-units}
\alias{.pt}
\alias{.stroke}
\title{Graphical units}
\usage{
.pt

.stroke
}
\description{
Multiply size in mm by these constants in order to convert to the units
that grid uses internally for \code{lwd} and \code{fontsize}.
}
\keyword{internal}
