\name{grobGrid}
\alias{grobGrid}
\alias{cbind.grobGrid}
\alias{colWidths}
\alias{cweave.grobGrid}
\alias{cweave.list}
\alias{cweave.matrix}
\alias{dim.grobGrid}
\alias{gTree.gridGrob}
\alias{grid.draw.gridGrob}
\alias{gridLayout}
\alias{grobCol}
\alias{grobGrid}
\alias{grobRow}
\alias{grobs.gridGrob}
\alias{print.grobGrid}
\alias{rbind.grobGrid}
\alias{rep.unit2}
\alias{rowHeights}
\alias{rweave.grobGrid}
\alias{rweave.list}
\alias{rweave.matrix}
\alias{spacer}
\alias{viewports.gridGrob}
\alias{as.list.unit}
\alias{interleave.default}
\alias{interleave.list}
\alias{interleave.unit}
\alias{max2}
\title{Experimental tools for create grids of grobs}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Still a work in progress.
}
\usage{grobGrid(name, nrow, ncol, grobs = NULL, widths = 1, heights = 1, clip = "on", default.units = "null", as.table = FALSE, respect = FALSE)}
\arguments{
\item{name}{}
\item{nrow}{}
\item{ncol}{}
\item{grobs}{}
\item{widths}{}
\item{heights}{}
\item{clip}{}
\item{default.units}{}
\item{as.table}{}
\item{respect}{}
}

\details{}

\examples{}
\keyword{internal}
