\name{guide_axis}
\alias{guide_axis}
\title{Grob axis}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Grob for axes
}
\usage{guide_axis(at, labels, position="right", theme)}
\arguments{
\item{at}{position of ticks}
\item{labels}{labels at ticks}
\item{position}{position of axis (top, bottom, left or right)}
\item{theme}{range of data values}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
