% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guides-.R
\name{guide-exts}
\alias{guide-exts}
\alias{guide_train}
\alias{guide_merge}
\alias{guide_geom}
\alias{guide_transform}
\alias{guide_gengrob}
\title{S3 generics for guides.}
\usage{
guide_train(guide, scale, aesthetic = NULL)

guide_merge(guide, new_guide)

guide_geom(guide, layers, default_mapping)

guide_transform(guide, coord, panel_params)

guide_gengrob(guide, theme)
}
\arguments{
\item{guide}{The guide object}
}
\description{
You will need to provide methods for these S3 generics if you want to
create your own guide object. They are currently undocumented; use at
your own risk!
}
\keyword{internal}
