\name{guide_grid}
\alias{guide_grid}
\title{Draw a grid}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Produce a grob to be used as a background in panels
}
\usage{guide_grid(theme, x.minor, x.major, y.minor, y.major)}
\arguments{
\item{theme}{theme to use to draw elements}
\item{x.minor}{minor breaks in x axis}
\item{x.major}{major breaks in x axis}
\item{y.minor}{minor breaks in y axis}
\item{y.major}{major breaks in y axis}
}




\keyword{internal}
