\name{guide_legend}
\alias{guide_legend}
\title{Build a legend grob}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Build the grob for a single legend.
}
\usage{guide_legend(legend, usage=usage, theme)}
\arguments{
\item{legend}{}
\item{usage}{}
\item{theme}{}
}
\value{A grid grob}
\details{}

\examples{}
\keyword{internal}
