% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guides-.R
\name{guide_data}
\alias{guide_data}
\title{Extract tick information from guides}
\usage{
guide_data(plot = last_plot(), aesthetic, i = 1L, j = 1L)
}
\arguments{
\item{plot}{A \code{ggplot} or \code{ggplot_build} object.}

\item{aesthetic}{A string that describes a single aesthetic for which to
extract guide information. For example: \code{"colour"}, \code{"size"}, \code{"x"} or
\code{"y.sec"}.}

\item{i, j}{An integer giving a row (i) and column (j) number of a facet for
which to return position guide information.}
}
\value{
A \code{data.frame} containing information extracted from the guide key,
a \code{list} when the coord doesn't support position axes, or \code{NULL} when no
such information could be found.
}
\description{
\code{guide_data()} builds a plot and extracts information from guide keys. This
information typically contains positions, values and/or labels, depending
on which aesthetic is queried or guide is used.
}
\examples{
# A standard plot
p <- ggplot(mtcars) +
  aes(mpg, disp, colour = drat, size = drat) +
  geom_point() +
  facet_wrap(vars(cyl), scales = "free_x")

# Guide information for legends
guide_data(p, "size")

# Note that legend guides can be merged
merged <- p + guides(colour = "legend")
guide_data(merged, "size")

# Guide information for positions
guide_data(p, "x", i = 1, j = 2)

# Coord polar doesn't support proper guides, so we get a list
polar <- p + coord_polar()
guide_data(polar, "theta", i = 1, j = 2)
}
\keyword{internal}
