% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guides-none.r
\name{guide_none}
\alias{guide_none}
\title{Empty guide}
\usage{
guide_none(title = waiver(), position = waiver())
}
\arguments{
\item{position}{Where this guide should be drawn: one of top, bottom,
left, or right.}
}
\description{
This guide draws nothing.
}
