\name{guide_range}
\alias{guide_range}
\title{Axis guide}
\usage{
  guide_range(title = NULL, label = NULL,
    label_pos = waiver(), length = 1, unit = NULL,
    position = NULL, location = NULL, space = "panel")
}
\arguments{
  \item{title}{NYI}

  \item{label}{}

  \item{label_pos}{one of "trbl" or waiver. position of
  label}

  \item{length}{}

  \item{unit}{character string}

  \item{position}{tlbr or float}

  \item{location}{numeric vector of x and y}

  \item{space}{"panel" or "data"}
}
\description{
  Axis guide
}

