% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guides-.r
\name{guides_gengrob}
\alias{guides_gengrob}
\title{guides_grob function}
\usage{
guides_gengrob(gdefs, theme)
}
\arguments{
\item{gdefs}{...}

\item{theme}{....}
}
\description{
generate grob from each gdef (needs to write this function?)
}
