\name{hexBin}
\alias{hexBin}
\title{Bin 2d plane into hexagons}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Wrapper around \code{\link[hexbin]{hcell2xy}} that returns a data frame
}
\usage{hexBin(x, y, binwidth)}
\arguments{
\item{x}{x positions}
\item{y}{y positions}
\item{binwidth}{numeric vector of length 2 giving binwidth in x and y directions}
}





