\name{hexGrob}
\alias{hexGrob}
\title{Draw hexagon grob}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Modified from code by Nicholas Lewin-Koh and Martin Maechler
}
\usage{hexGrob(x, y, size = rep(1, length(x)), colour = "grey50", fill = "grey90")}
\arguments{
\item{x}{x positions of hex centres}
\item{y}{y positions}
\item{size}{vector of hex sizes}
\item{colour}{border colour}
\item{fill}{fill colour}
}




\keyword{internal}
