\name{html_index_create}
\alias{html_index_create}
\title{Create physical file for html documentation index}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
See \code{\link{html_index}} for more details
}
\usage{html_index_create(path="web/")}
\arguments{
\item{path}{path to create file in}
}




\keyword{internal}
