\name{html_index_create}
\alias{html_index_create}
\title{Create physical file for html documentation index}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
See \code{\link{html_index}} for more details
}
\usage{html_index_create(path="web/")}
\arguments{
\item{path}{path to create file in}
\item{}{path to create files in}
}

\details{@arguments path to create file in
@keyword internal
Create all html documentation pages
Create all html pages including indices and templates.  Also converts
pdfs to pngs and optimises.}

\examples{}
\keyword{internal}
\keyword{internal}
