\name{html_template_create}
\alias{html_template_create}
\title{Create html file for templated files}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
See \code{\link{html_template}} for more details
}
\usage{html_template_create(name, path="web/")}
\arguments{
\item{name}{name of template}
\item{path}{path to create file in}
\item{}{path to create file in}
\item{}{list of objects}
}

\details{@arguments name of template
@arguments path to create file in
@keyword internal
Create all templates
Render all templates in templates directory

@arguments path to create file in
@keyword internal
Convenience function for generating lists of objects with their icons.
Build a bulleted list of objects with description and icons}

\examples{}
\keyword{internal}
\keyword{internal}
\keyword{internal}
