\name{html_template}
\alias{html_template}
\title{Generate html for templated files}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
See templates directory for examples.  Each template is auto linked.
}
\usage{html_template(name)}
\arguments{
\item{name}{name of template}
}




\keyword{internal}
