% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-plyr.R
\name{id}
\alias{id}
\title{Create an unique integer id for each unique row in a data.frame}
\usage{
id(.variables, drop = FALSE)
}
\arguments{
\item{.variables}{list of variables}

\item{drop}{Should unused factor levels be dropped?}
}
\value{
An integer vector with attribute \code{n} giving the total number of
possible unique rows
}
\description{
Properties:
\itemize{
\item \code{order(id)} is equivalent to \code{do.call(order, df)}
\item rows containing the same data have the same value
\item if \code{drop = FALSE} then room for all possibilites
}
}
\keyword{internal}
