% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ignoring_data}
\alias{ignoring_data}
\alias{.ignore_data}
\alias{.expose_data}
\title{Ignoring and exposing data}
\usage{
.ignore_data(data)

.expose_data(data)
}
\arguments{
\item{data}{A list of \verb{<data.frame>}s.}
}
\value{
A modified list of \verb{<data.frame>s}
}
\description{
The \code{.ignore_data()} function is used to hide \verb{<AsIs>} columns during
scale interactions in \code{ggplot_build()}. The \code{.expose_data()} function is
used to restore hidden columns.
}
\examples{
data <- list(
  data.frame(x = 1:3, y = I(1:3)),
  data.frame(w = I(1:3), z = 1:3)
)

ignored <- .ignore_data(data)
str(ignored)

.expose_data(ignored)
}
\keyword{internal}
