% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{interleave}
\alias{interleave}
\title{Interleave (or zip) multiple vectors into a single vector.}
\usage{
interleave(...)
}
\arguments{
\item{...}{vectors to interleave}
}
\description{
Interleave (or zip) multiple vectors into a single vector.
}
\keyword{internal}

