\name{is.constant}
\alias{is.constant}
\title{is.constant}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Determine if an expression represents a constant value
}
\usage{is.constant(x)}
\arguments{
\item{x}{}
}

\details{Used by qplot to determine whether a value should be mapped or set}

\examples{}

