% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{is.ggplot}
\alias{is.ggplot}
\title{Reports whether x is a ggplot object}
\usage{
is.ggplot(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Reports whether x is a ggplot object
}
\keyword{internal}
