% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order-.r
\name{is.order}
\alias{is.order}
\title{Is this object an ordering specification?}
\usage{
is.order(x)
}
\arguments{
\item{x}{object to test}
}
\description{
Is this object an ordering specification?
}
\keyword{internal}

