% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-checks.R
\name{is_tests}
\alias{is_tests}
\alias{is.ggplot}
\alias{is.mapping}
\alias{is.coord}
\alias{is.Coord}
\alias{is.ggproto}
\alias{is.facet}
\alias{is.layer}
\alias{is.geom}
\alias{is.stat}
\alias{is.position}
\alias{is.scale}
\alias{is.guide}
\alias{is.guides}
\alias{is.theme}
\alias{is.element}
\alias{is.margin}
\title{Reports whether x is a type of object}
\usage{
is.ggplot(x)

is.mapping(x)

is.coord(x)

is.Coord(x) # Deprecated

is.ggproto(x)

is.facet(x)

# Layer related tests
is.layer(x)

is.geom(x)

is.stat(x)

is.position(x)

# Scale and guide related tests
is.scale(x)

is.guide(x)

is.guides(x)

# Theme related tests
is.theme(x)

is.element(x)

is.margin(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Reports whether x is a type of object
}
\keyword{internal}
