% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-contour.r
\name{isoband_z_matrix}
\alias{isoband_z_matrix}
\title{Compute input matrix for isoband functions}
\usage{
isoband_z_matrix(data)
}
\arguments{
\item{data}{A data frame with columns \code{x}, \code{y}, and \code{z}.}
}
\value{
A \code{\link[=matrix]{matrix()}}
}
\description{
Note that grDevices::contourLines() needs transposed
output to the matrix returned by this function.
}
