% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/facet-labels.r
\name{label_parsed}
\alias{label_parsed}
\title{Label facets with parsed label.}
\usage{
label_parsed(variable, value)
}
\arguments{
\item{variable}{variable name passed in by facetter}

\item{value}{variable value passed in by facetter}
}
\description{
Label facets with parsed label.
}
\examples{
mtcars$cyl2 <- factor(mtcars$cyl, labels = c("alpha", "beta", "gamma"))
qplot(wt, mpg, data = mtcars) + facet_grid(. ~ cyl2)
qplot(wt, mpg, data = mtcars) + facet_grid(. ~ cyl2,
  labeller = label_parsed)
}
\seealso{
\code{\link{plotmath}}

Other facet labellers: \code{\link{label_both}};
  \code{\link{label_bquote}}; \code{\link{label_value}}
}

