% Generated by roxygen2 (4.0.0): do not edit by hand
\name{label_wrap_gen}
\alias{label_wrap_gen}
\title{Label facets with a word wrapped label.}
\usage{
label_wrap_gen(width = 25)
}
\arguments{
  \item{width}{integer, target column width for output.}
}
\description{
Uses \code{\link[base]{strwrap}} for line wrapping.
}
\examples{
set.seed(331)
x=runif(60)
y=rnorm(60)
speed=sample(c('Prime group', 'Rib group', 'No group'), 60, replace=TRUE)
group=sample(letters[1:3], 60, replace=TRUE)

df = data.frame(x=x, y=y, speed=as.factor(speed), group=as.factor(group))
group.names <- c('a'='First','b'='Second','c'="Don\\'t")

ggplot(df, aes(x, y)) + geom_point() + facet_grid(speed ~ group, labeller=label_wrap_gen(3))
ggplot(df, aes(x, y)) + geom_point() + facet_grid(speed ~ group, labeller=labeller(speed=label_wrap_gen(3), group=group.names))
}
\seealso{
, \code{\link{labeller}}
}

