% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/facet-labels.r
\name{labellers}
\alias{label_both}
\alias{label_bquote}
\alias{label_context}
\alias{label_parsed}
\alias{label_value}
\alias{label_wrap_gen}
\alias{labellers}
\title{Labeller functions}
\usage{
label_value(labels, multi_line = TRUE)

label_both(labels, multi_line = TRUE, sep = ": ")

label_context(labels, multi_line = TRUE, sep = ": ")

label_parsed(labels, multi_line = TRUE)

label_bquote(expr)

label_wrap_gen(width = 25, multi_line = TRUE)
}
\arguments{
\item{labels}{Data frame of labels. Usually contains only one
element, but facetting over multiple factors entails multiple
label variables.}

\item{multi_line}{Whether to display the labels of multiple factors
on separate lines.}

\item{sep}{String separating variables and values.}

\item{expr}{Backquoted labelling expression.}

\item{width}{Maximum number of characters before wrapping the strip.}
}
\description{
Labeller functions are in charge of formatting the strip labels of
facet grids and wraps. Most of them accept a \code{multi_line}
argument to control whether multiple factors (defined in formulae
such as \code{~first + second}) should be displayed on a single
line separated with commas, or each on their own line.
}
\details{
\code{label_value()} only displays the value of a factor while
\code{label_both()} displays both the variable name and the factor
value. \code{label_context()} is context-dependent and uses
\code{label_value()} for single factor facetting and
\code{label_both()} when multiple factors are
involved. \code{label_wrap_gen()} uses \code{\link[base]{strwrap}()}
for line wrapping.

\code{\link{label_parsed}()} interprets the labels as plotmath
expressions. \code{\link{label_bquote}()} offers a more flexible
way of constructing plotmath expressions. See examples and
\code{\link{bquote}()} for details on the syntax of the
argument.
}
\section{Writing New Labeller Functions}{


  A labeller function accepts a data frame of labels (character
  vectors) containing one column for each factor. Multiple factors
  occur with formula of the type \code{~first + second}.

  The return value must be a rectangular list where each 'row'
  characterises a single facet. The list elements can be either
  character vectors or lists of plotmath expressions. When multiple
  elements are returned, they get displayed on their own new lines
  (i.e., each facet gets a multi-line strip of labels).

  To illustrate, let's say your labeller returns a list of two
  character vectors of length 3. This is a rectangular list because
  all elements have the same length. The first facet will get the
  first elements of each vector and display each of them on their
  own line. Then the second facet gets the second elements of each
  vector, and so on.

  For compatibility with \code{\link{labeller}()}, each labeller
  function must have the \code{labeller} S3 class.
}
\examples{
mtcars$cyl2 <- factor(mtcars$cyl, labels = c("alpha", "beta", "gamma"))
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()

# Displaying only the values
p + facet_grid(. ~ cyl)
p + facet_grid(. ~ cyl, labeller = label_value)

# Displaying both the values and the variables
p + facet_grid(. ~ cyl, labeller = label_both)

# Displaying only the values or both the values and variables
# depending on whether multiple factors are facetted over
p + facet_grid(am ~ vs+cyl, labeller = label_context)

# Interpreting the labels as plotmath expressions
p + facet_grid(. ~ cyl2)
p + facet_grid(. ~ cyl2, labeller = label_parsed)
p + facet_wrap(~vs + cyl2, labeller = label_parsed)

# You can also provide a flexible backquoted expression. The labels
# must be backquoted and referred to by their names.
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
p + facet_grid(. ~ vs, labeller = label_bquote(alpha ^ .(vs)))
p + facet_grid(. ~ vs, labeller = label_bquote(.(vs) ^ .(vs)))
p + facet_grid(. ~ vs + am, labeller = label_bquote(.(am) ^ .(vs)))
}
\seealso{
\code{\link{labeller}()}
}

