\name{last_plot}
\alias{last_plot}
\title{Retrieve last plot modified/created.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Whenever a plot is created or modified, it is recorded.
}
\usage{last_plot()}
\arguments{

}

\details{}
\seealso{\code{\link{ggsave}}}
\examples{}
\keyword{hplot}
