\name{layout_grid}
\alias{layout_grid}
\title{Layout panels in a 2d grid.}
\usage{layout_grid(data, rows, cols, margins)}

\description{
  Layout panels in a 2d grid.
}
\value{a data frame with columns \code{PANEL}, \code{ROW} and \code{COL},
that match the facetting variable values up with their position in the 
grid}
\keyword{internal}
