\name{locate_grid}
\alias{locate_grid}
\title{Take single layer of data and combine it with panel information to split...}
\usage{locate_grid(data, panels, rows, cols, margins=FALSE)}

\description{
  Take single layer of data and combine it with panel
  information to split data into different panels.  Adds in
  extra data for missing facetting levels and for margins.
}
