% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot2.r
\docType{data}
\name{luv_colours}
\alias{luv_colours}
\title{\code{colors()} in Luv space.}
\format{A data frame with 657 observations and 4 variables:
\itemize{
\item{L,u,v}{Position in Luv colour space}
\item{col}{Colour name}
}}
\usage{
luv_colours
}
\description{
All built-in \code{\link{colors}()} translated into Luv colour space.
}
\keyword{datasets}

