\name{make_labels}
\alias{make_labels}
\title{Convert aesthetic mapping into text labels}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
This is used by ggplot and + to ensure consistent label formatting
}
\usage{make_labels(mapping)}
\arguments{
\item{mapping}{}
}




\keyword{internal}
