% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes-calculated.r
\name{mapped}
\alias{mapped}
\title{Mapped aesthetics}
\usage{
mapped(x)
}
\arguments{
\item{x}{An aesthetic expression using only variables available after
mapping.}
}
\description{
Usually you will want the mapping of aesthetics to be based on input data
but sometimes, e.g. for colour and fill, you'd want one aesthetic to be a
function of another. Using the \code{mapped()} function you can mark an aesthetic
to only be evaluated after all other aesthetics have been calculated and
mapped. This allows you to e.g. set fill to be a transparent verison of ¨
colour.
}
\note{
Using \code{mapped()} for an aesthetic will not create a guide for that
aesthetic, as it essentially a function of one of the other guides.
}
\examples{
ggplot(mpg, aes(class, hwy)) +
  geom_boxplot(aes(colour = class, fill = mapped(alpha(colour, 0.4))))

}
\seealso{
\code{\link[=stat]{stat()}} for marking aesthetics for evaluation after statistical
transformation
}
