% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/margins.R
\name{margin}
\alias{margin}
\title{Define margins.}
\usage{
margin(t, r, b, l, unit = "pt")
}
\arguments{
\item{t,b,r,l}{Dimensions of each margin. Uses CSS recyling rules, so if you
want the same margin in all directions, use one argument, if you want
the vertical and horizontal margins, use two arguments.
(To remember order, think trouble).}

\item{unit}{Default units of dimensions. Defaults to "pt" so it
can be most easily scaled with the text.}
}
\description{
This is a convenience function that creates a grid unit object of the
correct length to use for setting margins.
}
\examples{
margin(4)
margin(4, 2)
margin(4, 3, 2, 1)
}

