\name{match.fun.null}
\alias{match.fun.null}
\title{Cleaner version of match.fun}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Version of \code{\link{match.fun}} that returns NULL on failure
}
\usage{match.fun.null(x)}
\arguments{
\item{x}{function name to find (character vector)}
}
\value{function if found, otherwise NULL}
\details{}

\examples{}
\keyword{internal}
