% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-.r
\name{max_height}
\alias{max_height}
\alias{max_width}
\title{Get the maximal width/length of a list of grobs}
\usage{
max_height(grobs)

max_width(grobs)
}
\arguments{
\item{grobs}{A list of grobs}
}
\value{
The largest value. measured in cm as a unit object
}
\description{
Get the maximal width/length of a list of grobs
}
\keyword{internal}
