\name{mean_se}
\alias{mean_se}
\title{Mean + se's.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Mean and standard errors on either side.
}
\usage{mean_se(x, mult = 1)}
\arguments{
\item{x}{numeric vector}
\item{mult}{number of multiples of standard error}
}


\seealso{for use with \code{\link{stat_summary}}}


