% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-summary.r
\name{mean_se}
\alias{mean_se}
\title{Calculate mean and standard error}
\usage{
mean_se(x, mult = 1)
}
\arguments{
\item{x}{numeric vector}

\item{mult}{number of multiples of standard error}
}
\value{
A data frame with columns \code{y}, \code{ymin}, and \code{ymax}.
}
\description{
For use with \code{\link[=stat_summary]{stat_summary()}}
}
\examples{
x <- rnorm(100)
mean_se(x)
}
