\name{merge_legends}
\alias{merge_legends}
\title{Merge legends}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Merge multiple legend descriptions into one
}
\usage{merge_legends(legends)}
\arguments{
\item{legends}{list of legends to merge}
}

\details{Does not check that it makes sense to merge them.}

\examples{}
\keyword{internal}
