% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{merge_element}
\alias{merge_element}
\title{Merge a parent element into a child element}
\usage{
merge_element(new, old, ..., inherit = FALSE)
}
\arguments{
\item{new}{The child element in the theme hierarchy}

\item{old}{The parent element in the theme hierarchy}

\item{...}{Optional arguments, not currently used.}

\item{inherit}{Boolean whether to replace (\code{FALSE}, default) or inherit
(\code{TRUE}) parent values. Relevant for \code{\link[=rel]{rel()}} and \code{\link[=margin]{margin()}} values.}
}
\value{
A modified version of \code{new} updated with the properties of
\code{old}
}
\description{
This is a generic and element classes must provide an implementation of this
method
}
\examples{
new <- element_text(colour = "red")
old <- element_text(colour = "blue", size = 10)

# Adopt size but ignore colour
merge_element(new, old)

}
\keyword{internal}
