\name{Diamond prices}
\docType{data}
\alias{midwest}
\title{Prices of 50,000 round cut diamonds}
\description{

The variables are as follows:

\begin{itemize}
 \item PID
 \item county
 \item state
 \item area
 \item poptotal.  Total population
 \item popdensity. Population density
 \item popwhite.  Number of whites.
 \item popblack.  Number of blacks.
 \item popamerindian.  Number of American Indians.
 \item popasian.  Number of Asians.
 \item popother.  Number of other races.
 \item percwhite.  Percent white.
 \item percblack.  Percent black.
 \item percamerindan.  Percent American Indian.
 \item percasian. Percent Asian.
 \item percother. Percent other races.
 \item popadults.  Number of adults.
 \item perchsd.  
 \item percollege.  Percent college educated.
 \item percprof.  Percent profession.
 \item poppovertyknown. 
 \item percpovertyknown
 \item percbelowpoverty
 \item percchildbelowpovert
 \item percadultpoverty
 \item percelderlypoverty
 \item inmetro.  In a metro area.
 \item category'
\end{itemize}
}

\usage{data(midwest)}
\format{A data frame with 437 rows and 28 variables}
\keyword{datasets}
