\name{missing_colour}
\alias{missing_colour}
\title{Add a missing colour to a colour palette.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Convenient method.
}
\usage{missing_colour(palette, missing, na.colour)}
\arguments{
\item{palette}{}
\item{missing}{}
\item{na.colour}{}
}




\keyword{internal}
