% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mpg}
\alias{mpg}
\title{Fuel economy data from 1999 and 2008 for 38 popular models of car}
\format{A data frame with 234 rows and 11 variables
\itemize{
  \item manufacturer.
  \item model.
  \item displ. engine displacement, in litres
  \item year.
  \item cyl. number of cylinders
  \item trans. type of transmission
  \item drv. f = front-wheel drive, r = rear wheel drive, 4 = 4wd
  \item cty. city miles per gallon
  \item hwy. highway miles per gallon
  \item fl.
  \item class.
}}
\usage{
mpg
}
\description{
This dataset contains a subset of the fuel economy data that the EPA makes
available on \url{http://fueleconomy.gov}. It contains only models which
had a new release every year between 1999 and 2008 - this was used as a
proxy for the popularity of the car.
}
\keyword{datasets}

