\name{munch_data}
\alias{munch_data}
\title{For munching, only grobs are lines and polygons: everything else is}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
transfomed into those special cases by the geom.
}
\usage{munch_data(data, dist = NULL, segment_length = 0.01)}
\arguments{
\item{data}{distance, scaled from 0 to 1 (maximum distance on plot)}
\item{dist}{}
\item{segment_length}{}
}




\keyword{internal}
