\name{muted}
\alias{muted}
\title{Mute standard R colours.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
This produces colours with moderate luminance and saturation.s
}
\usage{muted(colour, l=30, c=70)}
\arguments{
\item{colour}{}
\item{l}{}
\item{c}{}
}

\details{}

\examples{}
\keyword{internal}
