\name{muted}
\alias{muted}
\title{Mute standard R colours.}
\usage{muted(colour, l=30, c=70)}

\description{
  Mute standard R colours.
}

\details{
  This produces colours with moderate luminance and
  saturation.
}
\keyword{internal}
\arguments{
  \item{colour}{colour to modify}
  \item{l}{new luminance}
  \item{c}{new chroma}
}
