% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide-.R
\name{new_guide}
\alias{new_guide}
\title{Guide constructor}
\usage{
new_guide(..., available_aes = "any", super)
}
\arguments{
\item{...}{Named arguments that match the parameters of \code{super$params} or
the theme elements in \code{super$elements}.}

\item{available_aes}{A vector of character strings listing the aesthetics
for which the guide can be drawn.}

\item{super}{The super class to use for the constructed guide. Should be a
Guide class object.}
}
\value{
A \code{Guide} ggproto object.
}
\description{
A constructor function for guides, which performs some standard compatability
checks between the guide and provided arguments.
}
\keyword{internal}
