\name{nice_ramp}
\alias{nice_ramp}
\title{Nice colour ramp}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Wrapper for colorRamp that deals with missing values and alpha
}
\usage{nice_ramp(ramp, x, alpha = 1)}
\arguments{
\item{ramp}{}
\item{x}{}
\item{alpha}{}
}




\keyword{internal}
