\name{nullGrob}
\alias{nullGrob}
\alias{widthDetails.nullGrob}
\alias{heightDetails.nullGrob}
\alias{grobWidth.nullGrob}
\alias{grobHeight.nullGrob}
\alias{drawDetails.nullGrob}
\title{Null grob}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
The null grob does nothing and has zero size.
}
\usage{nullGrob()}
\arguments{

}

\details{}

\examples{}
\keyword{internal}
