% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide-old.R
\name{old_guide}
\alias{old_guide}
\alias{guide_train}
\alias{guide_merge}
\alias{guide_geom}
\alias{guide_transform}
\alias{guide_gengrob}
\title{The previous S3 guide system}
\usage{
guide_train(guide, scale, aesthetic = NULL)

guide_merge(guide, new_guide)

guide_geom(guide, layers, default_mapping = NULL)

guide_transform(guide, coord, panel_params)

guide_gengrob(guide, theme)

old_guide(guide)
}
\arguments{
\item{guide}{An old guide object}
}
\description{
The guide system has been overhauled to use the ggproto infrastructure to
accommodate guide extensions with the same flexibility as layers, scales and
other ggplot2 objects. In rewriting, the old S3 system has become defunct,
meaning that the previous methods for guides have been superseded by ggproto
methods. As a fallback option, the generics, but not the methods, that the
previous S3 system used are encapsulated in the \code{GuideOld} ggproto class.
}
\keyword{internal}
