% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order-as.r
\name{order_as}
\alias{order_as}
\title{Order as specified by formula}
\usage{
order_as(ordering, facets = "all", order_f = max, na.rm = TRUE, ...)
}
\arguments{
\item{ordering}{a formula with the axes (of the tabular display) to order on
the LHS and the aesthetics (of the tabular display) to use for determining
the order on the RHS. Example x ~ y to order the x axis by values on y}

\item{facets}{How to handle ordering across facets. "within" will produce
different axis ordering within each facet panel. "betwee" will order the
panels themselves. "all" will do both.}

\item{order_f}{Function used to create ordering.}

\item{na.rm}{Should missing values be removed? Passed to order_f}

\item{...}{Additional arguments to pass to order_f}
}
\description{
Order as specified by formula
}

