% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order-null.r
\name{order_null}
\alias{order_null}
\title{Order specification: no ordering.}
\usage{
order_null()
}
\description{
Order specification: no ordering.
}
\examples{
# order_null is the default ordering specification if you
# don't override it with order_as
}

