\name{panelGrob}
\alias{panelGrob}
\title{ggplot plot}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Creates a complete ggplot grob.
}
\usage{panelGrob(plot, pieces = ggplot_build(plot))}
\arguments{
\item{plot}{plot object}
\item{pieces}{should the plot be wrapped up inside the pretty accoutrements (labels, legends, etc)}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
