\name{panels_default}
\alias{panels_default}
\title{Default panels function.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Place all grobs in the [x,y] position in panel\_x\_y viewport
}
\usage{panels_default(plot, grobs)}
\arguments{
\item{plot}{plot object}
\item{grobs}{}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
