\name{parse_unit_spec}
\alias{parse_unit_spec}
\title{Parse date time unit specification}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Parse the time unit specification used by \code{\link{cut.Date}} into something useful
}
\usage{parse_unit_spec(unitspec)}
\arguments{
\item{unitspec}{}
}




\keyword{internal}
