% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-partial.R
\name{partial-scales}
\alias{partial-scales}
\alias{scale_partial}
\alias{scale_x}
\alias{scale_y}
\title{Partial scales}
\usage{
scale_partial(aesthetic, ..., call = caller_call())

scale_x(...)

scale_y(...)
}
\arguments{
\item{aesthetic}{A string specifying an aesthetic to create a partial
scale for.}

\item{...}{Arguments passed onto full scales.}

\item{call}{A \verb{<call>} to report in warning and error messages.}
}
\value{
A \verb{<ScalePartial>} object that can be added to a plot.
}
\description{
Partial scales are useful for setting some scale parameters without
committing to any particular scale yet. Partial scales can be added to a plot
and will be combined with full scales.
}
\examples{
# A standard plot
p <- ggplot(mpg, aes(displ, hwy)) +
  geom_point()

# Adding a partial scale
p + scale_y(trans = "sqrt")

# Partial scales can be stacked
p + scale_y(trans = "sqrt") + scale_y(breaks = seq(15, 45, by = 5))

# When two scales declare the same parameter, the latter overrules the first
p + scale_y_continuous(name = "Highway Miles") +
  scale_y(name = "Title from partial scale")

# But other parameters are kept and not overruled
p + scale_y(name = "Highway Miles",
            breaks = c(20, 30, 40),
            labels = c("A", "B", "C")) +
  scale_y_continuous(name = "Title from full scale")
}
