% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-patterns.R
\name{pattern_alpha}
\alias{pattern_alpha}
\title{Modify transparency for patterns}
\usage{
pattern_alpha(x, alpha)
}
\arguments{
\item{x}{Object to be interpreted as pattern.}

\item{alpha}{A \code{numeric} vector between 0 and 1. If \code{NA}, alpha values
are preserved.}
}
\value{
\code{x} with modified transparency
}
\description{
This generic allows you to add your own methods for adding transparency to
pattern-like objects.
}
\keyword{internal}
