\name{plot_layout}
\alias{plot_layout}
\title{Plot layout}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Create \\code{\\link{grid.layout}} for plot
}
\usage{plot_layout(gm, rows, cols, row.labels, col.labels, axes_h, axes_v, aspect_ratio)}
\arguments{
\item{gm}{grob matrix}
\item{rows}{total number of rows}
\item{cols}{total number of columns}
\item{row.labels}{data frame of row labels}
\item{col.labels}{data frame of column labels}
\item{axes_h}{matrix of horizontal axis grobs}
\item{axes_v}{matrix of vertical axis grobs}
\item{aspect_ratio}{aspect ratio of cells (defaults to not preserved)}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
