% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/templates.r
\name{plot-templates}
\alias{ggfluctuation}
\alias{ggmissing}
\alias{ggorder}
\alias{ggpcp}
\alias{ggstructure}
\alias{plot-templates}
\alias{plotmatrix}
\title{Plot templates.}
\usage{
ggpcp(data, vars = names(data), ...)

ggfluctuation(table, type = "size", floor = 0, ceiling = max(table$freq,
  na.rm = TRUE))

ggmissing(data, avoid = "stack", order = TRUE, missing.only = TRUE)

ggstructure(data)

ggorder(data)

plotmatrix(data, mapping = aes(), colour = "black")
}
\description{
These plot templates are deprecated in an attempt to make ggplot2 as
streamlined as possible, and to avoid bugs in these poorly tested
functions. See the \pkg{GGally} package for some alternatives.
}
\keyword{internal}

