% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.r
\name{plot_theme}
\alias{plot_theme}
\title{plot_theme function}
\usage{
plot_theme(x)
}
\arguments{
\item{x}{plot object}
}
\description{
Combine plot defaults with current theme to get complete theme for a plot
}
