% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/position-beeswarm.r
\name{position_beeswarm}
\alias{position_beeswarm}
\title{Beeswarm-style plots to show overlapping points. x must be discrete.}
\usage{
position_beeswarm(width = NULL, height = NULL)
}
\arguments{
\item{width}{degree of jitter in x direction. Defaults to 90\% of the
resolution of the data.}

\item{height}{ignored}
}
\description{
Beeswarm-style plots to show overlapping points. x must be discrete.
}
\examples{
qplot(class, hwy, data = mpg, position="beeswarm")
}
\seealso{
Other position adjustments: \code{\link{position_dodge}};
  \code{\link{position_fill}};
  \code{\link{position_identity}};
  \code{\link{position_jitterdodge}};
  \code{\link{position_jitter}};
  \code{\link{position_stack}}
}

