% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-boxdodge.r
\name{position_boxdodge}
\alias{position_boxdodge}
\title{Position dodge for box plots}
\usage{
position_boxdodge(width = NULL, preserve = c("total", "single"))
}
\arguments{
\item{width}{Dodging width, when different to the width of the individual
elements. This is useful when you want to align narrow geoms with wider
geoms. See the examples.}

\item{preserve}{Should dodging preserve the total width of all elements
at a position, or the width of a single element?}
}
\description{
Position dodge for box plots
}
