% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-connection.R
\name{position_connect}
\alias{position_connect}
\title{Connect observations}
\usage{
position_connect(connection = "hv")
}
\arguments{
\item{connection}{A specification of how to points are connected. Can be one of the following:
\itemize{
\item A string giving a named connection. These options are:
\itemize{
\item \code{"hv"} to first jump horizontally, then vertically.
\item \code{"vh"} to first jump vertically, then horizontally.
\item \code{"mid"} to step half-way between adjacent x-values.
\item \code{"linear"} to use a straight segment.
}
\item A numeric matrix with two columns giving x and y coordinates respectively.
The coordinates should describe points on a path that connect point A at
location (0, 0) to point B at location (1, 1). At least one of these two
points is expected to be included in the coordinates.
}}
}
\description{
A line connecting two points is usually drawn as a straight segment. This
position adjustment gives additional options for how two points are connected.
}
\examples{
# Mirroring `geom_step()`
ggplot(head(economics, 20), aes(date, unemploy)) +
  geom_line(position = "connect")

# Making a histogram without bars
ggplot(faithful, aes(waiting)) +
  geom_area(
    stat = "bin", bins = 20, pad = TRUE,
    position = position_connect("mid")
  )

# Using custom connections with a matrix.
# Note that point A at (0, 0) is not included, but point B at (1, 1) is.
zigzag <- cbind(c(0.4, 0.6, 1), c(0.75, 0.25, 1))
x <- seq(0, 1, length.out = 20)[-1]
smooth <- cbind(x, scales::rescale(1 / (1 + exp(-(x * 10 - 5)))))

ggplot(head(huron, 10), aes(year, level)) +
  geom_line(position = position_connect(zigzag), aes(colour = "zigzag")) +
  geom_line(position = position_connect(smooth), aes(colour = "smooth")) +
  geom_point()
}
\seealso{
Other position adjustments: 
\code{\link{position_dodge}()},
\code{\link{position_identity}()},
\code{\link{position_jitter}()},
\code{\link{position_jitterdodge}()},
\code{\link{position_nudge}()},
\code{\link{position_stack}()}
}
\concept{position adjustments}
