% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-dodge2.r
\name{position_dodge2}
\alias{position_dodge2}
\title{Alternate method for dodging overlapping objects}
\usage{
position_dodge2(width = NULL, preserve = c("single", "total"),
  padding = 0)
}
\arguments{
\item{width}{Dodging width, when different to the width of the individual
elements. This is useful when you want to align narrow geoms with wider
geoms. See the examples.}

\item{preserve}{Should dodging preserve the total width of all elements
at a position, or the width of a single element?}

\item{padding}{Padding between boxes at the same position. Boxes are shrunk
by this proportion to make room for space between them. Defaults to 0.}
}
\description{
Dodging preserves the vertical position of an geom while adjusting the
horizontal position. \code{position_dodge2} is a special case of \code{position_dodge}
for arranging box plots, which can have variable widths. It also works with
bars and rectangles.
}
\examples{
ggplot(data = iris, aes(Species, Sepal.Length)) +
  geom_boxplot(aes(colour = Sepal.Width < 3.2))

ggplot(data = iris, aes(Species, Sepal.Length)) +
  geom_boxplot(aes(colour = Sepal.Width < 3.2), varwidth = TRUE)

ggplot(mtcars, aes(factor(cyl), fill = factor(vs))) +
  geom_bar(position = position_dodge2(preserve = "single"))

ggplot(mtcars, aes(factor(cyl), fill = factor(vs))) +
  geom_bar(position = position_dodge2(preserve = "total"))
}
\seealso{
Other position adjustments: \code{\link{position_dodge}},
  \code{\link{position_identity}},
  \code{\link{position_jitterdodge}},
  \code{\link{position_jitter}},
  \code{\link{position_nudge}},
  \code{\link{position_stack}}
}
