% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-identity.R
\name{position_identity}
\alias{position_identity}
\title{Don't adjust position}
\usage{
position_identity()
}
\description{
Don't adjust position
}
\seealso{
Other position adjustments:
\code{\link[=position_dodge]{position_dodge()}},
\code{\link[=position_jitter]{position_jitter()}},
\code{\link[=position_jitterdodge]{position_jitterdodge()}},
\code{\link[=position_nudge]{position_nudge()}},
\code{\link[=position_stack]{position_stack()}}
}
\concept{position adjustments}
