% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/position-jitterin.r
\name{position_jitterin}
\alias{position_jitterin}
\title{Adjust position by jittering inside the data range.}
\usage{
position_jitterin(width = NULL, height = NULL, bound = "y")
}
\arguments{
\item{width}{degree of jitter in x direction. Defaults to 5\% of
the resolution of the data.}

\item{height}{degree of jitter in y direction. Defaults to 0.}

\item{bound}{scales whose ranges bound the jitter. Can be
\code{"x"}, \code{"y"} or \code{c("x", "y")}. Defaults to
\code{"y"}.}
}
\description{
This produces nicer plots in some situations, especially for
logistic regression plots. The default values for \code{width} and
\code{height} have been tailored for plots of binary data against a
predictor. The width defaults to zero, while the height default to
5\% of the data resolution (0.05 in case of a 0-1 binary variable).
}
\examples{
movies_sub <- movies[sample(nrow(movies), 1000), ]
ggplot(movies_sub, aes(year, Comedy)) +
  geom_point(position = position_jitterin(), alpha = 0.5) +
  stat_smooth(method = "glm", family = "binomial", level = 0)
}
\seealso{
Other position adjustments: \code{\link{position_dodge}};
  \code{\link{position_fill}};
  \code{\link{position_identity}};
  \code{\link{position_jitterdodge}};
  \code{\link{position_jitter}};
  \code{\link{position_nudge}};
  \code{\link{position_stack}}
}

