% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-nudgestack.R
\name{position_nudgestack}
\alias{position_nudgestack}
\title{Simultaneously nudge and stack}
\usage{
position_nudgestack(x = 0, y = 0, vjust = 1, reverse = FALSE)
}
\arguments{
\item{x, y}{Amount of vertical and horizontal distance to move.}

\item{vjust}{Vertical adjustment for geoms that have a position
(like points or lines), not a dimension (like bars or areas). Set to
\code{0} to align with the bottom, \code{0.5} for the middle,
and \code{1} (the default) for the top.}

\item{reverse}{If \code{TRUE}, will reverse the default stacking order.
This is useful if you're rotating both the plot and legend.}
}
\description{
This is primarily used for set stacked columns between the ticks on the
x-axis.
}
\examples{
data <- mtcars
ggplot() +
  geom_col(
    data,
    aes(x = cyl, y = gear, fill = gear),
    position = position_nudgestack(x = 1)
  )
}
\seealso{
Other position adjustments: \code{\link{position_dodge}},
  \code{\link{position_identity}},
  \code{\link{position_jitterdodge}},
  \code{\link{position_jitter}},
  \code{\link{position_nudge}},
  \code{\link{position_stack}}
}
\concept{position adjustments}
