\name{Presidential terms}
\docType{data}
\alias{presidential}
\title{Terms of 10 presidents from Eisenhower to Bush W.}
\description{

The names of each president, the start and end date of their term, and their party of 10 US presidents from Eisenhower to Bush W.

}
\usage{data(presidential)}
\format{A data frame with 10 rows and 4 variables}
\keyword{datasets}
