\name{print.ggplot}
\alias{print.ggplot}
\title{Print ggplot}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Print generic for ggplot.  Plot on current graphics device.
}
\usage{print.ggplot(x, newpage = is.null(vp), vp = NULL, save=ggopt()$save, ...)}
\arguments{
\item{x}{plot to display}
\item{newpage}{draw new (empty) page first?}
\item{vp}{viewport to draw plot in}
\item{save}{other arguments passed on to \\code{\\link{ggplot_plot}}}
\item{...}{}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
