\name{print.ggarrange}
\alias{print.ggarrange}
\title{print ggarrange object}
\usage{
  \method{print}{ggarrange} (x, newpage = is.null(vp),
    vp = NULL, ...)
}
\arguments{
  \item{x}{ggarrange object to display}

  \item{...}{other arguments passed on to
  \code{\link{layer}}. This can include aesthetics whose
  values you want to set, not map. See \code{\link{layer}}
  for more details.}
}
\description{
  print ggarrange object
}

