\name{print.ggtable}
\alias{print.ggtable}
\title{Print ggtable object}
\usage{
  print.ggtable(x, newpage = is.null(vp), vp = NULL, ...)
}
\arguments{
  \item{x}{plot to display}

  \item{newpage}{draw new (empty) page first?}

  \item{vp}{viewport to draw plot in}

  \item{...}{other arguments not used by this method}
}
\description{
  Print ggtable object
}
\seealso{
  \code{\link{ggtable}}
}

