% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/proto2.r
\name{print.proto2}
\alias{print.proto2}
\title{Print a proto2 object}
\usage{
\method{print}{proto2}(x, ..., flat = TRUE)
}
\arguments{
\item{flat}{If \code{TRUE} (the default), show a flattened list of all local
  and inherited members. If \code{FALSE}, show the inheritance hierarchy.}
}
\description{
If a proto2 object has a \code{$print} method, this will call that method.
Otherwise, it will print out the members of the object, and optionally, the
members of the inherited objects.
}

