\name{ps}
\alias{ps}
\title{Abbreviated paste}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Alias for paste with a shorter name and convenient defaults
}
\usage{ps(..., sep="", collapse="")}
\arguments{
\item{...}{character vectors to be concatenated}
\item{sep}{default separator}
\item{collapse}{default collapser}
}

\details{}

\examples{}
\keyword{internal}
