% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-elements.r
\docType{data}
\name{R6ElementTree}
\alias{R6ElementTree}
\title{R6 Element Tree Definition}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
R6ElementTree
}
\description{
This is an R6 Class definition, to hold the element tree structure.
Contains a reserved default set for use by ggplot2, and contains the methods
that permit the user to add additional definitions, provided they do not conflict
with the reserved structure. The Final structure and indexing functions, ie, those that
permit equal substitution with previous ggplot2 methods, is produced via combining
the default and user-defined element trees.
}
\author{
Nicholas Hamilton, UNSW Sydney
}
\keyword{datasets}

