% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/range.r
\docType{data}
\name{Range}
\alias{Range}
\alias{RangeContinuous}
\alias{RangeDiscrete}
\alias{continuous_range}
\alias{discrete_range}
\title{Mutable range objects}
\usage{
continuous_range()

discrete_range()
}
\description{
Mutable ranges have a two methods (\code{train} and \code{reset}), and make
it possible to build up complete ranges with multiple passes.
}
\details{
These range objects should be instantiated with
\code{\link{continuous_range}} and \code{\link{discrete_range}}.
}
\keyword{internal}

