% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-plyr.R
\name{rbind_dfs}
\alias{rbind_dfs}
\title{Bind data frames together by common column names}
\usage{
rbind_dfs(dfs)
}
\arguments{
\item{dfs}{A list of data frames}
}
\value{
A data.frame with the union of all columns from the data frames given
in \code{dfs}
}
\description{
This function is akin to \code{plyr::rbind.fill}, \code{dplyr::bind_rows}, and
\code{data.table::rbindlist}. It takes data frames in a list and stacks them on
top of each other, filling out values with \code{NA} if the column is missing from
a data.frame
}
\keyword{internal}
