\name{rect_to_poly}
\alias{rect_to_poly}
\title{Convert rectangle to polygon}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Useful for non-Cartesian coordinate systems where it's easy to work purely in terms of locations, rather than locations and dimensions.
}
\usage{rect_to_poly(xmin, xmax, ymin, ymax)}
\arguments{
\item{xmin}{}
\item{xmax}{}
\item{ymin}{}
\item{ymax}{}
}





