% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R, R/utilities-grid.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{alpha}
\alias{unit}
\alias{arrow}
\title{Objects exported from other packages}
\examples{
ggplot(mpg, aes(displ, hwy)) +
  geom_point(alpha = 0.5, colour = "blue")

ggplot(mpg, aes(displ, hwy)) +
  geom_point(colour = alpha("blue", 0.5))
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{grid}{\code{\link[grid:arrow]{arrow()}}, \code{\link[grid:unit]{unit()}}}

  \item{scales}{\code{\link[scales:alpha]{alpha()}}}
}}

