% Generated by roxygen2 (4.0.0): do not edit by hand
\name{remove_geom}
\alias{remove_geom}
\title{Remove a Geom From a ggplot2 Object}
\usage{
remove_geom(plot_object, geom)
}
\arguments{
  \item{plot_object}{the ggplot2 object to work on}

  \item{geom}{name of geom to remove}
}
\description{
remove_geom allows you to remove layers of from a ggplot2 object
by specifying the name of a geom to remove. This can be useful
when dynamically building ggplot2 objects and want to remove
cluttered layers that have already been added to the object in
previous function calls.
}
\examples{
p <- ggplot(data.frame(a=1:10, b=1:10), aes(x=a, y=b)) + geom_point() + geom_line()
p_minus_point <- remove_geom(p, "point")
stopifnot(
  setdiff(
    sapply(p$layers, function(x) x$geom$objname),
    sapply(p_minus_point$layers, function(x) x$geom$objname)
  ) == "point"
)
}

