\name{rename_aes}
\alias{rename_aes}
\title{Rename aesthetics}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Rename aesthetics named in American spelling or with base R graphic parameter names to ggplot2 names
}
\usage{rename_aes(x)}
\arguments{
\item{x}{}
}

\details{}

\examples{}

