% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-plyr.R
\name{rename}
\alias{rename}
\title{Rename elements in a list, data.frame or vector}
\usage{
rename(x, replace)
}
\arguments{
\item{x}{A data.frame or a named vector or list}

\item{replace}{A named character vector. The names identifies the elements in
\code{x} that should be renamed and the values gives the new names.}
}
\value{
\code{x}, with new names according to \code{replace}
}
\description{
This is akin to \code{dplyr::rename} and \code{plyr::rename}. It renames elements given
as names in the \code{replace} vector to the values in the \code{replace} vector
without touching elements not referenced.
}
\keyword{internal}
