% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-.R
\name{render_axes}
\alias{render_axes}
\title{Render panel axes}
\usage{
render_axes(x = NULL, y = NULL, coord, theme, transpose = FALSE)
}
\arguments{
\item{x, y}{A list of ranges as available to the draw_panel method in
\code{Facet} subclasses.}

\item{coord}{A \code{Coord} object}

\item{theme}{A \code{theme} object}

\item{transpose}{Should the output be transposed?}
}
\value{
A list with the element "x" and "y" each containing axis
specifications for the ranges passed in. Each axis specification is a list
with a "top" and "bottom" element for x-axes and "left" and "right" element
for y-axis, holding the respective axis grobs. Depending on the content of x
and y some of the grobs might be zeroGrobs. If \code{transpose=TRUE} the
content of the x and y elements will be transposed so e.g. all left-axes are
collected in a left element as a list of grobs.
}
\description{
These helpers facilitates generating theme compliant axes when
building up the plot.
}
\keyword{internal}
