% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-.R
\name{render_strips}
\alias{render_strips}
\title{Render panel strips}
\usage{
render_strips(x = NULL, y = NULL, labeller = identity, theme)
}
\arguments{
\item{x, y}{A data.frame with a column for each variable and a row for each
combination to draw}

\item{labeller}{A labeller function}

\item{theme}{a \code{theme} object}
}
\value{
A list with an "x" and a "y" element, each containing a "top" and
"bottom" or "left" and "right" element respectively. These contains a list of
rendered strips as gtables.
}
\description{
All positions are rendered and it is up to the facet to decide which to use
}
\keyword{internal}
