% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reuse-scale.R
\name{reuse_all_colour_scales}
\alias{reuse_all_colour_scales}
\title{Find every \verb{scale_colour_*()} in the environment and remap it for a different colour-based aesthetic.
The new scale functions will be named \verb{scale_[new_aesthetics]_*()}.}
\usage{
reuse_all_colour_scales(new_aesthetics = c("fill"))
}
\arguments{
\item{new_aesthetics}{The name of one or more new colour-based aesthetics to create scales for}
}
\description{
Find every \verb{scale_colour_*()} in the environment and remap it for a different colour-based aesthetic.
The new scale functions will be named \verb{scale_[new_aesthetics]_*()}.
}
