% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-plyr.R
\name{revalue}
\alias{revalue}
\title{Replace specified values with new values, in a factor or character vector}
\usage{
revalue(x, replace)
}
\arguments{
\item{x}{A character or factor vector}

\item{replace}{A named character vector with the names corresponding to the
elements to replace and the values giving the replacement.}
}
\value{
A vector of the same class as \code{x} with the given values replaced
}
\description{
An easy to use substitution of elements in a string-like vector (character or
factor). If \code{x} is a character vector the matching elements will be replaced
directly and if \code{x} is a factor the matching levels will be replaced
}
\keyword{internal}
