% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rweave}
\alias{rweave}
\title{Row weave.}
\usage{
rweave(...)
}
\arguments{
  \item{...}{matrices to weave together}
}
\description{
Weave together two (or more) matrices by row.
}
\details{
Matrices must have same dimensions.
}
\keyword{internal}

