% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{S7_ggplot}
\alias{S7_ggplot}
\title{ggplot class}
\usage{
S7_ggplot(
  data = class_missing,
  layers = class_missing,
  scales = class_missing,
  guides = class_missing,
  mapping = class_missing,
  theme = class_missing,
  coordinates = class_missing,
  facet = class_missing,
  labels = class_missing,
  plot_env = class_missing
)
}
\arguments{
\item{data}{Any object that can be used with \code{\link[=fortify]{fortify()}} to yield a
\verb{<data.frame>}.}

\item{layers}{A \verb{<list>} containing \verb{<LayerInstance>} objects. Typically
an empty \verb{<list>} that will be filled when adding layers using \code{+}.}

\item{scales}{A \verb{<ScalesList>} ggproto object that manages scales that
are added to the plot.}

\item{guides}{A \verb{<Guides>} ggproto object that manages guides that are
added to the plot.}

\item{mapping}{An \verb{<uneval>} object constructed with \code{\link[=aes]{aes()}} containing
the default aesthetic mappings.}

\item{theme}{A \verb{<theme>} object constructed with \code{\link[=theme]{theme()}} containing
non-data visual settings.}

\item{coordinates}{A \verb{<Coord>} ggproto object that manages the interpretation
of position aesthetics.}

\item{facet}{A \verb{<Facet>} ggproto object that manages the display of data
subsets.}

\item{labels}{A named \verb{<list>} of \code{character}s and \code{expression}s giving
aesthetic-label pairs.}

\item{plot_env}{An \verb{<environment>} in which the plot was created.}
}
\description{
The ggplot class is implemented using S7 and has properties needed to
build and render a plot.
}
\details{
The purpose of the ggplot class object is to allow developers to extend
their own versions of a ggplot class. Users should instead use the
\code{\link[=ggplot]{ggplot()}} interface to construct a new plot.
}
