\name{same_file}
\alias{same_file}
\title{Test if all files are the same}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Uses md5 checksum to rapidly check if multiple files are equal.
}
\usage{same_file(...)}
\arguments{
\item{...}{character vector of paths}
}
\value{boolean}
\details{}

\examples{}
\keyword{internal}
