\name{scale_usage}
\alias{scale_usage}
\title{Compute usage of scales}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Builds a list of aesthetics and the geoms that they are used by.
}
\usage{scale_usage(plot)}
\arguments{
\item{plot}{ggplot object}
}

\details{Used for drawing legends.}

\examples{}
\keyword{internal}
