% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_alucard}
\alias{scale_fill_alucard}
\alias{scale_color_alucard}
\alias{scale_colour_alucard}
\title{Alucard Theme Scales for \code{ggplot2}}
\usage{
scale_fill_alucard(..., discrete = FALSE, aesthetics = "fill")

scale_color_alucard(..., discrete = FALSE, aesthetics = "color")

scale_colour_alucard(..., discrete = FALSE, aesthetics = "color")
}
\arguments{
\item{...}{Parameters passed on to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}} if
\code{discrete == TRUE}, or to \code{\link[ggplot2:scale_gradient]{ggplot2::scale_fill_gradientn()}} if \code{discrete == FALSE}.}

\item{discrete}{Whether the scale is discrete. Defaults to \code{FALSE}.}

\item{aesthetics}{The aesthetics for the plot.}
}
\description{
Alucard Theme Scales for \code{ggplot2}
}
