\name{ScaleAsn}
\alias{scale_asn}
\alias{ScaleAsn}
\alias{scale_x_asn}
\alias{scale_y_asn}
\alias{scale_z_asn}
\alias{scale_xend_asn}
\alias{scale_yend_asn}
\title{scale_asn}
\description{Continuous position scale}
\details{
This page describes scale_asn, see \code{\link{layer}} and \code{\link{qplot}} for how to create a complete plot from individual components.
}
\usage{scale_asn(...)}
\arguments{
 \item{...}{ignored }
}
\seealso{\itemize{
  \item \code{\link{scale_discrete}}: Discrete position scales
  \item \url{http://had.co.nz/ggplot/scale_asn.html}
}}
\value{A \code{\link{layer}}}
\author{Hadley Wickham, \url{http://had.co.nz/}}
\keyword{hplot}
