\name{ScaleAtanh}
\alias{scale_atanh}
\alias{ScaleAtanh}
\alias{scale_x_atanh}
\alias{scale_y_atanh}
\alias{scale_z_atanh}
\alias{scale_xend_atanh}
\alias{scale_yend_atanh}
\title{scale_atanh}
\description{Continuous position scale}
\details{
This page describes scale_atanh, see \code{\link{layer}} and \code{\link{qplot}} for how to create a complete plot from individual components.
}
\usage{scale_atanh(...)}
\arguments{
 \item{...}{ignored }
}
\seealso{\itemize{
  \item \code{\link{scale_discrete}}: Discrete position scales
  \item \url{http://had.co.nz/ggplot/scale_atanh.html}
}}
\value{A \code{\link{layer}}}
\author{Hadley Wickham, \url{http://had.co.nz/}}
\keyword{hplot}
