\name{ScaleColour}
\alias{scale_colour}
\alias{ScaleColour}
\alias{scale_colour_colour}
\alias{scale_fill_colour}
\title{scale_colour}
\description{Discrete position scale}
\details{
This page describes scale_colour, see \code{\link{layer}} and \code{\link{qplot}} for how to create a complete plot from individual components.
}
\usage{scale_colour(...)}
\arguments{
 \item{...}{ignored }
}
\seealso{\itemize{
  \item \url{http://had.co.nz/ggplot/scale_colour.html}
}}
\value{A \code{\link{layer}}}
\author{Hadley Wickham, \url{http://had.co.nz/}}
\keyword{hplot}
