% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scale-.r
\name{scale_dimension}
\alias{scale_dimension}
\title{The physical size of the scale.}
\usage{
scale_dimension(scale, expand = c(0, 0))
}
\description{
This always returns a numeric vector of length 2, giving the physical
dimensions of a scale.
}
\examples{
scale_dimension(xlim(0, 10))
scale_dimension(xlim("a", "b", "c"))
}
\keyword{internal}

