\name{ScaleExp}
\alias{scale_exp}
\alias{ScaleExp}
\alias{scale_x_exp}
\alias{scale_y_exp}
\alias{scale_z_exp}
\alias{scale_xend_exp}
\alias{scale_yend_exp}
\title{scale_exp}
\description{Continuous position scale}
\details{
This page describes scale_exp, see \code{\link{layer}} and \code{\link{qplot}} for how to create a complete plot from individual components.
}
\usage{scale_exp(...)}
\arguments{
 \item{...}{ignored }
}
\seealso{\itemize{
  \item \code{\link{scale_discrete}}: Discrete position scales
  \item \url{http://had.co.nz/ggplot/scale_exp.html}
}}
\value{A \code{\link{layer}}}
\author{Hadley Wickham, \url{http://had.co.nz/}}
\keyword{hplot}
