% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-geometry.R
\name{scale_geometry_discrete}
\alias{scale_geometry_discrete}
\title{Geometry scale}
\usage{
scale_geometry_discrete(data = NULL, id = NULL, ...)
}
\arguments{
\item{data}{An sf object with a simple feature list (geometry) column.}

\item{id}{A string giving a column name in \code{data} to use as feature identifiers.}

\item{...}{Passed on to \link{discrete_scale}.}
}
\description{
The geometry scale is a translation scale, transforming feature identifiers
into geometry using sf objects.
}
\examples{
\dontshow{if (require("rnaturalearth")) withAutoprint(\{ # examplesIf}
world <- rnaturalearth::ne_countries()

# 1:1 mapping using three-letter country codes
iso_a3 <- data.frame(id = c("AUS", "NZL", "PNG", "IDN"))

ggplot(iso_a3) +
  geom_sf(aes(geometry = id, fill = id)) +
  scale_geometry_discrete(world, "iso_a3")

# 1:many mapping using regions
regions <- data.frame(id = c("Asia", "Europe"))

ggplot(regions) +
  geom_sf(aes(geometry = id, fill = id)) +
  scale_geometry_discrete(world, "region_un")

# Subnational identities
states <- rnaturalearth::ne_states(country = "united states of america")

west_us <- data.frame(id = c("WA", "OR", "CA"))

ggplot(west_us) +
  geom_sf(aes(geometry = id, fill = id)) +
  scale_geometry_discrete(states, "postal")
\dontshow{\}) # examplesIf}
}
