% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-identity.R, R/zxx.R
\name{scale_identity}
\alias{scale_colour_identity}
\alias{scale_fill_identity}
\alias{scale_shape_identity}
\alias{scale_linetype_identity}
\alias{scale_linewidth_identity}
\alias{scale_alpha_identity}
\alias{scale_size_identity}
\alias{scale_discrete_identity}
\alias{scale_continuous_identity}
\alias{scale_color_identity}
\title{Use values without scaling}
\usage{
scale_colour_identity(
  name = waiver(),
  ...,
  guide = "none",
  aesthetics = "colour"
)

scale_fill_identity(name = waiver(), ..., guide = "none", aesthetics = "fill")

scale_shape_identity(
  name = waiver(),
  ...,
  guide = "none",
  aesthetics = "shape"
)

scale_linetype_identity(
  name = waiver(),
  ...,
  guide = "none",
  aesthetics = "linetype"
)

scale_linewidth_identity(
  name = waiver(),
  ...,
  guide = "none",
  aesthetics = "linewidth"
)

scale_alpha_identity(
  name = waiver(),
  ...,
  guide = "none",
  aesthetics = "alpha"
)

scale_size_identity(name = waiver(), ..., guide = "none", aesthetics = "size")

scale_discrete_identity(aesthetics, name = waiver(), ..., guide = "none")

scale_continuous_identity(aesthetics, name = waiver(), ..., guide = "none")
}
\arguments{
\item{name}{The name of the scale. Used as the axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}

\item{...}{Other arguments passed on to \code{\link[=discrete_scale]{discrete_scale()}} or
\code{\link[=continuous_scale]{continuous_scale()}}}

\item{guide}{Guide to use for this scale. Defaults to \code{"none"}.}

\item{aesthetics}{Character string or vector of character strings listing the
name(s) of the aesthetic(s) that this scale works with. This can be useful, for
example, to apply colour settings to the \code{colour} and \code{fill} aesthetics at the
same time, via \code{aesthetics = c("colour", "fill")}.}
}
\description{
Use this set of scales when your data has already been scaled, i.e. it
already represents aesthetic values that ggplot2 can handle directly.
These scales will not produce a legend unless you also supply the \code{breaks},
\code{labels}, and type of \code{guide} you want.
}
\details{
The functions \code{scale_colour_identity()}, \code{scale_fill_identity()}, \code{scale_size_identity()},
etc. work on the aesthetics specified in the scale name: \code{colour}, \code{fill}, \code{size},
etc. However, the functions \code{scale_colour_identity()} and \code{scale_fill_identity()} also
have an optional \code{aesthetics} argument that can be used to define both \code{colour} and
\code{fill} aesthetic mappings via a single function call. The functions
\code{scale_discrete_identity()} and \code{scale_continuous_identity()} are generic scales that
can work with any aesthetic or set of aesthetics provided via the \code{aesthetics}
argument.
}
\examples{
ggplot(luv_colours, aes(u, v)) +
  geom_point(aes(colour = col), size = 3) +
  scale_color_identity() +
  coord_fixed()

df <- data.frame(
  x = 1:4,
  y = 1:4,
  colour = c("red", "green", "blue", "yellow")
)
ggplot(df, aes(x, y)) + geom_tile(aes(fill = colour))
ggplot(df, aes(x, y)) +
  geom_tile(aes(fill = colour)) +
  scale_fill_identity()

# To get a legend guide, specify guide = "legend"
ggplot(df, aes(x, y)) +
  geom_tile(aes(fill = colour)) +
  scale_fill_identity(guide = "legend")
# But you'll typically also need to supply breaks and labels:
ggplot(df, aes(x, y)) +
  geom_tile(aes(fill = colour)) +
  scale_fill_identity("trt", labels = letters[1:4], breaks = df$colour,
  guide = "legend")

# cyl scaled to appropriate size
ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(size = cyl))

# cyl used as point size
ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(size = cyl)) +
  scale_size_identity()
}
\seealso{
The \href{https://ggplot2-book.org/scales-other#sec-scale-identity}{identity scales section} of the online ggplot2 book.

Other shape scales: \code{\link[=scale_shape]{scale_shape()}}, \code{\link[=scale_shape_manual]{scale_shape_manual()}}.

Other linetype scales: \code{\link[=scale_linetype]{scale_linetype()}}, \code{\link[=scale_linetype_manual]{scale_linetype_manual()}}.

Other alpha scales: \code{\link[=scale_alpha]{scale_alpha()}}, \code{\link[=scale_alpha_manual]{scale_alpha_manual()}}.

Other size scales: \code{\link[=scale_size]{scale_size()}}, \code{\link[=scale_size_manual]{scale_size_manual()}}.

Other colour scales:
\code{\link[=scale_alpha]{scale_alpha()}},
\code{\link[=scale_colour_brewer]{scale_colour_brewer()}},
\code{\link[=scale_colour_continuous]{scale_colour_continuous()}},
\code{\link[=scale_colour_discrete]{scale_colour_discrete()}},
\code{\link[=scale_colour_gradient]{scale_colour_gradient()}},
\code{\link[=scale_colour_grey]{scale_colour_grey()}},
\code{\link[=scale_colour_hue]{scale_colour_hue()}},
\code{\link[=scale_colour_manual]{scale_colour_manual()}},
\code{\link[=scale_colour_steps]{scale_colour_steps()}},
\code{\link[=scale_colour_viridis_d]{scale_colour_viridis_d()}}
}
\concept{colour scales}
