\name{ScaleInverse}
\alias{scale_inverse}
\alias{ScaleInverse}
\alias{scale_x_inverse}
\alias{scale_y_inverse}
\alias{scale_z_inverse}
\alias{scale_xend_inverse}
\alias{scale_yend_inverse}
\title{scale_inverse}
\description{Continuous position scale}
\details{
This page describes scale_inverse, see \code{\link{layer}} and \code{\link{qplot}} for how to create a complete plot from individual components.
}
\usage{scale_inverse(...)}
\arguments{
 \item{...}{ignored }
}
\seealso{\itemize{
  \item \code{\link{scale_discrete}}: Discrete position scales
  \item \url{http://had.co.nz/ggplot/scale_inverse.html}
}}
\value{A \code{\link{layer}}}
\author{Hadley Wickham, \url{http://had.co.nz/}}
\keyword{hplot}
