% Generated by roxygen2 (4.0.0): do not edit by hand
\name{scale_length_continuous}
\alias{scale_length}
\alias{scale_length_continuous}
\alias{scale_length_discrete}
\title{length scale.}
\usage{
scale_length_continuous(..., range = c(0.005, 0.1))

scale_length(..., range = c(0.005, 0.1))

scale_length_discrete(..., range = c(0.005, 0.1))
}
\arguments{
  \item{range}{a numeric vector of length 2 that specifies
  the minimum and maximum length of the plotting symbol
  after transformation.}

  \item{...}{common continuous scale parameters:
  \code{name}, \code{breaks}, \code{labels},
  \code{na.value}, \code{limits} and \code{trans}.  See
  \code{\link{continuous_scale}} for more details}
}
\description{
length scale.
}
\examples{
p <- ggplot(mtcars, aes(x=wt, y=mpg))
ggplot(mtcars) + geom_point(aes(x=wt, y=mpg))
ggplot(mtcars) + geom_point(aes(x=wt, y=mpg)) + geom_rug(aes(x=wt))
ggplot(mtcars) + geom_point(aes(x=wt, y=mpg)) + geom_rug(aes(x=wt))
mtQ <- data.frame(wt=quantile(x=mtcars$wt))
ggplot(mtcars) + geom_point(aes(x=wt, y=mpg)) + geom_rug(aes(x=wt)) + geom_rug(aes(x=wt), colour="red", data=mtQ)
ggplot(mtcars) + geom_point(aes(x=wt, y=mpg)) + geom_rug(aes(x=wt)) + geom_rug(aes(x=wt), colour="red", length=0.05, data=mtQ)
ggplot(mtcars) + geom_point(aes(x=wt, y=mpg)) + geom_rug(aes(x=wt, length=mpg))
}

