\name{ScaleLog}
\alias{scale_log}
\alias{ScaleLog}
\alias{scale_x_log}
\alias{scale_y_log}
\alias{scale_z_log}
\alias{scale_xend_log}
\alias{scale_yend_log}
\title{scale_log}
\description{Continuous position scale}
\details{
This page describes scale_log, see \code{\link{layer}} and \code{\link{qplot}} for how to create a complete plot from individual components.
}
\usage{scale_log(...)}
\arguments{
 \item{...}{ignored }
}
\seealso{\itemize{
  \item \code{\link{scale_discrete}}: Discrete position scales
  \item \url{http://had.co.nz/ggplot/scale_log.html}
}}
\value{A \code{\link{layer}}}
\author{Hadley Wickham, \url{http://had.co.nz/}}
\keyword{hplot}
