\name{ScaleLog10}
\alias{scale_log10}
\alias{ScaleLog10}
\alias{scale_x_log10}
\alias{scale_y_log10}
\alias{scale_z_log10}
\alias{scale_xend_log10}
\alias{scale_yend_log10}
\title{scale_log10}
\description{Continuous position scale}
\details{
This page describes scale_log10, see \code{\link{layer}} and \code{\link{qplot}} for how to create a complete plot from individual components.
}
\usage{scale_log10(...)}
\arguments{
 \item{...}{ignored }
}
\seealso{\itemize{
  \item \code{\link{scale_discrete}}: Discrete position scales
  \item \url{http://had.co.nz/ggplot/scale_log10.html}
}}
\value{A \code{\link{layer}}}
\author{Hadley Wickham, \url{http://had.co.nz/}}
\keyword{hplot}
