\name{ScaleLog2}
\alias{scale_log2}
\alias{ScaleLog2}
\alias{scale_x_log2}
\alias{scale_y_log2}
\alias{scale_z_log2}
\alias{scale_xend_log2}
\alias{scale_yend_log2}
\title{scale_log2}
\description{Continuous position scale}
\details{
This page describes scale_log2, see \code{\link{layer}} and \code{\link{qplot}} for how to create a complete plot from individual components.
}
\usage{scale_log2(...)}
\arguments{
 \item{...}{ignored }
}
\seealso{\itemize{
  \item \code{\link{scale_discrete}}: Discrete position scales
  \item \url{http://had.co.nz/ggplot/scale_log2.html}
}}
\value{A \code{\link{layer}}}
\author{Hadley Wickham, \url{http://had.co.nz/}}
\keyword{hplot}
