\name{ScaleLogit}
\alias{scale_logit}
\alias{ScaleLogit}
\alias{scale_x_logit}
\alias{scale_y_logit}
\alias{scale_z_logit}
\alias{scale_xend_logit}
\alias{scale_yend_logit}
\title{scale_logit}
\description{Continuous position scale}
\details{
This page describes scale_logit, see \code{\link{layer}} and \code{\link{qplot}} for how to create a complete plot from individual components.
}
\usage{scale_logit(...)}
\arguments{
 \item{...}{ignored }
}
\seealso{\itemize{
  \item \code{\link{scale_discrete}}: Discrete position scales
  \item \url{http://had.co.nz/ggplot/scale_logit.html}
}}
\value{A \code{\link{layer}}}
\author{Hadley Wickham, \url{http://had.co.nz/}}
\keyword{hplot}
