% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-.R
\name{scale_params}
\alias{scale_params}
\title{Setting scale parameters}
\usage{
scale_params(aesthetics, ...)
}
\arguments{
\item{aesthetics}{The name of the aesthetics for which to update the scale.}

\item{...}{Named arguments to one of the scale constructors,
\code{\link[=continuous_scale]{continuous_scale()}}, \code{\link[=discrete_scale]{discrete_scale()}} or \code{\link[=binned_scale]{binned_scale()}}.}
}
\value{
A \code{scale_params} object that can be added to a plot.
}
\description{
Setting scale parameters
}
\examples{
ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  scale_params("x", limits = c(0, 10)) +
  scale_params("y", transform = "sqrt")
}
