% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-partial.R
\name{scale_partial}
\alias{scale_partial}
\title{Partial scales constructor}
\usage{
scale_partial(
  aesthetic,
  name = waiver(),
  ...,
  .args = NULL,
  call = caller_call() \%||\% current_call()
)
}
\arguments{
\item{aesthetic}{A string for an aesthetic, like \code{"x"} or \code{"shape"}.}

\item{name}{A name for a scale.}

\item{..., .args}{Named arguments to pass on to the final scale. Use either
\code{...} as dynamic dots, or \code{.args} as a named list.}

\item{call}{A \code{call} object for reporting the scale in messages.}
}
\value{
A \verb{<ScalePartial>} ggproto object that can be added to a plot.
}
\description{
Constructor for partial scales.
}
\examples{
NULL
}
\keyword{internal}
