\name{ScalePow10}
\alias{scale_pow10}
\alias{ScalePow10}
\alias{scale_x_pow10}
\alias{scale_y_pow10}
\alias{scale_z_pow10}
\alias{scale_xend_pow10}
\alias{scale_yend_pow10}
\title{scale_pow10}
\description{Continuous position scale}
\details{
This page describes scale_pow10, see \code{\link{layer}} and \code{\link{qplot}} for how to create a complete plot from individual components.
}
\usage{scale_pow10(...)}
\arguments{
 \item{...}{ignored }
}
\seealso{\itemize{
  \item \code{\link{scale_discrete}}: Discrete position scales
  \item \url{http://had.co.nz/ggplot/scale_pow10.html}
}}
\value{A \code{\link{layer}}}
\author{Hadley Wickham, \url{http://had.co.nz/}}
\keyword{hplot}
