\name{ScaleProbit}
\alias{scale_probit}
\alias{ScaleProbit}
\alias{scale_x_probit}
\alias{scale_y_probit}
\alias{scale_z_probit}
\alias{scale_xend_probit}
\alias{scale_yend_probit}
\title{scale_probit}
\description{Continuous position scale}
\details{
This page describes scale_probit, see \code{\link{layer}} and \code{\link{qplot}} for how to create a complete plot from individual components.
}
\usage{scale_probit(...)}
\arguments{
 \item{...}{ignored }
}
\seealso{\itemize{
  \item \code{\link{scale_discrete}}: Discrete position scales
  \item \url{http://had.co.nz/ggplot/scale_probit.html}
}}
\value{A \code{\link{layer}}}
\author{Hadley Wickham, \url{http://had.co.nz/}}
\keyword{hplot}
