\name{ScaleReverse}
\alias{scale_reverse}
\alias{ScaleReverse}
\alias{scale_x_reverse}
\alias{scale_y_reverse}
\alias{scale_z_reverse}
\alias{scale_xend_reverse}
\alias{scale_yend_reverse}
\title{scale_reverse}
\description{Continuous position scale}
\details{
This page describes scale_reverse, see \code{\link{layer}} and \code{\link{qplot}} for how to create a complete plot from individual components.
}
\usage{scale_reverse(...)}
\arguments{
 \item{...}{ignored }
}
\seealso{\itemize{
  \item \code{\link{scale_discrete}}: Discrete position scales
  \item \url{http://had.co.nz/ggplot/scale_reverse.html}
}}
\value{A \code{\link{layer}}}
\author{Hadley Wickham, \url{http://had.co.nz/}}
\keyword{hplot}
