% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scale-size.r
\name{scale_size_area}
\alias{scale_size_area}
\title{Scale area, with fixed 0.}
\usage{
scale_size_area(..., max_size = 6)
}
\arguments{
\item{...}{Other arguments passed on to \code{\link{continuous_scale}}
to control name, limits, breaks, labels and so forth.}

\item{max_size}{Size of largest points.}
}
\description{
This scale is subtly different to \code{scale_area}: it ensures that a
value of zero is always mapped to size 0.
}
\details{
Note that this controls the size scale, so it will also control
the thickness of lines. Line thickness will be proportional to the square
root of the value, which is probably undesirable in most cases.
}

