\name{ScaleSize_discrete}
\alias{scale_size_discrete}
\alias{ScaleSize_discrete}
\title{scale_size_discrete}
\description{Size scale for discrete variables}
\details{
This page describes scale_size_discrete, see \code{\link{layer}} and \code{\link{qplot}} for how to create a complete plot from individual components.
}
\usage{scale_size_discrete(...)}
\arguments{
 \item{...}{ignored }
}
\seealso{\itemize{
  \item \url{http://had.co.nz/ggplot/scale_size_discrete.html}
}}
\value{A \code{\link{layer}}}
\author{Hadley Wickham, \url{http://had.co.nz/}}
\keyword{hplot}
