\name{ScaleSqrt}
\alias{scale_sqrt}
\alias{ScaleSqrt}
\alias{scale_x_sqrt}
\alias{scale_y_sqrt}
\alias{scale_z_sqrt}
\alias{scale_xend_sqrt}
\alias{scale_yend_sqrt}
\title{scale_sqrt}
\description{Continuous position scale}
\details{
This page describes scale_sqrt, see \code{\link{layer}} and \code{\link{qplot}} for how to create a complete plot from individual components.
}
\usage{scale_sqrt(...)}
\arguments{
 \item{...}{ignored }
}
\seealso{\itemize{
  \item \code{\link{scale_discrete}}: Discrete position scales
  \item \url{http://had.co.nz/ggplot/scale_sqrt.html}
}}
\value{A \code{\link{layer}}}
\author{Hadley Wickham, \url{http://had.co.nz/}}
\keyword{hplot}
