\name{scales_list}
\alias{scales_list}
\title{List of scales}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Make a list of scales, cloning if necessary
}
\usage{scales_list(scale, n, free)}
\arguments{
\item{scale}{input scale}
\item{n}{number of scales to produce in output}
\item{free}{should the scales be free (TRUE) or fixed (FALSE)}
}

\details{}

\examples{}

