% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggplot2.r
\docType{data}
\name{seals}
\alias{seals}
\title{Vector field of seal movements.}
\format{A data frame with 1155 rows and 4 variables}
\usage{
data(seals)
}
\description{
This vector field was produced from the data described in Brillinger, D.R.,
Preisler, H.K., Ager, A.A. and Kie, J.G. "An exploratory data analysis
(EDA) of the paths of moving animals". J. Statistical Planning and
Inference 122 (2004), 43-63, using the methods of Brillinger, D.R.,
"Learning a potential function from a trajectory", Signal Processing
 Letters. December (2007).
}
\references{
\url{http://www.stat.berkeley.edu/~brill/Papers/jspifinal.pdf}
}
\keyword{datasets}

