% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-edition.R
\name{set_edition}
\alias{set_edition}
\title{Set ggplot2 edition}
\usage{
set_edition(edition = NULL)
}
\arguments{
\item{edition}{An edition. Possible values currently include \code{"2026"} only.
Can be \code{NULL} (default) to unset an edition.}
}
\value{
The previous \code{edition} value. This function is called for the side
effect of setting the edition though.
}
\description{
ggplot2 uses the 'edition' concept to manage the lifecycles of functions and
arguments. Setting a recent edition opens up the latest features but also
closes down deprecated and superseded functionality.
}
\examples{
set_edition(2026)
}
