% Generated by roxygen2 (4.0.0): do not edit by hand
\name{set_geom_params}
\alias{set_geom_params}
\title{Set a Geom's Parameters of an Existing ggplot2 Object}
\usage{
set_geom_params(plot_object, geom, param_name, param_value)
}
\arguments{
  \item{plot_object}{the ggplot2 object to work on}

  \item{geom}{name of the geom to set the parameter for
  (e.g. "ribbon", "point")}

  \item{param_name}{name of parameter to set}

  \item{param_value}{desired value of the parameter}
}
\description{
Changes or sets a geom's parameters after a ggplot2 object has been created. This
function applies to all layers with a given geom type.
}
\examples{
p <- ggplot(data.frame(a=1:10, b=1:10), aes(x=a, y=b)) + geom_point() + geom_line()
p_set <- set_geom_params(p, "point", 'size', 10)
stopifnot(p_set$layers[[1]]$geom_params$size == 10)
}

