% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-last.R
\name{set_last_plot}
\alias{set_last_plot}
\title{Set the last plot to be fetched by lastplot()}
\usage{
set_last_plot(value)
}
\description{
Set the last plot to be fetched by lastplot()
}
\seealso{
\code{\link[=last_plot]{last_plot()}}
}
\keyword{internal}
