\name{setup_viewports}
\alias{setup_viewports}
\title{Setup viewports}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Setup matrix of viewports for a layout with given parameters
}
\usage{setup_viewports(type, rows=nrow(data), cols=ncol(data), data, offset=c(0,0), range, angle=0)}
\arguments{
\item{type}{viewport type}
\item{rows}{number of rows}
\item{cols}{number of columns}
\item{data}{optional data to compute rows and columns from}
\item{offset}{offset from top and left}
\item{range}{list containing x and y ranges}
\item{angle}{}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
