% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.R
\name{standardise_aes_names}
\alias{standardise_aes_names}
\title{Standardise aesthetic names}
\usage{
standardise_aes_names(x)
}
\arguments{
\item{x}{Character vector of aesthetics names, such as \code{c("colour", "size", "shape")}.}
}
\value{
Character vector of standardised names.
}
\description{
This function standardises aesthetic names by converting \code{color} to \code{colour}
(also in substrings, e.g. \code{point_color} to \code{point_colour}) and translating old style
R names to ggplot names (eg. \code{pch} to \code{shape}, \code{cex} to \code{size}).
}
\keyword{internal}
