\name{stat_abline}
\alias{stat_abline}
\alias{StatAbline}
\title{stat\_abline}
\description{Add a line with slope and intercept}
\details{
This page describes stat\_abline, see \code{\link{layer}} and \code{\link{qplot}} for how to create a complete plot from individual components.
}
\usage{stat_abline(mapping=NULL, data=NULL, geom="abline", position="identity", ...)}
\arguments{
 \item{mapping}{mapping between variables and aesthetics generated by aes}
 \item{data}{dataset used in this layer, if not specified uses plot dataset}
 \item{geom}{geometric used by this layer}
 \item{position}{position adjustment used by this layer}
 \item{...}{ignored }
}
\seealso{\itemize{
  \item \url{http://had.co.nz/ggplot2/stat_abline.html}
}}
\value{A \code{\link{layer}}}
\examples{\dontrun{

}}
\author{Hadley Wickham, \url{http://had.co.nz/}}
\keyword{hplot}
