% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stat-bin.r
\name{stat_bin}
\alias{stat_bin}
\title{Bin data.}
\usage{
stat_bin(mapping = NULL, data = NULL, geom = "bar", position = "stack",
  width = 0.9, drop = FALSE, right = FALSE, binwidth = NULL,
  origin = NULL, breaks = NULL, show_guide = NA, inherit.aes = TRUE,
  ...)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link{aes}} or \code{\link{aes_string}}. Only needs to be set
at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override
the plot defaults.}

\item{geom}{The geometric object to use display the data}

\item{position}{The position adjustment to use for overlappling points
on this layer}

\item{width}{Width of bars when used with categorical data}

\item{drop}{If TRUE, remove all bins with zero counts}

\item{right}{If \code{TRUE}, right-closed, left-open, if \code{FALSE},
the default, right-open, left-closed.}

\item{binwidth}{Bin width to use. Defaults to 1/30 of the range of the
data}

\item{origin}{Origin of first bin}

\item{breaks}{Actual breaks to use.  Overrides bin width and origin}

\item{show_guide}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{...}{other arguments passed on to \code{\link{layer}}. This can
include aesthetics whose values you want to set, not map. See
\code{\link{layer}} for more details.}
}
\value{
New data frame with additional columns:
  \item{count}{number of points in bin}
  \item{density}{density of points in bin, scaled to integrate to 1}
  \item{ncount}{count, scaled to maximum of 1}
  \item{ndensity}{density, scaled to maximum of 1}
}
\description{
Missing values are currently silently dropped.
}
\section{Aesthetics}{

\Sexpr[results=rd,stage=build]{ggplot2:::rd_aesthetics("stat", "bin")}
}
\examples{
\donttest{
simple <- data.frame(x = rep(1:10, each = 2))
base <- ggplot(simple, aes(x))
# By default, right = FALSE intervals are of the form [a, b)
base + stat_bin(binwidth = 1, drop = FALSE, right = FALSE, col = "black")
# If right = TRUE, and intervals are of the form (a, b]
base + stat_bin(binwidth = 1, drop = FALSE, right = TRUE, col = "black")

d <- ggplot(diamonds, aes(carat))
d + stat_bin()
d + stat_bin(binwidth = 0.1)
d + stat_bin(breaks=seq(0, 2, by = 0.01))
# See geom_histogram for more histogram examples

# To create a unit area histogram, use aes(y = ..density..)
linehist <- d + geom_freqpoly(aes(y = ..density..), binwidth = 0.1)
linehist
linehist + stat_density(colour = "blue", fill = NA)
}
}

