\name{stat_bin2d}
\alias{stat_bin2d}
\alias{StatBin2d}
\title{stat\_bin2d}
\description{2d binning}
\details{
This page describes stat\_bin2d, see \code{\link{layer}} and \code{\link{qplot}} for how to create a complete plot from individual components.
}
\section{Aesthetics}{
The following aesthetics can be used with stat\_bin2d.  Aesthetics are mapped to variables in the data with the \code{\link{aes}} function: \code{stat\_bin2d(\code{\link{aes}}(x = var))}
\itemize{
  \item \code{x}: x position (\strong{required}) 
  \item \code{y}: y position (\strong{required}) 
  \item \code{fill}: internal colour 
}
}
\usage{stat_bin2d(mapping=NULL, data=NULL, geom="rect", position="identity", bins=30, drop=TRUE, ...)}
\arguments{
 \item{mapping}{mapping between variables and aesthetics generated by aes}
 \item{data}{dataset used in this layer, if not specified uses plot dataset}
 \item{geom}{geometric used by this layer}
 \item{position}{position adjustment used by this layer}
 \item{bins}{NULL}
 \item{drop}{NULL}
 \item{...}{ignored }
}
\seealso{\itemize{
  \item \url{http://had.co.nz/ggplot2/stat_bin2d.html}
}}
\value{A \code{\link{layer}}}
\examples{\dontrun{

}}
\author{Hadley Wickham, \url{http://had.co.nz/}}
\keyword{hplot}
