\name{stat_hline}
\alias{stat_hline}
\alias{StatHline}
\title{stat\_hline}
\description{Add a horizontal line}
\details{
This page describes stat\_hline, see \code{\link{layer}} and \code{\link{qplot}} for how to create a complete plot from individual components.
}
\usage{stat_hline(mapping=NULL, data=NULL, geom="hline", position="identity", ...)}
\arguments{
 \item{mapping}{mapping between variables and aesthetics generated by aes}
 \item{data}{dataset used in this layer, if not specified uses plot dataset}
 \item{geom}{geometric used by this layer}
 \item{position}{position adjustment used by this layer}
 \item{...}{ignored }
}
\seealso{\itemize{
  \item \url{http://had.co.nz/ggplot2/stat_hline.html}
}}
\value{A \code{\link{layer}}}
\examples{\dontrun{

}}
\author{Hadley Wickham, \url{http://had.co.nz/}}
\keyword{hplot}
