% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/stat-km.R
\name{stat_km}
\alias{stat_km}
\title{Kaplan Meier Estimate of Survival}
\usage{
stat_km(mapping = NULL, data = NULL, geom = "km", position = "identity",
  ...)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link{aes}} or \code{\link{aes_string}}. Only needs to be set
at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override
the plot defaults.}

\item{geom}{The geometric object to use display the data}

\item{position}{The position adjustment to use for overlappling points
on this layer}

\item{...}{Other arguments passed to \link{survival::survfit.formula}}

\item{se}{display confidence interval around KM curve? (TRUE by default, use
\code{conf.int} to control significance level which is 0.95 by default)}
}
\value{
a data.frame with additional columns: \item{x}{x in data}
  \item{y}{Kaplan-Meier Survival Estimate at x} \item{ymin}{Lower confidence
  limit of KM curve, if \code{se = TRUE}} \item{ymax}{Upper confidence limit
  of KM curve, if \code{se = FALSE}}
}
\description{
Kaplan Meier Estimate of Survival
}
\examples{
\donttest{
ggplot(survival::lung, aes(time, status, color = sex)) + stat_km()
qplot(survival::lung, time, status = status, geom = "km")
}
}

