\name{StatSort_angle}
\alias{stat_sort_angle}
\alias{StatSort_angle}
\title{stat_sort_angle}
\description{}
\details{
This page describes stat_sort_angle, see \code{\link{layer}} and \code{\link{qplot}} for how to create a complete plot from individual components.
}
\section{Aesthetics}{
The following aesthetics can be used with stat_sort_angle.  Aesthetics are mapped to variables in the data with the \code{\link{aes}} function: \code{stat_sort_angle(\code{\link{aes}}(x = var))}
\itemize{
  \item \code{x}: x position (\strong{required}) 
  \item \code{y}: y position (\strong{required}) 
}
}
\usage{stat_sort_angle(...)}
\arguments{
 \item{...}{ignored }
}
\seealso{\itemize{
  \item \url{http://had.co.nz/ggplot/stat_sort_angle.html}
}}
\value{A \code{\link{layer}}}
\author{Hadley Wickham, \url{http://had.co.nz/}}
\keyword{hplot}
