% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stat-stepribbon.r
\name{stat_stepribbon}
\alias{stat_stepribbon}
\title{Plot stepped ribbons}
\usage{
stat_stepribbon(mapping = NULL, data = NULL, geom = "ribbon",
  position = "identity")
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link{aes}} or \code{\link{aes_string}}. Only needs to be set
at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override
the plot defaults.}

\item{geom}{The geometric object to use display the data}

\item{position}{The position adjustment to use for overlappling points
on this layer}

\item{direction}{Character. Can be "hv" or "vh" for the initial direction of steps.}
}
\value{
a data.frame with additional rows necessary to plot a step function
}
\description{
Plot stepped ribbons
}
\section{Aesthetics}{

\Sexpr[results=rd,stage=build]{ggplot2:::rd_aesthetics("stat", "stepribbon")}
}
\examples{
\donttest{
 time <- 1:50
 prob <- sort(runif(50), decreasing=TRUE)
 df <- data.frame(time, prob, lci = prob - 0.5, uci = prob + 0.7)

 ggplot(df, aes(x = time, y = prob, ymin = lci, ymax = uci)) +
 geom_step() +
 geom_ribbon(stat="stepribbon", alpha=0.3)
}
}
\seealso{
\code{\link{geom_ribbon}}, \code{\link{geom_step}}
}

