% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-summarise.R
\name{stat_summarise}
\alias{stat_summarise}
\title{Arbitrary summaries}
\usage{
stat_summarise(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  ...,
  fun = express_group,
  fun.args = list(),
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use to display the data for this layer.
When using a \verb{stat_*()} function to construct a layer, the \code{geom} argument
can be used to override the default coupling between stats and geoms. The
\code{geom} argument accepts the following:
\itemize{
\item A \code{Geom} ggproto subclass, for example \code{GeomPoint}.
\item A string naming the geom. To give the geom as a string, strip the
function name of the \code{geom_} prefix. For example, to use \code{geom_point()},
give the geom as \code{"point"}.
\item For more information and other ways to specify the geom, see the
\link[=layer_geoms]{layer geom} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[=layer_positions]{layer position} documentation.
}}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}'s \code{params} argument. These
arguments broadly fall into one of 4 categories below. Notably, further
arguments to the \code{position} argument, or aesthetics that are required
can \emph{not} be passed through \code{...}. Unknown arguments that are not part
of the 4 categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"}
or \code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics}
section that lists the available options. The 'required' aesthetics
cannot be passed on to the \code{params}. Please note that while passing
unmapped aesthetics as vectors is technically possible, the order and
required length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on
parameters to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation
lists which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[=layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[=draw_key]{key glyphs}, to change the display of the layer in the legend.
}}

\item{fun}{A function that is given the complete data per group and should
return a data.frame with variables compatible with the layer's \code{geom}
argument. Defaults to \code{express_group()} which will evaluate
\link[rlang:args_data_masking]{data-masked} expressions passed to \code{fun.args}.}

\item{fun.args}{Optional additional arguments to pass to \code{fun}. When
\code{fun = express_group} (default), can be a list of expressions created by
\code{vars()}.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display. To include legend keys for all levels, even
when no data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}
}
\description{
Arbitrary summaries
}
\section{Aesthetics}{

\code{stat_summarise()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
\item \code{\link[=aes_group_order]{group}}
}
Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.
}

\examples{
# Using a helper summary function
my_summary_fun <- function(data) {
  data.frame(
    x    = min(data$x) - 0.5,
    xend = max(data$x) + 0.5,
    y    = mean(data$y)
  )
}

ggplot(mpg, aes(drv, displ)) +
  geom_jitter(aes(colour = drv)) +
  stat_summarise(
    geom = "segment",
    fun = my_summary_fun
  )

# Making the same summary using the `exprs` argument
ggplot(mpg, aes(drv, displ)) +
  geom_jitter(aes(colour = drv)) +
  stat_summarise(
    geom = "segment",
    fun.args = vars(
      # Note: we use `xend` first to not mutate `x` too soon
      xend = max(x) + 0.5,
      x    = min(x) - 0.5,
      y    = mean(y)
    )
  )

# You can use a temporary variable that needn't conform to data.frame columns
# as long as it is removed afterwards
ggplot(mpg, aes(displ, colour = drv)) +
  stat_summarise(
    geom = "line",
    fun.args = vars(
      dens = density(x), # non-vector temporary variable
      x    = dens$x,
      y    = dens$y,
      dens = NULL # deleting the temporary variable
    )
  )
}
