\name{stat_vline}
\alias{stat_vline}
\alias{StatVline}
\title{stat\_vline}
\description{Add a vertical line}
\details{
This page describes stat\_vline, see \code{\link{layer}} and \code{\link{qplot}} for how to create a complete plot from individual components.
}
\usage{stat_vline(mapping=NULL, data=NULL, geom="vline", position="identity", intercept, ...)}
\arguments{
 \item{mapping}{mapping between variables and aesthetics generated by aes}
 \item{data}{dataset used in this layer, if not specified uses plot dataset}
 \item{geom}{geometric used by this layer}
 \item{position}{position adjustment used by this layer}
 \item{intercept}{NULL}
 \item{...}{ignored }
}
\seealso{\itemize{
  \item \url{http://had.co.nz/ggplot2/stat_vline.html}
}}
\value{A \code{\link{layer}}}
\examples{\dontrun{

}}
\author{Hadley Wickham, \url{http://had.co.nz/}}
\keyword{hplot}
