% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subplot.R
\name{subplot_wrap}
\alias{subplot_wrap}
\alias{subplot_grid}
\title{Create a ribbon or grid of subplots}
\usage{
subplot_wrap(facets, axis = "x", scales = "auto", ...)

subplot_grid(rows = NULL, cols = NULL, axis = "x", scales = "auto",
  ...)
}
\arguments{
\item{facets}{A set of variables or expressions quoted by \code{\link[=vars]{vars()}}
and defining faceting groups on the rows or columns dimension.
The variables can be named (the names are passed to \code{labeller}).

For compatibility with the classic interface, can also be a
formula or character vector. Use either a one sided formula, `~a
\itemize{
\item b\code{, or a character vector,}c("a", "b")`.
}}

\item{axis}{Axis to which the variables of the subplots are mapped.}

\item{scales}{Either "auto", in which case only the scale of the axis
designated by \code{axis} is free, or "free" in which case both scales are
free.}

\item{...}{Passed to \code{\link[=facet_wrap]{facet_wrap()}} or \code{\link[=facet_grid]{facet_grid()}}.}

\item{rows}{A set of variables or expressions quoted by
\code{\link[=vars]{vars()}} and defining faceting groups on the rows or columns
dimension. The variables can be named (the names are passed to
\code{labeller}).

For compatibility with the classic interface, \code{rows} can also be
a formula with the rows (of the tabular display) on the LHS and
the columns (of the tabular display) on the RHS; the dot in the
formula is used to indicate there should be no faceting on this
dimension (either row or column).}

\item{cols}{A set of variables or expressions quoted by
\code{\link[=vars]{vars()}} and defining faceting groups on the rows or columns
dimension. The variables can be named (the names are passed to
\code{labeller}).

For compatibility with the classic interface, \code{rows} can also be
a formula with the rows (of the tabular display) on the LHS and
the columns (of the tabular display) on the RHS; the dot in the
formula is used to indicate there should be no faceting on this
dimension (either row or column).}
}
\description{
\code{subplot_wrap} and \code{subplot_grid} are variations of \code{\link[=facet_wrap]{facet_wrap()}} and
\code{\link[=facet_grid]{facet_grid()}} that are more appropriate when the facets represent
different variables rather than different subsets of the same variable.
}
\examples{
# Stack different variables in the same column
# NB: see tidyr::gather for a more practical alternative
cars <- stack(mtcars, select=c(disp, drat, qsec))
cars$mpg <- mtcars$mpg
cars$cyl <- mtcars$cyl
head(cars)

# inspect the distribution of all stacked variables at once
p <- ggplot(cars) + geom_histogram(aes(values), bins=10)
p + facet_wrap(~ind, scales="free_x")
p + subplot_wrap(~ind)

# examine the relationship of all variables with mpg
p <- ggplot(cars) + geom_point(aes(values, mpg))
p + facet_wrap(~ind, scales="free_x")
p + subplot_wrap(~ind)
# or in the other direction
p <- ggplot(cars) + geom_point(aes(mpg, values))
p + facet_wrap(~ind, scales="free_y")
p + subplot_wrap(~ind, axis="y")

# do the same but with an additional conditional variable, hence
# generating a grid of subplots
p <- ggplot(cars) + geom_point(aes(values, mpg))
p + facet_grid(cyl~ind, scales="free_x")
p + subplot_grid(cyl~ind)

# the facet labels play the role of the axis title and respect its theme
ggplot(cars) + geom_point(aes(values, mpg)) + subplot_wrap(~ind)
theme_update(axis.title=element_text(colour="red"))
ggplot(cars) + geom_point(aes(values, mpg)) + subplot_wrap(~ind)
# but only when the plot if created *after* global theme updates
p <- ggplot(cars) + geom_point(aes(values, mpg)) + subplot_wrap(~ind)
theme_update(axis.title=element_text(colour="blue"))
p
p + theme(axis.title=element_text(colour="green"))
# here, in both cases, the facet labels stay blue, unfortunately
}
