% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise-plot.R
\name{summarise_coords}
\alias{summarise_coords}
\title{Get a summary of the coords in a built plot object}
\usage{
summarise_coords(p)
}
\description{
Get a summary of the coords in a built plot object
}
