% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise-plot.R
\name{summarise_layers}
\alias{summarise_layers}
\title{Get a summary of the layers in a built plot object}
\usage{
summarise_layers(p)
}
\description{
Get a summary of the layers in a built plot object
}
