% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise-plot.R
\name{summarise_layout}
\alias{summarise_layout}
\title{Get a summary of the layout in a built plot object}
\usage{
summarise_layout(p)
}
\description{
Get a summary of the layout in a built plot object
}
