% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise-plot.R
\name{summarise_plot}
\alias{summarise_plot}
\alias{summarise_layout}
\alias{summarise_coord}
\alias{summarise_layers}
\title{Summarise built plot objects}
\usage{
summarise_layout(p)

summarise_coord(p)

summarise_layers(p)
}
\arguments{
\item{p}{A ggplot_built object.}
}
\description{
These functions provide summarised information about built ggplot objects.
}
\examples{
p <- ggplot(mpg, aes(displ, hwy)) + geom_point() + facet_wrap(~class)
b <- ggplot_build(p)

summarise_layout(b)
summarise_coord(b)
summarise_layers(b)

}
