\name{summary.ggplot}
\alias{summary.ggplot}
\title{Summarise ggplot object}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Displays a useful description of a ggplot object
}
\usage{summary.ggplot(object, ...)}
\arguments{
\item{object}{}
\item{...}{}
}

\details{}

\examples{summary(qplot(mpg, wt, data=mtcars))}
\keyword{internal}
