\name{surround_viewports}
\alias{surround_viewports}
\title{Generate viewports for plot surroundings}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
This some pretty ugly code
}
\usage{surround_viewports(position, widths, heights, legend_vp)}
\arguments{
\item{position}{}
\item{widths}{}
\item{heights}{}
\item{legend_vp}{}
}





