\name{theme_grey}
\alias{theme_grey}
\alias{theme_gray}
\title{Grey theme}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Produce a theme with grey background and white gridlines
}
\usage{theme_grey(base_size = 12, base_family = "")}
\arguments{
\item{base_size}{base font size}
\item{base_family}{}
}




\keyword{dplot}
