\name{theme_line}
\alias{theme_line}
\title{Theme element: line}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
This element draws a line between two (or more) points
}
\usage{theme_line(colour = "black", size = 0.5, linetype = 1)}
\arguments{
\item{colour}{line color}
\item{size}{line size}
\item{linetype}{line type}
}


\seealso{\code{\link{polylineGrob}} for underlying grid function, \code{link{theme_segment}}}


