\name{theme_render}
\alias{theme_render}
\title{Render a theme element}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
This function is used internally for all drawing of plot surrounds etc
}
\usage{theme_render(theme, element, ..., name = NULL)}
\arguments{
\item{theme}{}
\item{element}{}
\item{...}{}
\item{name}{}
}

\details{It also names the created grobs consistently}

\examples{}

