\name{theme_text}
\alias{theme_text}
\title{text}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
family
}
\usage{theme_text(family = "", face = "plain", colour = "black", size = 10, hjust = 0.5, vjust = 0.5, angle = 0, lineheight = 1.1)}
\arguments{
\item{family}{}
\item{face}{}
\item{colour}{}
\item{size}{}
\item{hjust}{}
\item{vjust}{}
\item{angle}{}
\item{lineheight}{}
}

\details{}

\examples{}

