\name{theme_update}
\alias{theme_update}
\alias{theme_set}
\alias{theme_get}
\title{Get, set and update themes.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
These three functions get, set and update themes.
}
\usage{theme_update(...)}
\arguments{
\item{...}{}
}

\details{Use \code{theme_update} to modify a small number of elements of the current
theme or use \code{theme_set} to completely override it.}

\examples{}

