% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_alucard.R
\name{theme_alucard}
\alias{theme_alucard}
\title{Theme Alucard}
\usage{
theme_alucard()
}
\description{
Provides a minimal \code{ggplot2} theme with a Alucard, light backdrop.
}
\examples{
# Set the current `ggplot2` theme with `ggplot2::theme_set`:
library(ggplot2)
theme_set(theme_alucard())

# Or set it for the plot in the pipeline:
library(dplyr)
library(dRacula)
library(ggplot2)

mpg \%>\%
  filter(manufacturer \%in\% c("honda", "ford", "dodge", "audi")) \%>\%
  group_by(manufacturer) \%>\%
  summarize(mean_hwy = mean(hwy)) \%>\%
  ggplot(aes(x = manufacturer, y = mean_hwy, fill = manufacturer)) +
  theme(legend.position = "none") +
  coord_flip() + geom_col() +
  scale_fill_alucard(discrete = TRUE) +
  theme_alucard()
}
