\name{theme_blank}
\alias{theme_blank}
\alias{theme_line}
\alias{theme_rect}
\alias{theme_segment}
\alias{theme_text}
\title{Deprecated theme_xx functions}
\usage{
  theme_blank(...)
}
\description{
  The \code{theme_xx} functions have been deprecated. They
  are replaced with the \code{element_xx} functions.
}

