% Generated by roxygen2 (4.0.0): do not edit by hand
\name{theme_grey}
\alias{theme_gray}
\alias{theme_grey}
\title{A theme with grey background and white gridlines.}
\usage{
theme_grey(base_size = 12, base_family = "")
}
\arguments{
  \item{base_size}{base font size}

  \item{base_family}{base font family}
}
\description{
A theme with grey background and white gridlines.
}
\seealso{
Other themes: \code{\link{theme_bw}};
  \code{\link{theme_classic}}; \code{\link{theme_light}};
  \code{\link{theme_linedraw}}; \code{\link{theme_minimal}}
}

