% Generated by roxygen2 (4.0.0): do not edit by hand
\name{theme_light}
\alias{theme_light}
\title{A theme with white background and light grey lines}
\usage{
theme_light(base_size = 12, base_family = "")
}
\arguments{
  \item{base_size}{base font size}

  \item{base_family}{base font family}
}
\description{
A theme with white background and light grey lines
}
\examples{
p <- ggplot(mtcars) + geom_point(aes(x = wt, y = mpg))
p
# 'classic' dark-on-light theme
p + theme_bw()
# this theme
p + theme_light()
# variation with dark box and axes legends
p + theme_linedraw()
}
\seealso{
Other themes: \code{\link{theme_bw}};
  \code{\link{theme_classic}}; \code{\link{theme_gray}},
  \code{\link{theme_grey}}; \code{\link{theme_linedraw}};
  \code{\link{theme_minimal}}
}

