% Generated by roxygen2 (4.0.0): do not edit by hand
\name{theme_minimal}
\alias{theme_minimal}
\title{A minimalistic theme with no background annotations.}
\usage{
theme_minimal(base_size = 12, base_family = "")
}
\arguments{
  \item{base_size}{base font size}

  \item{base_family}{base font family}
}
\description{
A minimalistic theme with no background annotations.
}
\seealso{
Other themes: \code{\link{theme_bw}};
  \code{\link{theme_classic}}; \code{\link{theme_gray}},
  \code{\link{theme_grey}}; \code{\link{theme_light}};
  \code{\link{theme_linedraw}}
}

