\name{theme_rect}
\alias{theme_rect}
\title{Theme element: rectangle.}
\usage{
  theme_rect(fill = NA, colour = "black", size = 0.5,
  linetype = 1)
}
\arguments{
  \item{fill}{fill colour}

  \item{colour}{border color}

  \item{size}{border size}

  \item{linetype}{border linetype}
}
\description{
  Theme element: rectangle.
}
\details{
  Most often used for backgrounds and borders.
}
\seealso{
  \code{\link{rectGrob}} for underlying grid function
}
\keyword{dplot}

