% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-defaults.r
\name{theme_rtl}
\alias{theme_rtl}
\title{Right to Left Theme}
\usage{
theme_rtl()
}
\description{
Partial theme for plotting in right to left languages.
}
\details{
Sets legend left of plot and right-aligns title.
}
\examples{
df2 <- data.frame(x = letters[1:10], y = 1:10, z = rep(c("A", "B"), 5))
ggplot(df2, aes(x, y, col = z)) +
  geom_point() +
  coord_mirror() +
  theme_rtl() +
  ggtitle("Right Aligned Title")
}
\seealso{
\code{\link[=coord_mirror]{coord_mirror()}}
}
