\name{theme_segment}
\alias{theme_segment}
\title{Theme element: segments.}
\usage{theme_segment(colour="black", size=0.5, linetype=1)}

\description{
  Theme element: segments.
}

\details{
  This element draws segments between a set of points
}
\seealso{\code{\link{segmentsGrob}} for underlying grid function,
\code{link{theme_line}}}
\keyword{dplot}
\arguments{
  \item{colour}{line color}
  \item{size}{line size}
  \item{linetype}{line type}
}
