% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-tidy-eval.R
\name{tidyeval-compat}
\alias{tidyeval-compat}
\alias{quo}
\alias{quos}
\alias{quo_name}
\alias{ensym}
\alias{ensyms}
\alias{enexpr}
\alias{enexprs}
\title{Other tidy eval tools}
\description{
These tidy eval tools are no longer recommended for normal usage,
but are still exported for compatibility. See \code{\link[=tidyeval]{?tidyeval}}
for the recommended tools.
}
\keyword{internal}
