\name{tr}
\alias{tr}
\title{Traceback alias}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Alias of traceback with fewer keypresses, and severe restriction on number of lines for each function
}
\usage{tr(x = NULL)}
\arguments{
\item{x}{}
}

\details{}

\examples{}
\keyword{manip}
\keyword{internal}
