\name{transform_position}
\alias{transform_position}
\title{Convenience function to ensure that all position variables}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
(x, xmin, xmax, xend) are transformed in the same way
}
\usage{transform_position(df, trans_x = NULL, trans_y = NULL)}
\arguments{
\item{df}{}
\item{trans_x}{}
\item{trans_y}{}
}

\details{}

\examples{}

