% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-.r
\name{transform_position}
\alias{transform_position}
\title{Convenience function to transform all position variables.}
\usage{
transform_position(df, trans_x = NULL, trans_y = NULL, ...)
}
\arguments{
\item{trans_x, trans_y}{Transformation functions for x and y aesthetics.
(will transform x, xmin, xmax, xend etc)}

\item{...}{Additional arguments passed to \code{trans_x} and \code{trans_y}.}
}
\description{
Convenience function to transform all position variables.
}
\keyword{internal}
