% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-point.R
\name{translate_shape_string}
\alias{translate_shape_string}
\title{Translating shape strings}
\usage{
translate_shape_string(shape_string)
}
\arguments{
\item{shape_string}{A character vector giving point shapes.}
}
\value{
An integer vector with translated shapes.
}
\description{
\code{translate_shape_string()} is a helper function for translating point shapes
given as a character vector into integers that are interpreted by the
grid system.
}
\examples{
translate_shape_string(c("circle", "square", "triangle"))

# Strings with 1 or less characters are interpreted as symbols
translate_shape_string(c("a", "b", "?"))
}
\keyword{internal}
