\name{try_require}
\alias{try_require}
\title{Quietly try to require a package}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Queitly require a package, returning an error message if that package is not installed.
}
\usage{try_require(package)}
\arguments{
\item{package}{name of package}
}




\keyword{internal}
