\name{tryapply}
\alias{tryapply}
\alias{tryNULL}
\title{Apply with built in try}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Uses compact, lapply and tryNULL
}
\usage{tryapply(list, fun, ...)}
\arguments{
\item{list}{}
\item{fun}{}
\item{...}{}
}

\details{}

\examples{}
\keyword{internal}
