\name{uneval}
\alias{uneval}
\title{Uneval}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Convert an unevaluted list to a list of unevaluated objects
}
\usage{uneval(x)}
\arguments{
\item{x}{unevaluated list (create with substitute)}
}

\details{}

\examples{}
\keyword{manip}
\keyword{internal}
