\name{update_stat_defaults}
\alias{update_stat_defaults}
\title{Update geom defaults}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Modify geom aesthetic defaults for future plots
}
\usage{update_stat_defaults(geom, new)}
\arguments{
\item{geom}{}
\item{new}{}
}




\keyword{name of geom to modify}
\keyword{named list of aesthetics}
\keyword{hplot}
