% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-elements.r
\name{validate_element}
\alias{validate_element}
\title{Check that an element object has the proper class}
\usage{
validate_element(el, elname)
}
\arguments{
\item{el}{an element}

\item{elname}{the name of the element}
}
\description{
Given an element object and the name of the element, this function
checks it against the element inheritance tree to make sure the
element is of the correct class
}
\details{
It throws error if invalid, and returns invisible() if valid.
}

