% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-discrete-.r
\name{vec_arith.ggplot2_mapped_discrete}
\alias{vec_arith.ggplot2_mapped_discrete}
\title{Utilities for working with discrete values mapped to numeric domain}
\usage{
\method{vec_arith}{ggplot2_mapped_discrete}(op, x, y, ...)
}
\arguments{
\item{op}{The operator to apply}

\item{x, y}{items to apply the operator to}

\item{...}{passed on}
}
\description{
Utilities for working with discrete values mapped to numeric domain
}
\keyword{internal}
