\name{vp_name}
\alias{vp_name}
\title{Viewport name}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Compute viewport name
}
\usage{vp_name(row, col, type)}
\arguments{
\item{row}{row index}
\item{col}{column index}
\item{type}{viewport type}
}

\details{This helps ensure a common naming scheme throughout ggplot.}

\examples{}
\keyword{hplot}
\keyword{internal}
