\name{vp_path}
\alias{vp_path}
\title{Viewport path}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Calculate viewport path.
}
\usage{vp_path(row, col, type)}
\arguments{
\item{row}{row index}
\item{col}{column index}
\item{type}{viewport type}
\item{}{row index}
\item{}{column index}
\item{}{viewport type}
\item{}{viewport type}
\item{}{number of rows}
\item{}{number of columns}
\item{}{optional data to compute rows and columns from}
\item{}{offset from top and left}
\item{}{list containing x and y ranges}
}

\details{Convience method for calculating the viewport path to a particular
entry in a matrix viewport.  This helps ensure a common naming scheme throughout
ggplot/

@arguments row index
@arguments column index
@arguments viewport type
@keyword hplot
@keyword internal
Viewport name
Compute viewport name

This helps ensure a common naming scheme throughout ggplot.

@arguments row index
@arguments column index
@arguments viewport type
@keyword hplot
@keyword internal
Setup viewports
Setup matrix of viewports for a layout with given parameters}

\examples{}
\keyword{hplot}
\keyword{internal}
\keyword{hplot}
\keyword{internal}
\keyword{hplot}
\keyword{internal}
