\name{wrap_hmisc}
\alias{wrap_hmisc}
\alias{sum_mean_cl_boot}
\alias{sum_mean_cl_normal}
\alias{sum_mean_sdl}
\alias{sum_median_hilow}
\alias{sum_range}
\alias{mean_cl_boot}
\alias{mean_cl_normal}
\alias{mean_sdl}
\alias{median_hilow}
\title{Wrap Hmisc summary functions}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Wrap up a selection of Hmisc to make it easy to use them with \code{\link{stat_summary}}
}
\usage{wrap_hmisc(x, fun, ...)}
\arguments{
\item{x}{}
\item{fun}{}
\item{...}{}
}

\details{}

\examples{}
\keyword{internal}
