% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-wrap.R
\name{wrap_dims}
\alias{wrap_dims}
\title{Arrange 1d structure into a grid}
\usage{
wrap_dims(n, nrow = NULL, ncol = NULL)
}
\arguments{
\item{n}{length of structure}

\item{nrow, ncol}{desired dimensions for the grid}
}
\value{
the grid dimension as a vector with nrow and then ncol
}
\description{
Arrange 1d structure into a grid
}
\keyword{internal}
