\name{wrap_hmisc}
\title{Wrap up a selection of summary functions from Hmisc to make it easy to use...}
\usage{wrap_hmisc(fun)}

\description{
  Wrap up a selection of summary functions from Hmisc to
  make it easy to use with \code{\link{stat_summary}}
}

\details{
  See the Hmisc documentation for details of their options.
}
\seealso{\code{\link[Hmisc]{smean.cl.boot}},
\code{\link[Hmisc]{smean.cl.normal}}, \code{\link[Hmisc]{smean.sdl}},
\code{\link[Hmisc]{smedian.hilow}}}
\alias{mean_cl_boot}
\alias{mean_cl_normal}
\alias{mean_sdl}
\alias{median_hilow}
