\name{zeroGrob}
\alias{zeroGrob}
\alias{widthDetails.zeroGrob}
\alias{heightDetails.zeroGrob}
\alias{grobWidth.zeroGrob}
\alias{grobHeight.zeroGrob}
\alias{drawDetails.zeroGrob}
\alias{is.zero}
\title{Zero grob}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
The zero grob draws nothing and has zero size.
}
\usage{zeroGrob()}
\arguments{

}




\keyword{internal}
