\name{zlim}
\alias{zlim}
\title{Set z limits}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Convenience function to set the limits of the z axis.
}
\usage{zlim(min=NA, max=NA)}
\arguments{
\item{min}{lower limit}
\item{max}{upper limit}
}

\details{Works by creating a new continuous scale, so will only work for
continuous variables.}

\examples{}
\keyword{hplot}
