% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.r
\name{aes_all}
\alias{aes_all}
\title{Given a character vector, create a set of identity mappings}
\usage{
aes_all(vars)
}
\arguments{
\item{vars}{vector of variable names}
}
\description{
Given a character vector, create a set of identity mappings
}
\examples{
aes_all(names(mtcars))
aes_all(c("x", "y", "col", "pch"))
}
\keyword{internal}

