% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.R
\name{aes_auto}
\alias{aes_auto}
\title{Automatic aesthetic mapping}
\usage{
aes_auto(data = NULL, ...)
}
\arguments{
\item{data}{data.frame or names of variables}

\item{...}{aesthetics that need to be explicitly mapped.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
