% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.r
\name{aes_auto}
\alias{aes_auto}
\title{Automatic aesthetic mapping}
\usage{
aes_auto(data = NULL, ...)
}
\arguments{
\item{data}{data.frame or names of variables}

\item{...}{aesthetics that need to be explicitly mapped.}
}
\description{
Automatic aesthetic mapping
}
\keyword{internal}
