% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes-colour-fill-alpha.r
\name{aes_colour_fill_alpha}
\alias{aes_colour_fill_alpha}
\alias{colour}
\alias{color}
\alias{fill}
\alias{Learn}
\alias{more}
\alias{about}
\alias{setting}
\alias{these}
\alias{aesthetics}
\alias{in}
\alias{`vignette("ggplot2-specs")`.}
\title{Colour related aesthetics: colour, fill and alpha}
\description{
This page demonstrates the usage of a sub-group
of aesthetics: \code{colour}, \code{fill} and \code{alpha}.
}
\details{
The \code{colour} aesthetic changes the colour of a line (or outline),
\code{fill} changes the interior colouring of a geom, \code{alpha}
changes its transparency.

\code{colour} and \code{fill} aesthetics can be set with a color name or
an rgb specification. The alpha aesthetic value goes from 0 to 1,
where 0 is completely transparent and 1 is not transparent at all.
}
\examples{
\donttest{

# Bar chart example
c <- ggplot(mtcars, aes(factor(cyl)))
# Default plotting
c + geom_bar()
# To change the interior colouring use fill aesthetic
c + geom_bar(fill = "red")
# Compare with the colour aesthetic which changes just the bar outline
c + geom_bar(colour = "red")
# Combining both, you can see the changes more clearly
c + geom_bar(fill = "white", colour = "red")
# Both colour and fill can take an rgb specification.
c + geom_bar(fill = "#00abff")
# Use NA for a completely transparent colour.
c + geom_bar(fill = "NA", colour = "#00abff")

# The aesthetic fill also takes different colouring scales
# setting fill equal to a factor variable uses a discrete colour scale
k <- ggplot(mtcars, aes(factor(cyl), fill = factor(vs)))
k + geom_bar()

# Fill aesthetic can also be used with a continuous variable
m <- ggplot(faithfuld, aes(waiting, eruptions))
m + geom_raster()
m + geom_raster(aes(fill = density))

# Some geoms don't use both aesthetics (i.e. geom_point or geom_line)
b <- ggplot(economics, aes(x = date, y = unemploy))
b + geom_line()
b + geom_line(colour = "green")
b + geom_point()
b + geom_point(colour = "red")

# For large datasets with overplotting the alpha
# aesthetic will make the points more transparent.
df <- data.frame(x = rnorm(5000), y = rnorm(5000))
h  <- ggplot(df, aes(x,y))
h + geom_point()
h + geom_point(alpha = 0.5)
h + geom_point(alpha = 1/10)

# Alpha can also be used to add shading
j <- b + geom_line()
j
yrng <- range(economics$unemploy)
j <- j + geom_rect(aes(NULL, NULL, xmin = start, xmax = end, fill = party),
ymin = yrng[1], ymax = yrng[2], data = presidential)
j
j + scale_fill_manual(values = alpha(c("blue", "red"), .3))
}
}
