\name{annotate}
\alias{annotate}
\title{Annotate a plot}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Add annotations to a plot in a convenient manner
}
\usage{annotate(geom, x = NULL, y = NULL, xmin = NULL, xmax = NULL, ymin = NULL, ymax = NULL, ...)}
\arguments{
\item{geom}{}
\item{x}{}
\item{y}{}
\item{xmin}{}
\item{xmax}{}
\item{ymin}{}
\item{ymax}{}
\item{...}{}
}



\examples{annotate("text", x = 0, y = 0, label = "title")}
\keyword{internal}
