% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autolayer.R
\name{autolayer}
\alias{autolayer}
\title{Create a ggplot layer appropriate to a particular data type}
\usage{
autolayer(object, ...)
}
\arguments{
\item{object}{an object, whose class will determine the behaviour of autolayer}

\item{...}{other arguments passed to specific methods}
}
\value{
a ggplot layer
}
\description{
\code{autolayer()} uses ggplot2 to draw a particular layer for an object of a
particular class in a single command. This defines the S3 generic that
other classes and packages can extend.
}
\seealso{
\code{\link[=autoplot]{autoplot()}}, \code{\link[=ggplot]{ggplot()}} and \code{\link[=fortify]{fortify()}}
}
