\name{axis_vp_path}
\alias{axis_vp_path}
\title{Axis viewport path}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Compute viewport path for specified component of axis
}
\usage{axis_vp_path(position, name)}
\arguments{
\item{position}{position of axis}
\item{name}{component name}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
