% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench.r
\name{benchplot}
\alias{benchplot}
\title{Benchmark plot creation time.
Broken down into construct, build, render and draw times.}
\usage{
benchplot(x)
}
\arguments{
\item{x}{code to create ggplot2 plot}
}
\description{
Benchmark plot creation time.
Broken down into construct, build, render and draw times.
}
\examples{
benchplot(ggplot(mtcars, aes(mpg, wt)) + geom_point())
benchplot(ggplot(mtcars, aes(mpg, wt)) + geom_point() + facet_grid(. ~ cyl))
}
\keyword{internal}

