\name{bin}
\alias{bin}
\title{}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
if (binwidth < resolution(x)) warning("Binwidth is smaller than the resolution of the data")
}
\usage{bin(x, weight=NULL, binwidth=NULL, origin=NULL, breaks=NULL, range=NULL, width=0.9)}
\arguments{
\item{x}{range}
\item{weight}{interval size}
\item{binwidth}{}
\item{origin}{}
\item{breaks}{}
\item{range}{}
\item{width}{}
}

\details{if (binwidth < resolution(x)) warning("Binwidth is smaller than the resolution of the data")
Need to leave zeros in for non-bar representations
results <- subset(results, count > 0)
Generate sequence of fixed size intervals covering range
All locations are multiples of size}
\seealso{\code{\link{round_any}}}
\examples{}

