\name{bin}
\alias{bin}
\title{Bin data}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
This function powers \code{\link{stat_bin}}R
}
\usage{bin(x, weight=NULL, binwidth=NULL, origin=NULL, breaks=NULL, range=NULL, width=0.9)}
\arguments{
\item{x}{}
\item{weight}{}
\item{binwidth}{}
\item{origin}{}
\item{breaks}{}
\item{range}{}
\item{width}{}
}

\details{}

\examples{}
\keyword{internal}
