\name{check_breaks_and_labels}
\alias{check_breaks_and_labels}
\title{Check breaks and labels.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Ensure that breaks and labels are of the correct form
}
\usage{check_breaks_and_labels(breaks = NULL, labels = NULL)}
\arguments{
\item{breaks}{}
\item{labels}{}
}




\keyword{internal}
