\name{check_required_aesthetics}
\alias{check_required_aesthetics}
\title{Check required aesthetics are present}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
This is used by geoms and stats to give a more helpful error message
}
\usage{check_required_aesthetics(required, present, name)}
\arguments{
\item{required}{character vector of required aesthetics}
\item{present}{character vector of present aesthetics}
\item{name}{name of object for error message}
}




\keyword{internal}
