\name{check_required_aesthetics}
\alias{check_required_aesthetics}
\title{Check required aesthetics are presented}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{

}
\usage{check_required_aesthetics(required, present, name)}
\arguments{
\item{required}{}
\item{present}{}
\item{name}{}
}

\details{}

\examples{}
\keyword{internal}
