\name{chop}
\alias{chop}
\title{Chop}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Chop a continuous variable into a categorical variable.
}
\usage{chop(x, n=5, method="quantiles", midpoint=0, digits=2)}
\arguments{
\item{x}{}
\item{n}{}
\item{method}{}
\item{midpoint}{}
\item{digits}{}
}

\details{Chop provides a convenient interface to the main methods of
converting a continuous variable into a categorical variable.}
\seealso{\code{\link{chop.breaks}} to get breaks used}
\examples{}
\keyword{manip}
