\name{chop.breaks}
\alias{chop.breaks}
\title{Chop breaks}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Calculate breakpoints for chop function
}
\usage{chop.breaks(x, n, method, midpoint=NA)}
\arguments{
\item{x}{}
\item{n}{}
\item{method}{}
\item{midpoint}{}
}

\details{}

\examples{}
\keyword{manip}
\keyword{internal}
