\name{col2hcl}
\alias{col2hcl}
\title{Modify standard R colour in hcl colour space}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Transforms rgb to hcl, sets non-missing arguments and then backtransforms to rgb
}
\usage{col2hcl(colour, h, c, l, alpha = 1)}
\arguments{
\item{colour}{}
\item{h}{}
\item{c}{}
\item{l}{}
\item{alpha}{}
}

\details{@keyword internal
Mute standard R colours.
This produces colours with moderate luminance and saturation.s}

\examples{}
\keyword{internal}
\keyword{internal}
