\name{continuous_scale}
\alias{continuous_scale}
\title{Continuous scale constructor.}
\usage{
  continuous_scale(aesthetics, scale_name, palette, name =
  NULL, breaks = NULL, labels = NULL, legend = TRUE, limits
  = NULL, rescaler = rescale, oob = censor, expand = c(0,
  0), na.value = NA, trans = "identity")
}
\arguments{
  \item{aesthetics}{character}
}
\description{
  Continuous scale constructor.
}
\keyword{internal}

