% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-munch.r
\name{coord_munch}
\alias{coord_munch}
\title{Munch coordinates data}
\usage{
coord_munch(coord, data, range, segment_length = 0.01)
}
\arguments{
\item{coord}{Coordinate system definition.}

\item{data}{Data set to transform - should have variables \code{x} and
\code{y} are chopped up into small pieces (as defined by \code{group}).
All other variables are duplicated as needed.}

\item{range}{Panel range specification.}

\item{segment_length}{Target segment length}
}
\description{
This function "munches" lines, dividing each line into many small pieces
so they can be transformed independently. Used inside geom functions.
}
\keyword{internal}

