% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities-matrix.r
\name{cweave}
\alias{cweave}
\title{Col weave}
\usage{
cweave(...)
}
\arguments{
\item{...}{matrices to weave together}
}
\description{
Weave together two (or more) matrices by column
}
\details{
Matrices must have same dimensions
}
\keyword{internal}

