\name{dir_diff}
\alias{dir_diff}
\title{Compute the set of differences in file make up between two directories.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{

}
\usage{dir_diff(a, b)}
\arguments{
\item{a}{path a}
\item{b}{path b}
}
\value{list with components only_a, only_b and both}
\details{}

\examples{}
\keyword{internal}
