\name{dir_diff}
\alias{dir_diff}
\title{Directory diff}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Compute the set of differences in file make up between two directories.
}
\usage{dir_diff(a, b)}
\arguments{
\item{a}{path a}
\item{b}{path b}
}
\value{list with components only\_a, only\_b and both}



\keyword{internal}
