\name{discrete_scale}
\alias{discrete_scale}
\title{Discrete scale constructor.}
\usage{
  discrete_scale(aesthetics, scale_name, palette,
    name = NULL, breaks = waiver(), labels = waiver(),
    legend = NULL, limits = NULL, expand = c(0, 0),
    na.value = NA, drop = TRUE, guide = "legend")
}
\description{
  Discrete scale constructor.
}
\keyword{internal}

