% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legend-draw.r
\name{draw_key}
\alias{draw_key}
\alias{legend_abline}
\alias{legend_blank}
\alias{legend_boxplot}
\alias{legend_crossbar}
\alias{legend_dotplot}
\alias{legend_path}
\alias{legend_point}
\alias{legend_pointrange}
\alias{legend_polygon}
\alias{legend_smooth}
\alias{legend_text}
\alias{legend_vline}
\title{Key drawing functions}
\usage{
legend_point(self, data, params)

legend_abline(self, data, params)

legend_polygon(self, data, params)

legend_blank(self, data, params)

legend_boxplot(self, data, params)

legend_crossbar(self, data, params)

legend_path(self, data, params)

legend_dotplot(self, data, params)

legend_pointrange(self, data, params)

legend_smooth(self, data, params)

legend_text(self, data, params)

legend_vline(self, data, params)
}
\arguments{
\item{self}{The Geom object}

\item{data}{A single row data frame containing the scaled aesthetics to
display in this key}

\item{params}{A list of additional parameters supplied to the geom.}
}
\value{
A grid grob.
}
\description{
Each Geom has an associated function that draws the key when the geom needs
to be displayed in a legend. These are the options built into ggplot2.
}
\keyword{internal}

