\name{drop_calculated_aes}
\alias{drop_calculated_aes}
\title{Drop calculated aesthetics}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Remove aesthetics mapped to variables created by statistics
}
\usage{drop_calculated_aes(aesthetics)}
\arguments{
\item{aesthetics}{}
}

\details{}

\examples{}
\keyword{internal}
