\name{US economic time series}
\docType{data}
\alias{economics}
\title{US economic time series}
\description{

This dataset was produced from US economic time series data available from \url{http://research.stlouisfed.org/fred2}.

\begin{itemize}
	\item date.  Month of data collection

	\item psavert, personal savings rate, \url{http://research.stlouisfed.org/fred2/series/PSAVERT/}
	\item pce, personal consumption expenditures, in billions of dollars, \url{http://research.stlouisfed.org/fred2/series/PCE}
	\item unemploy, number of unemployed in thousands, \url{http://research.stlouisfed.org/fred2/series/UNEMPLOY}
	\item uempmed, median duration of unemployment, in week, \url{http://research.stlouisfed.org/fred2/series/UEMPMED}
	\item pop, total population, in thousands, \url{http://research.stlouisfed.org/fred2/series/POP}

\end{itemize}

}
\usage{data(economics)}
\format{A data frame with 478 rows and 6 variables}
\keyword{datasets}
