% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{economics}
\alias{economics}
\alias{economics_long}
\title{US economic time series}
\format{
A data frame with 574 rows and 6 variables:
\describe{
\item{date}{Month of data collection}
\item{pce}{personal consumption expenditures, in billions of dollars,
\url{https://fred.stlouisfed.org/series/PCE}}
\item{pop}{total population, in thousands,
\url{https://fred.stlouisfed.org/series/POP}}
\item{psavert}{personal savings rate,
\url{https://fred.stlouisfed.org/series/PSAVERT/}}
\item{uempmed}{median duration of unemployment, in weeks,
\url{https://fred.stlouisfed.org/series/UEMPMED}}
\item{unemploy}{number of unemployed in thousands,
\url{https://fred.stlouisfed.org/series/UNEMPLOY}}
}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2870 rows and 4 columns.
}
\usage{
economics

economics_long
}
\description{
This dataset was produced from US economic time series data available from
\url{https://fred.stlouisfed.org/}. \code{economics} is in "wide"
format, \code{economics_long} is in "long" format.
}
\keyword{datasets}
