% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-elements.r
\name{el_def}
\alias{el_def}
\title{Define an element's class and what other elements it inherits from}
\usage{
el_def(class = NULL, inherit = NULL, description = NULL)
}
\arguments{
\item{class}{The name of class (like "element_line", "element_text",
or the reserved "character", which means a character vector (not
"character" class)}

\item{inherit}{A vector of strings, naming the elements that this
element inherits from.}

\item{description}{The optional description of the particular element.}
}
\description{
Define an element's class and what other elements it inherits from
}

