\name{element_line}
\alias{element_line}
\title{Theme element: line.}
\usage{
  element_line(colour = NULL, size = NULL, linetype = NULL,
    lineend = NULL, color = NULL)
}
\arguments{
  \item{colour}{line colour}

  \item{size}{line size}

  \item{linetype}{line type}

  \item{lineend}{line end}

  \item{color}{an alias for \code{colour}}
}
\description{
  Theme element: line.
}

