% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/theme-elements.r
\name{element_text}
\alias{element_text}
\title{Theme element: text.}
\usage{
element_text(family = NULL, face = NULL, colour = NULL, size = NULL,
  hjust = NULL, vjust = NULL, angle = NULL, lineheight = NULL,
  color = NULL)
}
\arguments{
\item{family}{font family}

\item{face}{font face ("plain", "italic", "bold", "bold.italic")}

\item{colour}{text colour}

\item{size}{text size (in pts)}

\item{hjust}{horizontal justification (in [0, 1])}

\item{vjust}{vertical justification (in [0, 1])}

\item{angle}{angle (in [0, 360])}

\item{lineheight}{line height}

\item{color}{an alias for \code{colour}}
}
\description{
Theme element: text.
}

