% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-expantion.r
\name{expand_range4}
\alias{expand_range4}
\title{Expand a numeric range}
\usage{
expand_range4(limits, expand)
}
\arguments{
\item{limits}{A numeric vector of length 2 giving the
range to expand.}

\item{expand}{A numeric vector of length 2 (\code{c(add, mult)})
or length 4 (\code{c(mult_left, add_left, mult_right, add_right)}),
as generated by \code{\link[=expand_scale]{expand_scale()}}.}
}
\value{
The expanded \code{limits}
}
\description{
Expand a numeric range
}
