\name{facet}
\alias{facet}
\title{Facet specification.}
\usage{
  facet(..., shrink = TRUE, subclass = c())
}
\arguments{
  \item{...}{object fields}

  \item{shrink}{shrink scales to fit output of statistics,
  not raw data}
}
\description{
  Facet specification.
}
\details{
  Create new facetting specification.  For internal use
  only.
}
\keyword{internal}

