% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-.r
\name{facet}
\alias{facet}
\title{Facet specification.}
\usage{
facet(..., shrink = TRUE, subclass = c())
}
\arguments{
\item{...}{object fields}

\item{shrink}{shrink scales to fit output of statistics, not raw data}
}
\description{
Create new facetting specification.  For internal use only.
}
\keyword{internal}

