\name{fortify}
\alias{fortify}
\title{Fortify a model with data}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Generic method to supplement the original data with model fit statistics
}
\usage{fortify(model, data, ...)}
\arguments{
\item{model}{}
\item{data}{}
\item{...}{}
}

\details{}
\seealso{\code{\link{fortify.lm}}}
\examples{}

