% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.R
\name{fortify}
\alias{fortify}
\title{Fortify a model with data.}
\usage{
fortify(model, data, ...)
}
\arguments{
\item{model}{model or other R object to convert to data frame}

\item{data}{original dataset, if needed}

\item{...}{Arguments passed to methods.}
}
\description{
Rather than using this function, I now recommend using the \pkg{broom}
package, which implements a much wider range of methods. \code{fortify()}
may be deprecated in the future.
}
\seealso{
\code{\link[=fortify.lm]{fortify.lm()}}

Other plotting automation topics: 
\code{\link{autolayer}()},
\code{\link{automatic_plotting}},
\code{\link{autoplot}()}
}
\concept{plotting automation topics}
