\name{fullseq}
\alias{fullseq}
\title{Generate sequence of fixed size intervals covering range}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
All locations are multiples of size
}
\usage{fullseq(range, size, pad = FALSE)}
\arguments{
\item{range}{range}
\item{size}{interval size}
\item{pad}{}
}


\seealso{\code{\link{reshape}{round_any}}}

\keyword{internal}
