% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom-.r
\docType{data}
\name{Geom}
\alias{Geom}
\title{The base geom proto object}
\format{A proto object}
\usage{
Geom
}
\description{
If you are creating a new geom in another package, you'll need to extend
from this object.
}
\keyword{internal}

