\name{geom_contour}
\alias{geom_contour}
\title{Display contours of a 3d surface in 2d.}
\usage{geom_contour(mapping = NULL, data = NULL, stat = "contour",
    position = "identity", lineend = "butt", linejoin =
    "round", linemitre = 1, na.rm = FALSE, ...)}

\description{
  Display contours of a 3d surface in 2d.
}
\seealso{\code{\link{geom_density2d}}: 2d density contours}
\examples{# See stat_contour for examples}

