% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom-point-count.r
\name{geom_count}
\alias{geom_count}
\title{Count the number of observations at each location.}
\usage{
geom_count(mapping = NULL, data = NULL, stat = "sum",
  position = "identity", na.rm = FALSE, show_guide = NA, ...)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link{aes}} or \code{\link{aes_string}}. Only needs to be set
at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override
the plot defaults.}

\item{stat}{The statistical transformation to use on the data for this
layer.}

\item{position}{The position adjustment to use for overlapping points
on this layer}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with
a warning.  If \code{TRUE} silently removes missing values.}

\item{show_guide}{logical. Should a legend be drawn?}

\item{...}{other arguments passed on to \code{\link{layer}}. Typically,
these are aesthetics whose values you want to set, not map. e.g.
\code{color = "red"} or \code{size = 3}.}
}
\description{
This is a variant \code{\link{geom_point}} that counts the number of
observations at each location, then maps the count to point size. It
useful when you have discrete data.
}
\section{Aesthetics}{

\Sexpr[results=rd,stage=build]{ggplot2:::rd_aesthetics("geom", "point")}
}
\examples{
ggplot(mpg, aes(cty, hwy)) +
 geom_point()

ggplot(mpg, aes(cty, hwy)) +
 geom_count()

# Best used in conjunction with scale_size_area which ensures that
# counts of zero would be given size 0. Doesn't make much different
# here because the smallest count is already close to 0.
ggplot(mpg, aes(cty, hwy)) +
 geom_count()
 scale_size_area()
}

