\name{geom_crossbar}
\alias{geom_crossbar}
\title{Hollow bar with middle indicated by horizontal line.}
\usage{
  geom_crossbar(mapping = NULL, data = NULL, stat =
  "identity", position = "identity", fatten = 2, ...)
}
\arguments{
  \item{fatten}{a multiplicate factor to fatten middle bar
  by}
}
\description{
  Hollow bar with middle indicated by horizontal line.
}
\examples{
# See geom_linerange for examples
}
\seealso{
  \code{\link{geom_errorbar}} for error bars,
  \code{\link{geom_pointrange}} and
  \code{\link{geom_linerange}} for other ways of showing
  mean + error, \code{\link{stat_summary}} to compute
  errors from the data, \code{\link{geom_smooth}} for the
  continuous analog.
}

