% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom-curve.r
\name{geom_curve}
\alias{geom_curve}
\title{Single curved line segments.}
\usage{
geom_curve(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", curvature = 1, angle = 90, ncp = 1,
  arrow = NULL, lineend = "butt", na.rm = FALSE, ...)
}
\arguments{
\item{curvature}{A numeric value giving the amount of curvature. Negative values produce left-hand curves, positive values produce right-hand curves, and zero produces a straight line. (see curveGrob)}

\item{angle}{A numeric value between 0 and 180, giving an amount to skew the control points of the curve. Values less than 90 skew the curve towards the start point and values greater than 90 skew the curve towards the end point. (see curveGrob)}

\item{ncp}{The number of control points used to draw the curve. More control points creates a smoother curve. (see curveGrob)}

\item{arrow}{specification for arrow heads, as created by arrow()}

\item{lineend}{Line end style (round, butt, square)}

\item{...}{Arguments to be passed to \code{curveGrob}.}
}
\description{
Single curved line segments.
}
\section{Aesthetics}{

\Sexpr[results=rd,stage=build]{ggplot2:::rd_aesthetics("geom", "curve")}
}
\examples{
# Adding curve segments
library(grid) # needed for arrow function
b <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
b + geom_curve(aes(x = 2, y = 15, xend = 2, yend = 25), curvature = 0.2)
b + geom_curve(aes(x = 2, y = 15, xend = 3, yend = 15), ncp = 2)
b + geom_curve(aes(x = 5, y = 30, xend = 3.5, yend = 25), arrow = arrow(length = unit(0.5, "cm")))
}
\seealso{
\code{\link{geom_segment}}, \code{\link{geom_path}} and \code{\link{geom_line}} for multi-
segment lines and paths.
}

