\name{geom_density}
\alias{geom_density}
\title{Display a smooth density estimate.}
\usage{
  geom_density(mapping = NULL, data = NULL, stat =
  "density", position = "identity", na.rm = FALSE, ...)
}
\description{
  A smooth density estimate calculated by
  \code{\link{stat_density}}.
}
\examples{
# See stat_density for examples
}
\seealso{
  \code{\link{geom_histogram}} for the histogram
}

