\name{geom_density}
\alias{geom_density}
\title{Display a smooth density estimate.}

\description{
  Display a smooth density estimate.
}

\details{
  A smooth density estimate calculated by
  \code{\link{stat_density}}.
}
\seealso{\code{\link{geom_histogram}} for the histogram}
\examples{# See stat_density for examples}
