% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom-freqpoly.r
\name{geom_freqpoly}
\alias{geom_freqpoly}
\title{Frequency polygon.}
\usage{
geom_freqpoly(mapping = NULL, data = NULL, stat = "bin",
  position = "identity", show_guide = NA, ...)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link{aes}} or \code{\link{aes_string}}. Only needs to be set
at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override
the plot defaults.}

\item{stat}{The statistical transformation to use on the data for this
layer.}

\item{position}{The position adjustment to use for overlapping points
on this layer}

\item{show_guide}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{...}{other arguments passed on to \code{\link{layer}}. Typically,
these are aesthetics whose values you want to set, not map. e.g.
\code{color = "red"} or \code{size = 3}.}
}
\description{
Frequency polygon.
}
\section{Aesthetics}{

\Sexpr[results=rd,stage=build]{ggplot2:::rd_aesthetics("geom", "freqpoly")}
}
\examples{
m <- ggplot(diamonds, aes(carat))
m + geom_freqpoly(binwidth = 0.1)
m + geom_freqpoly(binwidth = 0.01)

p <- ggplot(diamonds, aes(price))
p + geom_freqpoly(binwidth = 1000)
p + geom_freqpoly(aes(colour = color), binwidth = 1000)
p + geom_freqpoly(aes(y = ..density.., colour = color),
                  binwidth = 1000)
}
\seealso{
\code{\link{geom_histogram}}: histograms
}

