\name{geom_hex}
\alias{geom_hex}
\alias{GeomHex}
\title{geom\_hex}
\description{}
\details{
This page describes geom\_hex, see \code{\link{layer}} and \code{\link{qplot}} for how to create a complete plot from individual components.
}
\section{Aesthetics}{
The following aesthetics can be used with geom\_hex.  Aesthetics are mapped to variables in the data with the \code{\link{aes}} function: \code{geom\_hex(\code{\link{aes}}(x = var))}
\itemize{
  \item \code{x}: x position (\strong{required}) 
  \item \code{y}: y position (\strong{required}) 
  \item \code{colour}: border colour 
  \item \code{fill}: internal colour 
  \item \code{size}: size 
}
}
\usage{geom_hex(mapping=NULL, data=NULL, stat="binhex", position="identity", ...)}
\arguments{
 \item{mapping}{mapping between variables and aesthetics generated by aes}
 \item{data}{dataset used in this layer, if not specified uses plot dataset}
 \item{stat}{statistic used by this layer}
 \item{position}{position adjustment used by this layer}
 \item{...}{ignored }
}
\seealso{\itemize{
  \item \url{http://had.co.nz/ggplot2/geom_hex.html}
}}
\value{A \code{\link{layer}}}
\examples{\dontrun{
  # Coming soon
}}
\author{Hadley Wickham, \url{http://had.co.nz/}}
\keyword{hplot}
