% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geom-hex.r
\name{geom_hex}
\alias{geom_hex}
\title{Hexagon bining.}
\usage{
geom_hex(mapping = NULL, data = NULL, stat = "binhex",
  position = "identity", ...)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link{aes}} or \code{\link{aes_string}}. Only needs to be set
at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override
the plot defaults.}

\item{stat}{The statistical transformation to use on the data for this
layer.}

\item{position}{The position adjustment to use for overlapping points
on this layer}

\item{...}{other arguments passed on to \code{\link{layer}}. This can
include aesthetics whose values you want to set, not map. See
\code{\link{layer}} for more details.}
}
\description{
Hexagon bining.
}
\section{Aesthetics}{

\Sexpr[results=rd,stage=build]{ggplot2:::rd_aesthetics("geom", "hex")}
}
\examples{
# See ?stat_binhex for examples
}

