% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom-km.r
\name{geom_km}
\alias{geom_km}
\title{Add a Kaplan-Meier survival curve}
\usage{
geom_km(mapping = NULL, data = NULL, stat = "km", position = "identity",
  ...)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link{aes}} or \code{\link{aes_string}}. Only needs to be set
at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override
the plot defaults.}

\item{stat}{The statistical transformation to use on the data for this
layer.}

\item{position}{The position adjustment to use for overlapping points
on this layer}

\item{...}{other arguments passed on to \code{\link{layer}}. This can
include aesthetics whose values you want to set, not map. See
\code{\link{layer}} for more details.}
}
\description{
Add a Kaplan-Meier survival curve
}
\section{Aesthetics}{

\Sexpr[results=rd,stage=build]{ggplot2:::rd_aesthetics("geom", "km")}
}
\examples{
sex <- rbinom(250, 1, .5)
df <- data.frame(time = exp(rnorm(250, mean = sex)), status = rbinom(250, 1, .75), sex = sex)
ggplot(df, aes(x = time, status = status, color = factor(sex))) + geom_km()
}
\seealso{
The default stat for this geom is \code{\link{stat_km}} see
  that documentation for more options to control the underlying statistical transformation.
}

