\name{geom_pointrange}
\alias{geom_pointrange}
\title{An interval represented by a vertical line, with a point in the middle.}
\usage{
  geom_pointrange(mapping = NULL, data = NULL, stat =
  "identity", position = "identity", ...)
}
\description{
  An interval represented by a vertical line, with a point
  in the middle.
}
\examples{
# See geom_linerange for examples
}
\seealso{
  \code{\link{geom_errorbar}} for error bars,
  \code{\link{geom_linerange}} for range indicated by
  straight line, + examples, \code{\link{geom_crossbar}}
  for hollow bar with middle indicated by horizontal line,
  \code{\link{stat_summary}} for examples of these guys in
  use, \code{\link{geom_smooth}} for continuous analog"
}

