% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom-pointrange.r
\name{geom_pointrange}
\alias{geom_pointrange}
\title{An interval represented by a vertical line, with a point in the middle.}
\usage{
geom_pointrange(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", show_guide = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link{aes}} or \code{\link{aes_string}}. Only needs to be set
at the layer level if you are overriding the plot defaults.}

\item{data}{A data frame. If specified, overrides the default data frame
defined at the top level of the plot.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{show_guide}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{...}{other arguments passed on to \code{\link{layer}}. There are
  three types of arguments you can use here:

  \itemize{
  \item Aesthetics: to set an aesthetic to a fixed value, like
     \code{color = "red"} or \code{size = 3}.
  \item Other arguments to the layer, for example you override the
    default \code{stat} associated with the layer.
  \item Other arguments passed on to the stat.
  }}
}
\description{
An interval represented by a vertical line, with a point in the middle.
}
\section{Aesthetics}{

\Sexpr[results=rd,stage=build]{ggplot2:::rd_aesthetics("geom", "pointrange")}
}
\examples{
# See geom_linerange for examples
}
\seealso{
\code{\link{geom_errorbar}} for error bars,
 \code{\link{geom_linerange}} for range indicated by straight line, + examples,
 \code{\link{geom_crossbar}} for hollow bar with middle indicated by horizontal line,
 \code{\link{stat_summary}} for examples of these guys in use,
 \code{\link{geom_smooth}} for continuous analog"
}

