% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom-quantile.r
\name{geom_quantile}
\alias{geom_quantile}
\title{Add quantile lines from a quantile regression.}
\usage{
geom_quantile(mapping = NULL, data = NULL, stat = "quantile",
  position = "identity", lineend = "butt", linejoin = "round",
  linemitre = 1, na.rm = FALSE, show_guide = NA, ...)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link{aes}} or \code{\link{aes_string}}. Only needs to be set
at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override
the plot defaults.}

\item{stat}{The statistical transformation to use on the data for this
layer.}

\item{position}{The position adjustment to use for overlapping points
on this layer}

\item{lineend}{Line end style (round, butt, square)}

\item{linejoin}{Line join style (round, mitre, bevel)}

\item{linemitre}{Line mitre limit (number greater than 1)}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with
a warning.  If \code{TRUE} silently removes missing values.}

\item{show_guide}{logical. Should a legend be drawn?}

\item{...}{other arguments passed on to \code{\link{layer}}. Typically,
these are aesthetics whose values you want to set, not map. e.g.
\code{color = "red"} or \code{size = 3}.}
}
\description{
This can be used as a continuous analogue of a geom_boxplot.
}
\section{Aesthetics}{

\Sexpr[results=rd,stage=build]{ggplot2:::rd_aesthetics("geom", "quantile")}
}
\examples{
# See stat_quantile for examples
}
\seealso{
See \code{\link{stat_quantile}} for examples.
}

