\name{geom_quantile}
\alias{geom_quantile}
\title{Add quantile lines from a quantile regression.}
\usage{
  geom_quantile(mapping = NULL, data = NULL, stat =
  "quantile", position = "identity", lineend = "butt",
  linejoin = "round", linemitre = 1, na.rm = FALSE, ...)
}
\description{
  This can be used as a continuous analogue of a
  geom_boxplot.
}
\examples{
# See stat_quantile for examples
}

