% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ggorder}
\alias{ggorder}
\title{A plot to investigate the order in which observations were recorded.}
\usage{
ggorder(data)
}
\arguments{
  \item{data}{data set to plot}
}
\description{
A plot to investigate the order in which observations were recorded.
}

