\name{ggorder}
\alias{ggorder}
\title{A plot to investigate the order in which observations were recorded.}
\usage{ggorder(data, scale = "rank")}

\description{
  A plot to investigate the order in which observations
  were recorded.
}
\arguments{
  \item{data}{data set to plot}
  \item{scale}{type of scaling to use.  See \code{\link[reshape]{rescaler}}
for options}
}

