\name{ggplot}
\alias{ggplot}
\alias{ggplot.default}
\title{Create a new plot}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Create a new ggplot plot
}
\usage{ggplot(data = NULL, ...)}
\arguments{
\item{data}{}
\item{...}{}
}

\details{}
\seealso{\url{http://had.co.nz/ggplot2}}
\examples{}
\keyword{hplot}
