\name{ggplot}
\alias{ggplot}
\title{Create a new ggplot plot}
\usage{
  ggplot(data = NULL, ...)
}
\arguments{
  \item{data}{default data set}

  \item{...}{other arguments passed to specific methods}
}
\description{
  Create a new ggplot plot
}
\seealso{
  \url{http://had.co.nz/ggplot2}
}
\keyword{internal}

