% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-.R, R/axis-secondary.R, R/coord-transform.R
\docType{data}
\name{ggplot2-ggproto}
\alias{ggplot2-ggproto}
\alias{AxisSecondary}
\alias{CoordTrans}
\title{Base ggproto classes for ggplot2}
\description{
If you are creating a new geom, stat, position, or scale in another package,
you'll need to extend from \code{ggplot2::Geom}, \code{ggplot2::Stat},
\code{ggplot2::Position}, or \code{ggplot2::Scale}.
}
\seealso{
ggproto
}
\keyword{datasets}
\keyword{internal}
