\name{ggplot2-package}
\alias{ggplot2-package}
\alias{ggplot2}
\docType{package}
\title{
\packageTitle{ggplot2}
}
\description{
\packageDescription{ggplot2}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{ggplot2}

Maintainer: \packageMaintainer{ggplot2}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = ggplot2)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{ggplot2}}
%\section{Documentation Index}{\packageIndices{ggplot2}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
