% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.r
\docType{package}
\name{ggplot2-package}
\alias{ggplot2}
\alias{ggplot2-package}
\title{ggplot2: An Implementation of the Grammar of Graphics}
\description{
ggplot2 is a system for declaratively creating graphics. You provide the
data and tell ggplot2 how to map variables to aesthetics. ggplot2 is a
complete plotting system and is quite different to other graphics systems in
R like base graphics and lattice.  If you've never used ggplot2 before, I
recomend starting with a comprehensive introduction such as
\url{http://r4ds.had.co.nz/data-visualisation.html}. For more
documentation and examples, see see \url{http://ggplot2.org}.
}
\examples{
ggplot(mpg, aes(displ, cty)) +
  geom_point()
}

