% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.r
\docType{package}
\name{ggplot2-package}
\alias{ggplot2}
\alias{ggplot2-package}
\title{ggplot2: An Implementation of the Grammar of Graphics}
\description{
An implementation of the grammar of graphics in R. It combines the
advantages of both base and lattice graphics: conditioning and shared axes
are handled automatically, and you can still build up a plot step by step
from multiple data sources. It also implements a sophisticated
multidimensional conditioning system and a consistent interface to map
data to aesthetic attributes. See http://ggplot2.org for more information,
documentation and examples.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://ggplot2.tidyverse.org}
  \item \url{https://github.com/tidyverse/ggplot2}
  \item Report bugs at \url{https://github.com/tidyverse/ggplot2/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Authors:
\itemize{
  \item Winston Chang \email{winston@rstudio.com}
}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}

