% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.r
\docType{package}
\name{ggplot2-package}
\alias{ggplot2}
\alias{ggplot2-package}
\title{Create Elegant Data Visualisations Using the Grammar of Graphics}
\description{
A system for 'declaratively' creating graphics,
based on "The Grammar of Graphics". You provide the data, tell 'ggplot2'
how to map variables to aesthetics, what graphical primitives to use,
and it takes care of the details.
}
\keyword{internal}

