% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.r
\docType{package}
\name{ggplot2-package}
\alias{ggplot2}
\alias{ggplot2-package}
\title{ggplot2: Create Elegant Data Visualisations Using the Grammar of Graphics}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

A system for 'declaratively' creating graphics,
    based on "The Grammar of Graphics". You provide the data, tell 'ggplot2'
    how to map variables to aesthetics, what graphical primitives to use,
    and it takes care of the details.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://ggplot2.tidyverse.org}
  \item \url{https://github.com/tidyverse/ggplot2}
  \item Report bugs at \url{https://github.com/tidyverse/ggplot2/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Authors:
\itemize{
  \item Winston Chang
  \item Lionel Henry
  \item Thomas Lin Pedersen
  \item Kohske Takahashi
  \item Claus Wilke
  \item Kara Woo
  \item Hiroaki Yutani
}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
