% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggproto.r
\name{ggproto}
\alias{ggproto}
\title{Create a new ggproto object}
\usage{
ggproto(`_class` = NULL, `_inherit` = NULL, ...)
}
\arguments{
\item{_class}{Class name to assign to the object. This is stored as the class
attribute of the object. If \code{NULL} (the default), no class name will
be added to the object.}

\item{_inherit}{ggproto object to inherit from. If \code{NULL}, don't inherit
from any object.}

\item{...}{A list of members in the ggproto object.}
}
\description{
ggproto is inspired by the proto package, but it has some important
differences. Notably, it cleanly supports cross-package inheritance, and has
faster performance.
}
\details{
Methods in ggproto objects can take optional \code{self} and \code{super}
arguments. If the arguments are present, then, when the method is called, it
will be passed the current object (as \code{self}) and the object that it
inherits from (as \code{super}).
}

