% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scale-.r
\name{ggscale}
\alias{ggscale}
\title{Components of a scale:}
\description{
Guide related:
}
\details{
\itemize{
  \item name
  \item breaks
  \item labels
  \item expand
}

Mapping related:
\itemize{
  \item aesthetic
  \item limits
  \item palette
  \item trans
}

Scales are an S3 class with a single mutable component implemented with
a reference class - the range of the data.  This mutability makes working
with scales much easier, because it makes it possible to distribute the
training, without having to worry about collecting all the pieces back
together again.
}
\keyword{internal}

