\name{ggstrip}
\alias{ggstrip}
\title{Grob strip}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Grob for strip labels
}
\usage{ggstrip(text, horizontal=TRUE, theme)}
\arguments{
\item{text}{text to display}
\item{horizontal}{orientation, horizontal or vertical}
\item{theme}{}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
