\name{ggstructure}
\alias{ggstructure}
\title{A plot which aims to reveal gross structural anomalies in the data.}
\usage{
  ggstructure(data)
}
\arguments{
  \item{data}{data set to plot}
}
\description{
  A plot which aims to reveal gross structural anomalies in
  the data.
}
\examples{
ggstructure(mtcars)
}

