% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom-.r
\docType{data}
\name{graphical-units}
\alias{.pt}
\alias{.stroke}
\alias{graphical-units}
\title{Graphical units}
\format{\preformatted{ num 2.83
}}
\usage{
.pt

.stroke
}
\description{
Multiply size in mm by these constants in order to convert to the units
that grid uses internally for \code{lwd} and \code{fontsize}.
}
\keyword{datasets}

