% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-.R
\docType{data}
\name{graphical-units}
\alias{.pt}
\alias{.stroke}
\title{Graphical units}
\format{
An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.
}
\usage{
.pt

.stroke
}
\description{
Multiply size in mm by these constants in order to convert to the units
that grid uses internally for \code{lwd} and \code{fontsize}.
}
\keyword{datasets}
\keyword{internal}
