% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-.r
\docType{data}
\name{graphical-units}
\alias{.pt}
\alias{.stroke}
\alias{graphical-units}
\title{Graphical units}
\format{An object of class \code{numeric} of length 1.}
\usage{
.pt

.stroke
}
\description{
Multiply size in mm by these constants in order to convert to the units
that grid uses internally for \code{lwd} and \code{fontsize}.
}
\keyword{datasets}

