\name{guide_legends}
\alias{guide_legends}
\title{Build all legend grob}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Build legends, merging where possible
}
\usage{guide_legends(scales, usage, theme)}
\arguments{
\item{scales}{}
\item{usage}{}
\item{theme}{}
}
\value{A list of grobs}
\details{}

\examples{}
\keyword{internal}
