% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guides-axis.r
\name{guide_axis}
\alias{guide_axis}
\title{Axis guide}
\usage{
guide_axis(label = TRUE, label.theme = NULL, line.theme = NULL,
  tick.theme = NULL, tick.length = NULL, order = 0,
  position = waiver(), ...)
}
\arguments{
\item{label}{logical. If \code{TRUE} then the labels are drawn. If
\code{FALSE} then the labels are invisible.}

\item{label.theme}{A theme object for rendering the label text. Usually the
object of \code{\link[=element_text]{element_text()}} is expected. By default, the theme is
specified by \code{legend.text} in \code{\link[=theme]{theme()}}.}

\item{line.theme}{An \code{\link[=element_line]{element_line()}} to use as a template for the line
along the axis. Usually set with \link[=theme]{theme(axis.line = ...)}.}

\item{tick.theme}{An \code{\link[=element_line]{element_line()}} to use as a template for the ticks
along the axis. Usually set with \link[=theme]{theme(axis.ticks = ...)}.}

\item{tick.length}{A \code{\link[grid:unit]{grid::unit()}}. Usually set with
\link[=theme]{theme(axis.ticks.length = ...)}.}

\item{order}{positive integer less than 99 that specifies the order of
this guide among multiple guides. This controls the order in which
multiple guides are displayed, not the contents of the guide itself.
If 0 (default), the order is determined by a secret algorithm.}

\item{position}{Where this guide should be drawn: one of top, bottom,
left, or right.}

\item{...}{ignored.}
}
\description{
Axis guides are the visual representation of position scales like those
created with \link[=scale_x_continuous]{scale_(x|y)_continuous()} and
\link[=scale_x_discrete]{scale_(x|y)_discrete()}.
}
