% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide-axis-theta.R
\name{guide_axis_theta}
\alias{guide_axis_theta}
\title{Angle axis guide}
\usage{
guide_axis_theta(
  title = waiver(),
  angle = waiver(),
  cap = "none",
  order = 0,
  major.length = 1,
  minor.length = 0.75,
  minor.ticks = element_blank(),
  position = waiver()
)
}
\arguments{
\item{title}{A character string or expression indicating a title of guide.
If \code{NULL}, the title is not shown. By default
(\code{\link[=waiver]{waiver()}}), the name of the scale object or the name
specified in \code{\link[=labs]{labs()}} is used for the title.}

\item{angle}{Compared to setting the angle in \code{\link[=theme]{theme()}} / \code{\link[=element_text]{element_text()}},
this also uses some heuristics to automatically pick the \code{hjust} and \code{vjust} that
you probably want. Can be one of the following:
\itemize{
\item \code{NULL} to take the angles and \code{hjust}/\code{vjust} directly from the theme.
\item \code{waiver()} to allow reasonable defaults in special cases.
\item A number representing the text angle in degrees.
}}

\item{cap}{A \code{character} to cut the axis line back to the last breaks. Can
be \code{"none"} (default) to draw the axis line along the whole panel, or
\code{"upper"} and \code{"lower"} to draw the axis to the upper or lower break, or
\code{"both"} to only draw the line in between the most extreme breaks. \code{TRUE}
and \code{FALSE} are shorthand for \code{"both"} and \code{"none"} respectively.}

\item{order}{A positive \code{integer} of length 1 that specifies the order of
this guide among multiple guides. This controls in which order guides are
merged if there are multiple guides for the same position. If 0 (default),
the order is determined by a secret algorithm.}

\item{major.length, minor.length}{A \code{numeric} of length 1 giving the length of
major and minor ticks relative to the theme's setting.}

\item{minor.ticks}{A theme element inheriting from \code{element_line} or
\code{element_blank} for drawing minor ticks. Alternatively, a \code{logical} of
length 1 as shorthand for \code{element_line()} (\code{TRUE}) or \code{element_blank}
(\code{FALSE}). \code{minor.ticks = element_line(...)} can be used to style the
minor ticks.}

\item{position}{Where this guide should be drawn: one of top, bottom,
left, or right.}
}
\description{
This is a specialised guide used in \code{coord_polar2()} to represent the theta
position scale.
}
\note{
The axis labels in this guide are insensitive to \code{hjust} and \code{vjust}
settings. The distance from the tick marks to the labels is determined by
the largest \code{margin} size set in the theme.
}
\examples{
# A basic polar plot
p <- ggplot(mtcars, aes(disp, mpg)) +
  geom_point() +
  coord_polar2()

# The `angle` argument can be used to set relative angles
p + guides(theta = guide_axis_theta(angle = 0))

# Minor ticks can be activated by providing a line element
p + guides(theta = guide_axis_theta(minor.ticks = element_line()))
}
