% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide-data.r
\name{guide_data}
\alias{guide_data}
\title{Extract tick information from guides}
\usage{
guide_data(plot = last_plot(), aesthetic, i = 1L, j = 1L)
}
\arguments{
\item{plot}{A \code{ggplot} object, or \code{ggplot_build} object.}

\item{aesthetic}{A scalar character that describes an aesthetic for which
to extract guide information. For example \code{"colour"}, \code{"size"}, \code{"x"} or
\code{"y.sec"}.}

\item{i, j}{An integer giving a row (i) or column (j) number of a facet for
which to return position guide information.}
}
\value{
A \code{data.frame} containing information extracted from the guide key,
or \code{NULL} when no such information could be found.
}
\description{
\code{guide_data()} builds a plot and extracts information from guide keys. This
information typically contains positions, values and/or labels, depending
on which aesthetic is queried or guide is used.
}
\details{
When used with plots containing \code{coord_sf()}, an aesthetic of \code{"x"}
is used for the bottom axis, \code{"x.sec"} for the top axis, \code{"y"} for the
left axis and \code{"y.sec"} for the right axis. This is also the default for
other coords, but this can be changed for other coords with the \code{guides()}
function.
}
\examples{
# A typical plot
p <- ggplot(mtcars) +
  aes(mpg, disp, colour = drat, size = drat) +
  geom_point() +
  facet_wrap(vars(cyl), scales = "free_x")

# Getting guide data for position aesthetic in particular panel
guide_data(p, "x", i = 1, j = 2)

# Getting guide data for a legend
guide_data(plot, "size")

# If guides are merged, `guide_data()` reports merged guide data
guide_data(plot + guides(colour = "legend"), "size")
}
