% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide-none.R
\name{guide_none}
\alias{guide_none}
\title{Empty guide}
\usage{
guide_none(title = waiver(), position = waiver())
}
\arguments{
\item{title}{A character string or expression indicating a title of guide.
If \code{NULL}, the title is not shown. By default
(\code{\link[=waiver]{waiver()}}), the name of the scale object or the name
specified in \code{\link[=labs]{labs()}} is used for the title.}

\item{position}{Where this guide should be drawn: one of top, bottom,
left, or right.}
}
\description{
This guide draws nothing.
}
