\name{html_index_create}
\alias{html_index_create}
\title{Create physical file for html documentation index}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{

}
\usage{html_index_create(path="web/")}
\arguments{
\item{path}{path to create file in}
}

\details{}

\examples{}
\keyword{internal}
