\name{html_linked_list}
\alias{html_linked_list}
\title{Convenience function for generating lists of objects with their icons.}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{

}
\usage{html_linked_list(objects)}
\arguments{
\item{objects}{list of objects}
}

\details{}

\examples{}
\keyword{internal}
