\name{html_template_create}
\alias{html_template_create}
\title{Create html file for templated files}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
See \code{\link{html_template}} for more details
}
\usage{html_template_create(name, path="web/")}
\arguments{
\item{name}{name of template}
\item{path}{path to create file in}
}

\details{}

\examples{}
\keyword{internal}
