% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities-matrix.r
\name{interleave}
\alias{interleave}
\title{Interleave (or zip) multiple vectors into a single vector.}
\usage{
interleave(...)
}
\arguments{
\item{...}{vectors to interleave}
}
\description{
Interleave (or zip) multiple vectors into a single vector.
}
\keyword{internal}

