% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggproto.R, R/aes.R, R/geom-.R, R/coord-.R,
%   R/facet-.R, R/stat-.R, R/guide-.R, R/layer.R, R/guides-.R, R/margins.R,
%   R/plot.R, R/position-.R, R/scale-.R, R/theme.R
\name{is_ggproto}
\alias{is_ggproto}
\alias{is.ggproto}
\alias{is_mapping}
\alias{is_geom}
\alias{is_coord}
\alias{is.Coord}
\alias{is_facet}
\alias{is.facet}
\alias{is_stat}
\alias{is_guide}
\alias{is_layer}
\alias{is_guides}
\alias{is_margin}
\alias{is_tests}
\alias{is_ggplot}
\alias{is.ggplot}
\alias{is_position}
\alias{is_scale}
\alias{is_theme}
\alias{is.theme}
\title{Reports wether \code{x} is a type of object}
\usage{
is_ggproto(x)

is.ggproto(x) # Deprecated

is_mapping(x)

is_geom(x)

is_coord(x)

is.Coord(x) # Deprecated

is_facet(x)

is.facet(x) # Deprecated

is_stat(x)

is_guide(x)

is_layer(x)

is_guides(x)

is_margin(x)

is_ggplot(x)

is.ggplot(x) # Deprecated

is_position(x)

is_scale(x)

is_theme(x)

is.theme(x) # Deprecated
}
\arguments{
\item{x}{An object to test}
}
\description{
Reports wether \code{x} is a type of object
}
\keyword{internal}
