% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggproto.R, R/aes.R, R/geom-.R, R/coord-.R,
%   R/facet-.R, R/stat-.R, R/theme-elements.R, R/guide-.R, R/layer.R,
%   R/guides-.R, R/margins.R, R/plot.R, R/position-.R, R/scale-.R, R/theme.R
\name{is.ggproto}
\alias{is.ggproto}
\alias{is.mapping}
\alias{is.geom}
\alias{is.coord}
\alias{is.Coord}
\alias{is.facet}
\alias{is.stat}
\alias{is.element}
\alias{is.guide}
\alias{is.layer}
\alias{is.guides}
\alias{is.margin}
\alias{is_tests}
\alias{is.ggplot}
\alias{is.position}
\alias{is.scale}
\alias{is.theme}
\title{Reports whether x is a type of object}
\usage{
is.ggproto(x)

is.mapping(x)

is.geom(x)

is.coord(x)

is.Coord(x) # Deprecated

is.facet(x)

is.stat(x)

is.element(x)

is.guide(x)

is.layer(x)

is.guides(x)

is.margin(x)

is.ggplot(x)

is.position(x)

is.scale(x)

is.theme(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Reports whether x is a type of object
}
\keyword{internal}
