% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/facet-labels.r
\name{label_both}
\alias{label_both}
\title{Label facets with value and variable.}
\usage{
label_both(variable, value)
}
\arguments{
\item{variable}{variable name passed in by facetter}

\item{value}{variable value passed in by facetter}
}
\description{
Label facets with value and variable.
}
\examples{
p <- qplot(wt, mpg, data = mtcars)
p + facet_grid(. ~ cyl)
p + facet_grid(. ~ cyl, labeller = label_both)
}
\seealso{
Other facet labellers: \code{\link{label_bquote}};
  \code{\link{label_parsed}}; \code{\link{label_value}}
}

