% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/facet-labels.r
\name{label_wrap_gen}
\alias{label_wrap_gen}
\title{Label facets with a word wrapped label.}
\usage{
label_wrap_gen(width = 25)
}
\arguments{
\item{width}{integer, target column width for output.}
}
\description{
Uses \code{\link[base]{strwrap}} for line wrapping.
}
\seealso{
, \code{\link{labeller}}
}

