% Generated by roxygen2 (4.0.0): do not edit by hand
\name{labeller}
\alias{labeller}
\title{Generic labeller function for facets}
\usage{
labeller(keep.as.numeric = FALSE, ...)
}
\arguments{
  \item{keep.as.numbers}{logical, default TRUE. When FALSE,
  converts numeric values supplied as margins to the facet
  to characters.}
}
\value{
Function to supply to \code{\link{facet_grid}} for the argument \code{labeller}.
}
\description{
One-step function for providing methods or named character vectors
as labels in facets.
}
\examples{
numbers <- c(`4`='four', `6`='six', `8`='eight')
vs <- c(`0`='No vs', `1`='vs')
p <- ggplot(mtcars, aes(mpg, wt)) + geom_point()
p + facet_grid(vs~cyl, labeller=labeller(cyl=numbers, vs=vs))
}

