% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/facet-labels.r
\name{labeller}
\alias{labeller}
\title{Generic labeller function for facets}
\usage{
labeller(rows = NULL, cols = NULL, keep.as.numeric = NULL,
  as_character = FALSE, multi_line = TRUE, default = label_value, ...)
}
\arguments{
\item{rows}{Labeller for the rows. It is passed to
\code{\link{as_labeller}()} and can be a lookup table, a function
taking and returning character vectors, or simply a labeller
function.}

\item{cols}{Labeller for the columns. Also processed with
\code{\link{as_labeller}()}.}

\item{keep.as.numeric}{Deprecated, use \code{.as_character} instead.}

\item{as_character}{Logical, default FALSE. When TRUE, converts
numeric labels to characters.}

\item{multi_line}{Whether to display the labels of multiple factors
on separate lines. This is passed to the labeller function.}

\item{default}{Default labeller for the rows or the columns when
they are not specified. Also used with lookup tables or
non-labeller functions.}

\item{...}{Deprecated. Use \code{rows} and \code{cols} instead.}
}
\value{
A labeller function to supply to \code{\link{facet_grid}}
  for the argument \code{labeller}.
}
\description{
This function makes it easy to assign different labellers to
different factors. The labeller can be a function or it can be a
named character vectors that will serve as a lookup table.
}
\details{
In case of functions, if the labeller has class \code{labeller}, it
is directly applied on the data frame of labels. Otherwise, it is
applied to the columns of the data frame of labels. The data frame
is then processed with the function specified in the
\code{.default} argument. This is intended to be used with
functions taking a character vector such as
\code{\link[Hmisc]{capitalize}}.
}
\examples{
\donttest{
p1 <- ggplot(mtcars, aes(x = mpg, y = wt)) + geom_point()
p1 + facet_grid(vs + am ~ gear,
  labeller = labeller(rows = label_both, cols = label_value))


capitalize <- function(string) {
  substr(string, 1, 1) <- toupper(substr(string, 1, 1))
  string
}
conservation_status <- c(
  cd = "Conservation Dependent",
  en = "Endangered",
  lc = "Least concern",
  nt = "Near Threatened",
  vu = "Vulnerable",
  domesticated = "Domesticated"
)
## Source: http://en.wikipedia.org/wiki/Wikipedia:Conservation_status

p2 <- ggplot(msleep, aes(x = sleep_total, y = awake)) + geom_point()
p2 + facet_grid(vore ~ conservation, labeller = labeller(rows = capitalize))

p2 + facet_grid(vore ~ conservation, labeller = labeller(
  rows = capitalize,
  cols = conservation_status
))

# In the following example, we rename the levels to the long form,
# then apply a wrap labeller to the columns to prevent cropped text
msleep$conservation2 <- plyr::revalue(msleep$conservation,
  conservation_status)

p2 \%+\% msleep +
  facet_grid(vore ~ conservation2)
p2 \%+\% msleep +
  facet_grid(vore ~ conservation2,
    labeller = labeller(cols = label_wrap_gen(10))
  )
}
}
\seealso{
\link{labellers}
}

