% Generated by roxygen2 (4.0.0): do not edit by hand
\name{labeller}
\alias{labeller}
\title{Generic labeller function for facets}
\usage{
labeller(..., keep.as.numeric = FALSE)
}
\arguments{
  \item{keep.as.numbers}{logical, default TRUE. When FALSE,
  converts numeric values supplied as margins to the facet
  to characters.}
}
\value{
Function to supply to \code{\link{facet_grid}} for the argument \code{labeller}.
}
\description{
One-step function for providing methods or named character vectors
as labels in facets.
}
\details{
Provided methods are parsed two variables,
\code{variable} and \code{values} (in that order),
where \code{variable} is the name of the facetting variable and
\code{values} are the different values this variable takes.
If the provided method takes no arguments, an error is raised.
If the provided method requires one argument, \code{values} are mapped to this.
If the provided method requires more than two, an error is raised.
If the provided method has zero or one required argument,
but multiple optional argument, you are on our own.

If the result is NA, the function defaults to return the requsted value.

\code{NA}'s are replaced with strings of 'NA'.
}
\examples{
data(mpg)
ggplot(mpg, aes(cty, hwy)) + geom_point() + facet_grid(cyl ~ class, labeller=label_both)
ggplot(mpg, aes(cty, hwy)) + geom_point() + facet_grid(cyl ~ class, labeller=labeller(cyl=label_both))


library(Hmisc) ## for the capitalize function
conservation_status <- c('cd'='Conservation Dependent',
                         'en'='Endangered',
                         'lc'='Least concern',
                         'nt'='Near Threatened',
                         'vu'='Vulnerable',
                         'domesticated'='Domesticated')
Source: http://en.wikipedia.org/wiki/Wikipedia:Conservation_status

ggplot(msleep, aes(x=sleep_total, y=awake)) + geom_point() + facet_grid(vore ~ conservation, labeller=labeller(vore=capitalize))
ggplot(msleep, aes(x=sleep_total, y=awake)) + geom_point() + facet_grid(vore ~ conservation, labeller=labeller(vore=capitalize, conservation=conservation_status ))

# We could of course have renamed the levels; then we can apply another nifty function:
library(plyr)
msleep$conservation2 <- revalue(msleep$conservation, conservation_status)
ggplot(msleep, aes(x=sleep_total, y=awake)) + geom_point() + facet_grid(vore ~ conservation2, labeller=labeller(vore=capitalize))
ggplot(msleep, aes(x=sleep_total, y=awake)) + geom_point() + facet_grid(vore ~ conservation2, labeller=labeller(conservation2=label_wrap_gen(10) ))
}

