% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-chain.R
\name{link_stat}
\alias{link_stat}
\title{Parameterise a statistic computation}
\usage{
link_stat(stat, ..., mapping = aes())
}
\arguments{
\item{stat}{The statistical transformation to use on the data. The \code{stat}
argument accepts the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}.
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, for \code{stat_count()}, give
the string \code{"count"}.
}}

\item{...}{Other arguments passed to the stat as a parameter.}

\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} to be
evaluated only after the stat has been computed.}
}
\value{
A list bundling the stat, parameters and mapping.
}
\description{
This is a helper function for \code{\link[=stat_chain]{stat_chain()}} to pass parameters and declare
mappings.
}
\examples{
# See `?stat_chain`
}
\seealso{
\code{\link[=stat_chain]{stat_chain()}}
}
