% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{margin}
\alias{margin}
\title{Define margins.}
\usage{
margin(t = 0, r = 0, b = 0, l = 0, unit = "pt")
}
\arguments{
\item{t, r, b, l}{Dimensions of each margin. (To remember order, think trouble).}

\item{unit}{Default units of dimensions. Defaults to "pt" so it
can be most easily scaled with the text.}
}
\description{
This is a convenient function that creates a grid unit object of the
correct length to use for setting margins.
}
\examples{
margin(4)
margin(4, 2)
margin(4, 3, 2, 1)
}

