% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-.r
\name{max_height}
\alias{max_height}
\alias{max_width}
\title{Get the maximal width/length of a list of grobs}
\usage{
max_height(grobs, value_only = FALSE)

max_width(grobs, value_only = FALSE)
}
\arguments{
\item{grobs}{A list of grobs}

\item{value_only}{Should the return be a simple scalar numeric giving the
length in cm, or a unit.}
}
\value{
The largest value. measured in cm as a unit object
}
\description{
Get the maximal width/length of a list of grobs
}
\keyword{internal}
