% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-.R
\name{max_height}
\alias{max_height}
\alias{max_width}
\title{Get the maximal width/length of a list of grobs}
\usage{
max_height(grobs, value_only = FALSE)

max_width(grobs, value_only = FALSE)
}
\arguments{
\item{grobs}{A list of grobs}

\item{value_only}{Should the return value be a simple numeric vector giving
the maximum in cm}
}
\value{
The largest value. measured in cm as a unit object or a numeric
vector depending on \code{value_only}
}
\description{
Get the maximal width/length of a list of grobs
}
\keyword{internal}
