% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{midwest}
\alias{midwest}
\title{Midwest demographics.}
\format{A data frame with 437 rows and 28 variables
\itemize{
 \item PID
 \item county
 \item state
 \item area
 \item poptotal.  Total population
 \item popdensity. Population density
 \item popwhite.  Number of whites.
 \item popblack.  Number of blacks.
 \item popamerindian.  Number of American Indians.
 \item popasian.  Number of Asians.
 \item popother.  Number of other races.
 \item percwhite.  Percent white.
 \item percblack.  Percent black.
 \item percamerindan.  Percent American Indian.
 \item percasian. Percent Asian.
 \item percother. Percent other races.
 \item popadults.  Number of adults.
 \item perchsd.
 \item percollege.  Percent college educated.
 \item percprof.  Percent profession.
 \item poppovertyknown.
 \item percpovertyknown
 \item percbelowpoverty
 \item percchildbelowpovert
 \item percadultpoverty
 \item percelderlypoverty
 \item inmetro.  In a metro area.
 \item category'
}}
\usage{
midwest
}
\description{
Demographic information of midwest counties
}
\keyword{datasets}

