% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msleep}
\alias{msleep}
\title{An updated and expanded version of the mammals sleep dataset.}
\format{A data frame with 83 rows and 11 variables
\itemize{
  \item name. common name
  \item genus.
  \item vore. carnivore, omnivore or herbivore?
  \item order.
  \item conservation. the conservation status of the animal
  \item sleep\_total. total amount of sleep, in hours
  \item sleep\_rem. rem sleep, in hours
  \item sleep\_cycle. length of sleep cycle, in hours
  \item awake. amount of time spent awake, in hours
  \item brainwt. brain weight in kilograms
  \item bodywt. body weight in kilograms
}}
\usage{
msleep
}
\description{
This is an updated and expanded version of the mammals sleep dataset.
Updated sleep times and weights were taken from V. M. Savage and G. B.
West. A quantitative, theoretical framework for understanding mammalian
sleep. Proceedings of the National Academy of Sciences, 104 (3):1051-1056,
2007.
}
\details{
Additional variables order, conservation status and vore were added from
wikipedia.
}
\keyword{datasets}

