\name{nice_ramp}
\alias{nice_ramp}
\title{Wrapper for colorRamp that deals with missing values and alpha}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{

}
\usage{nice_ramp(ramp, x, alpha)}
\arguments{
\item{ramp}{}
\item{x}{}
\item{alpha}{}
}

\details{}

\examples{}
\keyword{internal}
