\name{nice_ramp}
\alias{nice_ramp}
\title{Wrapper for colorRamp that deals with missing values and alpha}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{

}
\usage{nice_ramp(ramp, x, alpha)}
\arguments{
\item{ramp}{colour}
\item{x}{alpha level [0,1]}
\item{alpha}{}
}

\details{@keyword internal
alpha
Give a colour an alpha level}

\examples{}
\keyword{internal}
\keyword{internal}
