% Generated by roxygen2 (4.0.0): do not edit by hand
\name{opts}
\alias{opts}
\title{Build a theme (or partial theme) from theme elements}
\usage{
opts(...)
}
\arguments{
  \item{...}{Arguments to be passed on to the \code{theme}
  function.}
}
\description{
\code{opts} is deprecated. See the \code{\link{theme}} function.
}

