% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-partial.R
\name{partial-scales}
\alias{partial-scales}
\alias{scale_x}
\alias{scale_y}
\title{Partial scales}
\usage{
scale_x(name = waiver(), ...)

scale_y(name = waiver(), ...)
}
\arguments{
\item{name}{The name of the scale that can be used as guide title.}

\item{...}{Other parameters passed to scales. Many of these depend on the
eventual scale type, but \code{limits}, \code{breaks}, \code{labels} and \code{guide} are
commonplace.}
}
\value{
A \verb{<ScalePartial>} object that can be added to a plot.
}
\description{
Partial scales can be used to adjust scale parameters without
explicitly committing to a particular type of scale.
}
\details{
Parameters set via partial scales may be overridden by adding 'full' scales
to a plot. If no full scales are added to a plot, the parameters from
the partial scales are passed to the default full scale. Therefore, some
care must be taken to not accidentally pass a continuous scale parameter,
like \code{trans}, to discrete scales, or pass a \code{palette} to a position scale.
}
\examples{
# TODO
}
