% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-boxdodge.r
\name{position_boxdodge}
\alias{position_boxdodge}
\title{Position dodge for box plots}
\usage{
position_boxdodge(width = NULL, preserve = c("single", "total"),
  padding = 0.1)
}
\arguments{
\item{width}{Dodging width, when different to the width of the individual
elements. This is useful when you want to align narrow geoms with wider
geoms. See the examples.}

\item{preserve}{Should dodging preserve the total width of all elements
at a position, or the width of a single element?}

\item{padding}{Padding between boxes at the same position. Boxes are shrunk
by this proportion to make room for space between them.}
}
\description{
Dodging preserves the vertical position of an geom while adjusting the
horizontal position. \code{position_boxdodge} is a special case of
\code{position_dodge} for arranging box plots, which can have variable widths.
}
\examples{
ggplot(data = iris, aes(Species, Sepal.Length)) +
geom_boxplot(aes(colour = Sepal.Width < 3.2))

ggplot(data = iris, aes(Species, Sepal.Length)) +
geom_boxplot(aes(colour = Sepal.Width < 3.2), varwidth = TRUE)
}
\seealso{
Other position adjustments: \code{\link{position_dodge}},
  \code{\link{position_identity}},
  \code{\link{position_jitterdodge}},
  \code{\link{position_jitter}},
  \code{\link{position_nudge}},
  \code{\link{position_stack}}
}
