\name{position_identity}
\alias{position_identity}
\alias{PositionIdentity}
\title{position\_identity}
\description{Don't adjust position}
\details{
This page describes position\_identity, see \code{\link{layer}} and \code{\link{qplot}} for how to create a complete plot from individual components.
}
\usage{position_identity(width = NULL, height = NULL, ...)}
\arguments{
 \item{width}{NULL}
 \item{height}{NULL}
 \item{...}{ignored }
}
\seealso{\itemize{
  \item \url{http://had.co.nz/ggplot2/position_identity.html}
}}
\value{A \code{\link{layer}}}
\examples{\dontrun{
  # Coming soon
}}
\author{Hadley Wickham, \url{http://had.co.nz/}}
\keyword{hplot}
