\name{PositionIdentity}
\alias{position_identity}
\alias{PositionIdentity}
\title{position_identity}
\description{Don't adjust position}
\details{
This page describes \code{\link{position_identity}}, see \code{\link{layer}} and \code{\link{qplot}} for how to create a complete plot from individual components.
}
\usage{position_identity(...)}
\arguments{
 \item{...}{ignored }
}
\seealso{\itemize{
  \item \url{http://had.co.nz/ggplot/position_identity.html}
}}
\value{A \code{\link{layer}}}
\examples{\dontrun{
  # Coming soon
}}
\author{Hadley Wickham, \url{http://had.co.nz/}}
\keyword{hplot}
