\name{position_identity}
\alias{position_identity}
\title{Don't adjust position}
\usage{
  position_identity(width = NULL, height = NULL, ...)
}
\arguments{
  \item{width}{Horizontal width from the center point to
  dodge.}

  \item{height}{Vertical height from the center point to
  dodge.}

  \item{...}{Other arguments to be passed to the function.}
}
\description{
  Don't adjust position
}

