% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/position-identity.r
\name{position_identity}
\alias{position_identity}
\title{Don't adjust position}
\usage{
position_identity(width = NULL, height = NULL)
}
\arguments{
\item{width}{Manually specify width (does not affect all position
adjustments)}

\item{height}{Manually specify height (does not affect all position
adjustments)}
}
\description{
Don't adjust position
}
\seealso{
Other position adjustments: \code{\link{position_dodge}};
  \code{\link{position_fill}};
  \code{\link{position_jitterdodge}};
  \code{\link{position_jitter}};
  \code{\link{position_stack}}
}

