% Generated by roxygen2 (4.0.0): do not edit by hand
\name{position_jitterdodge}
\alias{position_jd}
\alias{position_jitterdodge}
\title{Jitter-dodge points to align them with a boxplot including fill aesthetic}
\usage{
position_jitterdodge(width = NULL, height = NULL)

position_jd(width = NULL, height = NULL)
}
\arguments{
  \item{width}{degree of jitter in x direction. Defaults to
  40\% of the resolution of the data.}

  \item{height}{degree of jitter in y direction. Defaults
  to 40\% of the resolution of the data}
}
\description{
Jitter-dodge points to align them with a boxplot including fill aesthetic
}
\examples{
dsub <- diamonds[ sample(1:nrow(diamonds), 1000), ]
ggplot(dsub, aes(x=cut, y=carat, fill=clarity)) +
  geom_boxplot(outlier.size=0) +
  geom_point( pch=21, position=position_jitterdodge() )
}
\seealso{
Other position adjustments: \code{\link{position_dodge}};
  \code{\link{position_fill}};
  \code{\link{position_identity}};
  \code{\link{position_jitter}};
  \code{\link{position_stack}}
}

