% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-sina.r
\name{position_sina}
\alias{position_sina}
\title{Point density based jitter.}
\usage{
position_sina(binwidth = 0.2, scale = TRUE, neighbour_limit = 1,
  method = c("density", "neighbourhood"), adjust = 1)
}
\arguments{
\item{binwidth}{Binning factor. The range of the values in \code{aes(y)}
are binned in windows of length \code{(max(y) - min(y)) * y_fraction}.
Samples within the same bin belong to the same "neighbourhood".}

\item{scale}{logical. When set to \code{TRUE} x-coordinate widths across all
groups are scaled based on the densiest are in the plot. Default: \code{TRUE}}

\item{neighbour_limit}{if the samples within the same y-axis bin are more
than neighbour_limit, the samples's X coordinates will be adjusted.}

\item{method}{choose the method to spread the samples within the same
neighbourhood along the x-axis. Available methods: "density",
"neighbourhood" (can be abbreciated, e.g. "d"). See \code{Details}.}

\item{adjust}{adjusts the bandwidth of the density kernel when
\code{method == "density"} (see \code{\link[stats]{density}}) or the spread
of the samples within the same group along the x-axis.}
}
\description{
Point density based jitter.
}
\examples{
# See geom_sina for examples.
}
\seealso{
Other position adjustments: \code{\link{position_dodge}},
  \code{\link{position_fill}},
  \code{\link{position_identity}},
  \code{\link{position_jitterdodge}},
  \code{\link{position_jitter}},
  \code{\link{position_nudge}}
}

