\name{predictdf}
\alias{predictdf}
\alias{predictdf.default}
\alias{predictdf.glm}
\title{Prediction data frame}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Get predictions with standard errors into data frame
}
\usage{predictdf(model, xseq, se, level)}
\arguments{
\item{model}{}
\item{xseq}{}
\item{se}{}
\item{level}{}
}

\details{}

\examples{}

