\name{presidential}
\alias{presidential}
\title{Terms of 10 presidents from Eisenhower to Bush W.}
\usage{data(presidential)}

\description{
  Terms of 10 presidents from Eisenhower to Bush W.
}

\details{
  The names of each president, the start and end date of
  their term, and their party of 10 US presidents from
  Eisenhower to Bush W.
}
\docType{data}
\keyword{datasets}
