% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{data}
\name{presidential}
\alias{presidential}
\title{Terms of 10 presidents from Eisenhower to Bush W.}
\format{A data frame with 10 rows and 4 variables}
\usage{
data(presidential)
}
\description{
The names of each president, the start and end date of their term, and
their party of 10 US presidents from Eisenhower to Bush W.
}
\keyword{datasets}

