% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{presidential}
\alias{presidential}
\title{Terms of 12 presidents from Eisenhower to Trump}
\format{
A data frame with 12 rows and 4 variables:
\describe{
\item{name}{Last name of president}
\item{start}{Presidency start date}
\item{end}{Presidency end date}
\item{party}{Party of president}
}
}
\usage{
presidential
}
\description{
The names of each president, the start and end date of their term, and
their party of 12 US presidents from Eisenhower to Trump.  This data is
in the public domain.
}
\keyword{datasets}
