\name{print.ggplot}
\alias{print.ggplot}
\title{Draw plot on current graphics device.}
\usage{
  \method{print}{ggplot} (x, newpage = is.null(vp),
    vp = NULL, ...)
}
\arguments{
  \item{x}{plot to display}

  \item{newpage}{draw new (empty) page first?}

  \item{vp}{viewport to draw plot in}

  \item{...}{other arguments not used by this method}
}
\description{
  Draw plot on current graphics device.
}
\keyword{hplot}

