% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-render.r
\name{print.ggplot}
\alias{plot.ggplot}
\alias{print.ggplot}
\title{Draw plot on current graphics device.}
\usage{
\method{print}{ggplot}(x, newpage = is.null(vp), vp = NULL, ...)

\method{plot}{ggplot}(x, newpage = is.null(vp), vp = NULL, ...)
}
\arguments{
\item{x}{plot to display}

\item{newpage}{draw new (empty) page first?}

\item{vp}{viewport to draw plot in}

\item{...}{other arguments not used by this method}
}
\description{
Draw plot on current graphics device.
}
\keyword{hplot}

