% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/proto2.r
\name{proto2}
\alias{proto2}
\title{Create a new proto2 object}
\usage{
proto2(`_class` = NULL, `_inherit` = NULL, ...)
}
\arguments{
\item{_class}{Class name to assign to the object. This is stored as the class
attribute of the object. If \code{NULL} (the default), no class name will
be added to the object.}

\item{_inherit}{proto2 object to inherit from. If \code{NULL}, don't inherit
from any object.}

\item{...}{A list of members in the proto2 object.}
}
\description{
Create a new proto2 object
}

