% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/proto2.r
\name{proto2}
\alias{proto2}
\title{Create a new proto2 object}
\usage{
proto2(`_class`, `_inherit`, ...)
}
\arguments{
\item{...}{A list of members in the proto2 object.}

\item{`_class`}{Name of new class}

\item{`_inherit`}{proto2 object to inherit from. Use \code{NULL} to inherit
from the "base" class.}
}
\description{
Create a new proto2 object
}

