% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick-plot.r
\name{qplot}
\alias{qplot}
\alias{quickplot}
\title{Quick plot}
\usage{
qplot(
  x,
  y,
  ...,
  data,
  facets = NULL,
  margins = FALSE,
  geom = "auto",
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  log = "",
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  asp = NA
)

quickplot(
  x,
  y,
  ...,
  data,
  facets = NULL,
  margins = FALSE,
  geom = "auto",
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  log = "",
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  asp = NA
)
}
\arguments{
\item{x, y, ...}{Aesthetics passed into each layer}

\item{data}{Data frame to use (optional).  If not specified, will create
one, extracting vectors from the current environment.}

\item{facets}{faceting formula to use. Picks \code{\link[=facet_wrap]{facet_wrap()}} or
\code{\link[=facet_grid]{facet_grid()}} depending on whether the formula is one-
or two-sided}

\item{margins}{See \code{facet_grid()}: display marginal facets?}

\item{geom}{Character vector specifying geom(s) to draw. Defaults to
"point" if x and y are specified, and "histogram" if only x is specified.}

\item{xlim, ylim}{X and y axis limits}

\item{log}{Which variables to log transform ("x", "y", or "xy")}

\item{main, xlab, ylab}{Character vector (or expression) giving plot title,
x axis label, and y axis label respectively.}

\item{asp}{The y/x aspect ratio}
}
\description{
\code{qplot()} is now deprecated in order to encourage the users to
learn \code{\link[=ggplot]{ggplot()}} as it makes it easier to create complex graphics.
}
