\name{rdoc_auto_link}
\alias{rdoc_auto_link}
\title{rdoc auto link}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Automatically link functions used in rdoc
}
\usage{rdoc_auto_link(input, skip="")}
\arguments{
\item{input}{input rdoc string}
\item{skip}{functions to omit}
}




\keyword{internal}
