% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-elements.r
\name{register_theme_elements}
\alias{register_theme_elements}
\title{Register new theme elements}
\usage{
register_theme_elements(
  ...,
  complete = TRUE,
  element_tree = NULL,
  reset_all = FALSE,
  modify_current = TRUE
)
}
\arguments{
\item{...}{Element specifications}

\item{complete}{If \code{TRUE} (the default), elements are set to inherit from blank elements.}

\item{element_tree}{Addition of or modification to the element tree, which specifies the
inheritance relationship of the theme elements. The element tree must be provided as
a list of named element definitions created with \code{\link[=el_def]{el_def()}}.}

\item{reset_all}{If \code{TRUE}, resets all theme elements and the element tree to the ggplot2
defaults. This also resets the currently active theme if \code{modify_current} is set to
\code{TRUE} as well.}

\item{modify_current}{If \code{TRUE} (the default), the new theme elements are also added to
the currently active theme set by \code{theme_set()}.}
}
\description{
This function allows users to globally register new theme elements with ggplot2.
In general, for each new theme element an element definition and a corresponding entry
in the element tree should be provided. See \code{\link[=el_def]{el_def()}} for examples.
}
\details{
Extension packages should call this function from their \code{.onLoad()} function.
}
\keyword{internal}
