% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-elements.r
\name{rel}
\alias{rel}
\title{Relative sizing for theme elements}
\usage{
rel(x)
}
\arguments{
\item{x}{A number representing the relative size}
}
\description{
Relative sizing for theme elements
}
\examples{
df <- data.frame(x = 1:3, y = 1:3)
ggplot(df, aes(x, y)) +
  geom_point() +
  theme(axis.title.x = element_text(size = rel(2.5)))
}

