\name{rescale}
\alias{rescale}
\title{Rescale numeric vector}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Rescale numeric vector to have specified minimum and maximum.
}
\usage{rescale(x, to=c(0,1), from=range(x, na.rm=TRUE), clip = TRUE)}
\arguments{
\item{x}{}
\item{to}{}
\item{from}{}
\item{clip}{}
}

\details{}

\examples{}
\keyword{manip}
