\name{resolution}
\alias{resolution}
\title{Resolution}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Compute the "resolution" of a data vector, ie. what is the smallest non-zero
}
\usage{resolution(x, zero = TRUE)}
\arguments{
\item{x}{numeric vector}
\item{zero}{}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
