% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reuse-scale.R
\name{reuse_scale}
\alias{reuse_scale}
\title{Make a copy of a color scale for a different aesthetic.
Example use: \code{scale_outlinecolor_hue <- reuse_scale(scale_color_hue, "outlinecolor")}}
\usage{
reuse_scale(scale, aesthetics)
}
\arguments{
\item{scale}{A color scale to copy from}

\item{aesthetic}{The name of the new aesthetic or aesthetics to apply the color scale to}
}
\description{
Make a copy of a color scale for a different aesthetic.
Example use: \code{scale_outlinecolor_hue <- reuse_scale(scale_color_hue, "outlinecolor")}
}
