% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities-matrix.r
\name{rweave}
\alias{rweave}
\title{Row weave.}
\usage{
rweave(...)
}
\arguments{
\item{...}{matrices to weave together}
}
\description{
Weave together two (or more) matrices by row.
}
\details{
Matrices must have same dimensions.
}
\keyword{internal}

