\name{save_examples}
\alias{save_examples}
\title{Save examples}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Cache output from all examples in ggplot directory
}
\usage{save_examples(name = get_rev(), verbose = FALSE)}
\arguments{
\item{name}{}
\item{verbose}{}
}

\details{Produces:
* png for each example
* csv with hash, code and timing info}

\examples{}
\keyword{internal}
