\name{scale_alpha_discrete}
\alias{scale_alpha_discrete}
\title{Alpha scale for discrete variable.}
\usage{
  scale_alpha_discrete(..., range = c(0.1, 1))
}
\arguments{
  \item{range}{range of output alpha values.  Should lie
  between 0 and 1.}

  \item{...}{Other arguments passed on to
  \code{\link{discrete_scale}} to control name, limits,
  breaks, labels and so forth.}
}
\description{
  Alpha scale for discrete variable.
}

