\name{scale_alpha_discrete}
\alias{scale_alpha_discrete}
\title{Alpha scale for discrete variable.}
\usage{scale_alpha_discrete(..., range=c(1, 6))}

\description{
  Alpha scale for discrete variable.
}
\arguments{
  \item{...}{Other arguments passed on to \code{\link{discrete_scale}} 
to control name, limits, breaks, labels and so forth.}
}
