\name{scale_area}
\alias{scale_area}
\title{Scale area instead of radius (for size).}
\usage{
  scale_area(..., range = c(1, 6))
}
\arguments{
  \item{...}{Other arguments passed on to
  \code{\link{continuous_scale}} to control name, limits,
  breaks, labels and so forth.}

  \item{range}{Range of output sizes.  Should be greater
  than 0.}
}
\description{
  Scale area instead of radius (for size).
}

