% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scale-area.r
\name{scale_area}
\alias{scale_area}
\title{Scale area instead of radius (for size).}
\usage{
scale_area(..., range = c(1, 6))
}
\arguments{
\item{...}{Other arguments passed on to \code{\link{continuous_scale}}
to control name, limits, breaks, labels and so forth.}

\item{range}{Range of output sizes.  Should be greater than 0.}
}
\description{
\code{\link{scale_area}} is deprecated and will be removed in a future
version of ggplot2. Use \code{\link{scale_size_area}} instead. Note that the
default behavir of \code{\link{scale_size_area}} is slightly different: by
default, it makes the area proportional to the numeric value.
}

