% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-binned.R
\name{scale_binned}
\alias{scale_x_binned}
\alias{scale_y_binned}
\title{Positional scales for binning continuous data (x & y)}
\usage{
scale_x_binned(name = waiver(), n_bins = 10, breaks = waiver(),
  labels = waiver(), limits = NULL, expand = waiver(),
  oob = squish, na.value = NA_real_, right = TRUE,
  trans = "identity", position = "bottom")

scale_y_binned(name = waiver(), n_bins = 10, breaks = waiver(),
  labels = waiver(), limits = NULL, expand = waiver(),
  oob = squish, na.value = NA_real_, right = TRUE,
  trans = "identity", position = "left")
}
\description{
\code{scale_x_binned()} and \code{scale_y_binned()} are scales that discretize
continuous position data. You can use these scales to transform continuous
inputs before using it with a geom that requires discrete positions. An
example is using \code{scale_x_binned()} with \code{\link[=geom_bar]{geom_bar()}} to create a histogram.
}
\details{
#' @inheritParams binned_scale
}
\examples{
# Create a histogram by binning the x-axis
ggplot(mtcars) +
  geom_bar(aes(mpg)) +
  scale_x_binned()
}
\seealso{
Other position scales: \code{\link{scale_x_continuous}},
  \code{\link{scale_x_date}},
  \code{\link{scale_x_discrete}}
}
\concept{position scales}
