% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-colour.r
\name{scale_colour_continuous}
\alias{scale_colour_continuous}
\alias{scale_fill_continuous}
\title{Continuous colour scales}
\usage{
scale_colour_continuous(...,
  type = pkgconfig::get_config("ggplot2::continuous_colour", "gradient"))

scale_fill_continuous(...,
  type = pkgconfig::get_config("ggplot2::continuous_fill", "gradient"))
}
\arguments{
\item{...}{Additional parameters passed on to the scale type}

\item{type}{One of "gradient" (the default) or "viridis" indicating the
colour scale to use}
}
\description{
Continuous colour scales
}
\seealso{
\code{\link[=scale_colour_gradient]{scale_colour_gradient()}} and \code{\link[=scale_colour_viridis_c]{scale_colour_viridis_c()}}
}
