% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scale-size.r
\name{scale_size}
\alias{scale_radius}
\alias{scale_size}
\alias{scale_size_continuous}
\alias{scale_size_discrete}
\title{Scale size (area or radius).}
\usage{
scale_size_continuous(name = NULL, breaks = waiver(), labels = waiver(),
  limits = NULL, range = c(1, 6), trans = "identity", guide = "legend")

scale_radius(name = NULL, breaks = waiver(), labels = waiver(),
  limits = NULL, range = c(1, 6), trans = "identity", guide = "legend")

scale_size(name = NULL, breaks = waiver(), labels = waiver(),
  limits = NULL, range = c(1, 6), trans = "identity", guide = "legend")

scale_size_discrete(..., range = c(2, 6))
}
\arguments{
\item{name}{The name of the scale. Used as axis or legend title. If
\code{NULL}, the default, the name of the scale is taken from the first
mapping used for that aesthetic.}

\item{breaks}{One of: \itemize{
  \item \code{NULL} for no breaks
  \item \code{waiver()} for the default breaks computed by the
    transformation object
  \item A numeric vector of positions
  \item A function that takes the limits as input and returns breaks
    as output
}}

\item{labels}{One of: \itemize{
  \item \code{NULL} for no labels
  \item \code{waiver()} for the default labels computed by the
    transformation object
  \item A character vector giving labels (must be same length as \code{breaks})
  \item A function that takes the breaks as input and returns labels
    as output
}}

\item{limits}{A numeric vector of length two providing limits of the scale.
Use \code{NA} to refer to the existing minimum or maximum.}

\item{range}{a numeric vector of length 2 that specifies the minimum and
maximum size of the plotting symbol after transformation.}

\item{trans}{Either the name of a transformation object, or the
  object itself. Built-in transformations include "asn", "atanh",
  "boxcox", "exp", "identity", "log", "log10", "log1p", "log2",
  "logit", "probability", "probit", "reciprocal", "reverse" and "sqrt".

  A transformation object bundles together a transform, it's inverse,
  and methods for generating breaks and labels. Transformation objects
  are defined in the scales package, and are called \code{name_trans}, e.g.
  \code{\link[scales]{boxcox_trans}}. You can create your own
  transformation with \code{\link[scales]{trans_new}}.}

\item{guide}{Name of guide object, or object itself.}

\item{...}{Additional arguments passed on to \code{\link{discrete_scale}}.}
}
\description{
\code{scale_size} scales area, \code{scale_radius} scales radius. The size
aesthetic is most commonly used for points and text, and humans percieve
the area of points (not their radius), so this provides for optimal
perception.
}
\examples{
\donttest{
p <- ggplot(mtcars, aes(mpg, cyl, size = cyl)) +
   geom_point()
p
p + scale_size("cylinders")
p + scale_size("number\\nof\\ncylinders")

p + scale_size(range = c(0, 10))
p + scale_size(range = c(1, 2))

# If you want zero value to have zero size, use scale_size_area:
p + scale_size_area()

# You can map size to a factor, but it's not a terribly good
# idea, unless your factor is ordered, as in this example
ggplot(mtcars, aes(mpg, cyl, size = factor(cyl))) +
  geom_point()

# To control the size mapping for discrete variable, use
# scale_size_manual:
ggplot(mtcars, aes(mpg, cyl, size = factor(cyl))) +
  geom_point() +
  scale_size_manual(values = c(2, 4, 6))
}
}
\seealso{
\code{\link{scale_size_area}} if you want 0 values to be mapped
  to points with size 0.
}

