% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-viridis.r, R/zxx.r
\name{scale_colour_viridis_d}
\alias{scale_colour_viridis_d}
\alias{scale_fill_viridis_d}
\alias{scale_colour_viridis_c}
\alias{scale_fill_viridis_c}
\alias{scale_colour_ordinal}
\alias{scale_fill_ordinal}
\alias{scale_color_viridis_d}
\alias{scale_color_viridis_c}
\title{Viridis colour scales from viridisLite}
\usage{
scale_colour_viridis_d(..., alpha = 1, begin = 0, end = 1,
  direction = 1, option = "D")

scale_fill_viridis_d(..., alpha = 1, begin = 0, end = 1, direction = 1,
  option = "D")

scale_colour_viridis_c(..., alpha = 1, begin = 0, end = 1,
  direction = 1, option = "D", values = NULL, space = "Lab",
  na.value = "grey50", guide = "colourbar")

scale_fill_viridis_c(..., alpha = 1, begin = 0, end = 1, direction = 1,
  option = "D", values = NULL, space = "Lab", na.value = "grey50",
  guide = "colourbar")
}
\arguments{
\item{...}{Other arguments passed on to \code{\link[=discrete_scale]{discrete_scale()}} or
\code{\link[=continuous_scale]{continuous_scale()}} to control name, limits, breaks, labels and so forth.}

\item{alpha}{The alpha transparency, a number in [0,1], see argument alpha in
\code{\link[grDevices]{hsv}}.}

\item{begin}{The (corrected) hue in [0,1] at which the viridis colormap begins.}

\item{end}{The (corrected) hue in [0,1] at which the viridis colormap ends.}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors
are ordered from darkest to lightest. If -1, the order of colors is reversed.}

\item{option}{A character string indicating the colormap option to use. Four
options are available: "magma" (or "A"), "inferno" (or "B"), "plasma" (or "C"),
and "viridis" (or "D", the default option).}

\item{values}{if colours should not be evenly positioned along the gradient
this vector gives the position (between 0 and 1) for each colour in the
\code{colours} vector. See \code{\link{rescale}} for a convience function
to map an arbitrary range to between 0 and 1.}

\item{space}{colour space in which to calculate gradient. Must be "Lab" -
other values are deprecated.}

\item{na.value}{Missing values will be replaced with this value.}

\item{guide}{A function used to create a guide or its name. See
\code{\link[=guides]{guides()}} for more info.}
}
\description{
The \code{viridis} scales provide color maps that are perceptually uniform in both
color and black-and-white. They are also designed to be perceived by viewers
with common forms of color blindness. See also
\url{https://bids.github.io/colormap/}.
}
\examples{
dsamp <- diamonds[sample(nrow(diamonds), 1000), ]
(d <- ggplot(dsamp, aes(carat, price)) +
  geom_point(aes(colour = clarity)))
d + scale_colour_viridis_d()

# Change scale label
d + scale_colour_viridis_d("Diamond\\nclarity")

# Select palette to use, see ?scales::viridis_pal for more details
d + scale_colour_viridis_d(option = "plasma")
d + scale_colour_viridis_d(option = "inferno")

\donttest{
# scale_fill_viridis_d works just the same as
# scale_colour_viridis_d but for fill colours
p <- ggplot(diamonds, aes(x = price, fill = cut)) +
  geom_histogram(position = "dodge", binwidth = 1000)
p + scale_fill_viridis_d()
# the order of colour can be reversed
p + scale_fill_viridis_d(direction = -1)
}

# Use viridis_c with continous data
v <- ggplot(faithfuld) +
  geom_tile(aes(waiting, eruptions, fill = density))
v
v + scale_fill_viridis_c()
v + scale_fill_viridis_c(option = "plasma")
}
\seealso{
Other colour scales: \code{\link{scale_alpha}},
  \code{\link{scale_colour_brewer}},
  \code{\link{scale_colour_gradient}},
  \code{\link{scale_colour_grey}},
  \code{\link{scale_colour_hue}}
}
