\name{setup_viewports}
\alias{setup_viewports}
\title{Setup viewports}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Setup matrix of viewports for a layout with given parameters
}
\usage{setup_viewports(type, data, offset = c(0,0), clip = "on")}
\arguments{
\item{type}{viewport type}
\item{data}{number of rows}
\item{offset}{number of columns}
\item{clip}{optional data to compute rows and columns from}
\item{}{offset from top and left}
\item{}{list containing x and y ranges}
}

\details{}

\examples{}
\keyword{hplot}
\keyword{internal}
