\name{stat_binhex}
\alias{stat_binhex}
\alias{StatBinhex}
\title{stat\_binhex}
\description{}
\details{
This page describes stat\_binhex, see \code{\link{layer}} and \code{\link{qplot}} for how to create a complete plot from individual components.
}
\section{Aesthetics}{
The following aesthetics can be used with stat\_binhex.  Aesthetics are mapped to variables in the data with the \code{\link{aes}} function: \code{stat\_binhex(\code{\link{aes}}(x = var))}
\itemize{
  \item \code{x}: x position (\strong{required}) 
  \item \code{y}: y position (\strong{required}) 
  \item \code{fill}: internal colour 
}
}
\usage{stat_binhex(mapping=NULL, data=NULL, geom="hex", position="identity", bins=30, na.rm=FALSE, ...)}
\arguments{
 \item{mapping}{mapping between variables and aesthetics generated by aes}
 \item{data}{dataset used in this layer, if not specified uses plot dataset}
 \item{geom}{geometric used by this layer}
 \item{position}{position adjustment used by this layer}
 \item{bins}{NULL}
 \item{na.rm}{NULL}
 \item{...}{ignored }
}
\seealso{\itemize{
  \item \url{http://had.co.nz/ggplot2/stat_binhex.html}
}}
\value{A \code{\link{layer}}}
\examples{\dontrun{
  # Coming soon
}}
\author{Hadley Wickham, \url{http://had.co.nz/}}
\keyword{hplot}
