\name{stat_boxplot}
\alias{stat_boxplot}
\alias{StatBoxplot}
\title{stat\_boxplot}
\description{Calculate components of box and whisker plot}
\details{
This page describes stat\_boxplot, see \code{\link{layer}} and \code{\link{qplot}} for how to create a complete plot from individual components.
}
\section{Aesthetics}{
The following aesthetics can be used with stat\_boxplot.  Aesthetics are mapped to variables in the data with the aes function: \code{stat\_boxplot(aes(x = var))}
\itemize{
  \item \code{x}: x position (\strong{required}) 
  \item \code{y}: y position (\strong{required}) 
}
}
\usage{stat_boxplot(mapping = NULL, data = NULL, geom = "boxplot", position = "dodge", 
    na.rm = FALSE, coef = 1.5, ...)}
\arguments{
 \item{mapping}{mapping between variables and aesthetics generated by aes}
 \item{data}{dataset used in this layer, if not specified uses plot dataset}
 \item{geom}{geometric used by this layer}
 \item{position}{position adjustment used by this layer}
 \item{na.rm}{NULL}
 \item{coef}{NULL}
 \item{...}{ignored }
}
\seealso{\itemize{
  \item \url{http://had.co.nz/ggplot2/stat_boxplot.html}
}}
\value{A \code{\link{layer}}}
\examples{\dontrun{
# See geom_boxplot for examples
}}
\author{Hadley Wickham, \url{http://had.co.nz/}}
\keyword{hplot}
