% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-chain.R
\name{stat_chain}
\alias{stat_chain}
\title{Chain statistic computation}
\usage{
stat_chain(
  mapping = NULL,
  data = NULL,
  geom = "path",
  position = "identity",
  ...,
  stats = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}}. If specified and \code{inherit.aes = TRUE} (the default), it is combined with the default mapping at the top level
of the plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three options:
\itemize{
\item \code{NULL} (default): the data is inherited from the plot data as specified
in the call to \code{\link[=ggplot]{ggplot()}}.
\item A \code{data.frame}, or other object, will override the plot data. All objects
will be fortified to produce a data frame. See \code{\link[=fortify]{fortify()}} for which
variables will be created.
\item A \code{function} will be called with a single argument, the plot data. The return
value must be a \code{data.frame}, and will be used as the layer data. A
\code{function} can be created from a \code{formula} (e.g. \code{~ head(.x, 10)}).
}}

\item{geom}{The geometric object to use to display the data for this layer. When using a
\verb{stat_*()} function to construct a layer, the \code{geom} argument can be used to
override the default coupling between stats and geoms. The \code{geom} argument
accepts the following:
\itemize{
\item A \code{Geom} ggproto subclass, for example \code{GeomPoint}.
\item A string naming the geom. To give the geom as a string, strip the
function name of the \code{geom_} prefix. For example, to use \code{geom_point()},
give the geom as \code{"point"}.
\item For more information and other ways to specify the geom, see the
\link[=layer_geoms]{layer geom} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This can be used in
various ways, including to prevent overplotting and improving the display.
The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example, to
use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[=layer_positions]{layer position} documentation.
}}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}'s \code{params} argument. These arguments
broadly fall into one of 4 categories below. Notably, further arguments to
the \code{position} argument, or aesthetics that are required can
\emph{not} be passed through \code{...}. Unknown arguments that are not part of the 4
categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"} or
\code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics} section
that lists the available options. The 'required' aesthetics cannot be
passed on to the \code{params}. Please note that while passing unmapped
aesthetics as vectors is technically possible, the order and required
length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation lists
which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[=layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[=draw_key]{key glyphs}, to change the display of the layer in the legend.
}}

\item{stats}{A character vector or list of statistical transformations to use
for this layer. Every element needs to be one of the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, to use \code{stat_count()},
give the stat as \code{"count"}.
\item The result of \code{\link[=link_stat]{link_stat()}} to pass parameters or mapping instructions.
}}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with a warning. If
\code{TRUE}, missing values are silently removed.}

\item{show.legend}{Logical. Should this layer be included in the legends? \code{NA}, the default,
includes if any aesthetics are mapped. \code{FALSE} never includes, and \code{TRUE}
always includes. It can also be a named logical vector to finely select the
aesthetics to display. To include legend keys for all levels, even when no
data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend, but
unobserved levels are omitted.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather than combining with
them. This is most useful for helper functions that define both data and
aesthetics and shouldn't inherit behaviour from the default plot
specification, e.g. \code{\link[=annotation_borders]{annotation_borders()}}.}
}
\description{
This statistic layer can take multiple stats and chain these together
to transform the data in a series of computations.
}
\details{
The procedure in which stats are chained are as follows. First, the
layer-level, undelayed aesthetics in the \code{mapping} argument are evaluated.
The data that results from that evaluation is passed to the first stat in
the \code{stats} argument to perform that stat's computation. If that first stat
is a \code{\link{link_stat}} with an \code{after.stat} component, the \code{after.stat} component
is evaluated before passing on the data to the next stat in the \code{stats}
argument. The next components in the \code{stats} argument work the same: the
data is passed on to compute the stat, then \code{after.stat} is evaluated. In
essence, the \code{after.stat} allows control over how computed variables are
passed to the next stat in the chain. Finally, once all components in the
\code{stats} arguments have been handled, the staged after stat components of
the layer-level \code{mapping} is evaluated. Per usual, the data are then handled
by the position and geom parts of a layer.
}
\examples{
p <- ggplot(mpg, aes(displ, colour = drv))
# Binning unique observations
p + stat_chain(stats = c("unique", "bin"))
# Controlling parameters
p + stat_chain(
  stats = list("unique", link_stat("bin", bins = 10))
)
# Evaluate expressions after computing stats
p + stat_chain(stats = list(
  link_stat("unique",  after.stat = aes(x = x + 1)),
  link_stat("density", after.stat = aes(y = density))
))
# Note that the last `after.stat` argument serves the same role as the
# `after_stat()` function in the layer mapping, so the following is
# equivalent to the previous plot
p + stat_chain(
  mapping = aes(y = after_stat(density)),
  stats = list(link_stat("unique", after.stat = aes(x = x + 1)), "density")
)
}
\seealso{
\code{\link[=link_stat]{link_stat()}}
}
