% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stat-vline.r
\name{stat_hline}
\alias{stat_hline}
\title{Add a horizontal line}
\usage{
stat_hline(mapping = NULL, data = NULL, geom = "hline",
  position = "identity", yintercept, ...)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link{aes}} or \code{\link{aes_string}}. Only needs to be set
at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override
the plot defaults.}

\item{geom}{The geometric object to use display the data}

\item{position}{The position adjustment to use for overlappling points
on this layer}

\item{...}{other arguments passed on to \code{\link{layer}}. This can
include aesthetics whose values you want to set, not map. See
\code{\link{layer}} for more details.}
}
\description{
Add a horizontal line
}
\examples{
# see geom_hline
}
\seealso{
\code{\link{geom_hline}} for code examples.
}
\keyword{internal}

