\name{StatIdentity}
\alias{stat_identity}
\alias{StatIdentity}
\title{stat_identity}
\description{Don't transform data}
\details{
This page describes \code{\link{stat_identity}}, see \code{\link{layer}} and \code{\link{qplot}} for how to create a complete plot from individual components.
}
\usage{stat_identity(...)}
\arguments{
 \item{...}{ignored }
}
\seealso{\itemize{
  \item \url{http://had.co.nz/ggplot/stat_identity.html}
}}
\value{A \code{\link{layer}}}
\examples{\dontrun{
    # Doesn't do anything, so hard to come up a useful example
}}
\author{Hadley Wickham, \url{http://had.co.nz/}}
\keyword{hplot}
