\name{stat_identity}
\alias{stat_identity}
\alias{StatIdentity}
\title{stat\_identity}
\description{Don't transform data}
\details{
This page describes stat\_identity, see \code{\link{layer}} and \code{\link{qplot}} for how to create a complete plot from individual components.
}
\usage{stat_identity(mapping=NULL, data=NULL, geom="point", position="identity", ...)}
\arguments{
 \item{mapping}{mapping between variables and aesthetics generated by aes}
 \item{data}{dataset used in this layer, if not specified uses plot dataset}
 \item{geom}{geometric used by this layer}
 \item{position}{position adjustment used by this layer}
 \item{...}{ignored }
}
\seealso{\itemize{
  \item \url{http://had.co.nz/ggplot2/stat_identity.html}
}}
\value{A \code{\link{layer}}}
\examples{\dontrun{
# Doesn't do anything, so hard to come up a useful example
}}
\author{Hadley Wickham, \url{http://had.co.nz/}}
\keyword{hplot}
