% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.r
\name{summary.ggplot}
\alias{summary.ggplot}
\title{Displays a useful description of a ggplot object}
\usage{
\method{summary}{ggplot}(object, ...)
}
\arguments{
\item{object}{ggplot2 object to summarise}

\item{...}{other arguments ignored (for compatibility with generic)}
}
\description{
Displays a useful description of a ggplot object
}
\examples{
p <- ggplot(mtcars, aes(mpg, wt)) +
  geom_point()
summary(p)
}
\keyword{internal}
