\name{theme_text}
\alias{theme_text}
\title{Theme element: text}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
This element adds text
}
\usage{theme_text(family = "", face = "plain", colour = "black", size = 10, hjust = 0.5, vjust = 0.5, angle = 0, lineheight = 1.1)}
\arguments{
\item{family}{font family}
\item{face}{font face ("plain", "italic", "bold")}
\item{colour}{text colour}
\item{size}{text size (in pts)}
\item{hjust}{horizontal justification (in [0, 1])}
\item{vjust}{vertical justification (in [0, 1])}
\item{angle}{angle (in [0, 360])}
\item{lineheight}{line height}
}


\seealso{\code{\link{textGrob}} for underlying grid function}


