% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/theme-elements.r
\name{theme_blank}
\alias{theme_blank}
\alias{theme_line}
\alias{theme_rect}
\alias{theme_segment}
\alias{theme_text}
\title{Deprecated theme_xx functions}
\usage{
theme_blank(...)

theme_rect(...)

theme_line(...)

theme_segment(...)

theme_text(...)
}
\arguments{
\item{...}{Arguments to be passed to the appropriate \code{element_xx}
  function.}
}
\description{
The \code{theme_xx} functions have been deprecated. They are replaced
with the \code{element_xx} functions.
}

