\name{theme_bw}
\alias{theme_bw}
\title{A theme with white background and black gridlines.}
\usage{
  theme_bw(base_size = 12, base_family = "Helvetica")
}
\arguments{
  \item{base_size}{base font size}

  \item{base_family}{base font family}
}
\description{
  A theme with white background and black gridlines.
}

