% Generated by roxygen2 (4.0.0): do not edit by hand
\name{theme_classic}
\alias{theme_classic}
\title{A classic-looking theme, with x and y axis lines and no gridlines.}
\usage{
theme_classic(base_size = 12, base_family = "")
}
\arguments{
  \item{base_size}{base font size}

  \item{base_family}{base font family}
}
\description{
A classic-looking theme, with x and y axis lines and no gridlines.
}
\seealso{
Other themes: \code{\link{theme_bw}};
  \code{\link{theme_gray}}, \code{\link{theme_grey}};
  \code{\link{theme_light}}; \code{\link{theme_linedraw}};
  \code{\link{theme_minimal}}
}

