\name{theme_grey}
\title{A theme with grey background and white gridlines.}
\usage{theme_grey(base_size = 12, base_family = "")}

\description{
  A theme with grey background and white gridlines.
}
\keyword{dplot}
\alias{theme_gray}
\alias{theme_grey}
\arguments{
  \item{base_size}{base font size}
  \item{base_family}{base font family}
}

