\name{theme_grey}
\alias{theme_gray}
\alias{theme_grey}
\title{A theme with grey background and white gridlines.}
\usage{
  theme_grey(base_size = 12, base_family = "Helvetica")
}
\arguments{
  \item{base_size}{base font size}

  \item{base_family}{base font family}
}
\description{
  A theme with grey background and white gridlines.
}

