% Generated by roxygen2 (4.0.0): do not edit by hand
\name{theme_light}
\alias{theme_light}
\title{A theme with white background and light grey lines}
\usage{
theme_light(base_size = 12, base_family = "")
}
\arguments{
  \item{base_size}{base font size}

  \item{base_family}{base font family}
}
\description{
A theme with white background and light grey lines
}
\seealso{
Other themes: \code{\link{theme_bw}};
  \code{\link{theme_classic}}; \code{\link{theme_gray}},
  \code{\link{theme_grey}}; \code{\link{theme_linedraw}};
  \code{\link{theme_minimal}}
}

