\name{theme_line}
\alias{theme_line}
\title{Theme element: line.}
\usage{
  theme_line(colour = "black", size = 0.5, linetype = 1)
}
\arguments{
  \item{colour}{line color}

  \item{size}{line size}

  \item{linetype}{line type}
}
\description{
  This element draws a line between two (or more) points
}
\seealso{
  \code{\link{polylineGrob}} for underlying grid function,
  \code{link{theme_segment}}
}
\keyword{dplot}

