% Generated by roxygen2 (4.0.0): do not edit by hand
\name{theme_linedraw}
\alias{theme_linedraw}
\title{A theme with white background and black gridlines.}
\usage{
theme_linedraw(base_size = 12, base_family = "")
}
\arguments{
  \item{base_size}{base font size}

  \item{base_family}{base font family}
}
\description{
A theme with white background and black gridlines.
}
\seealso{
Other themes: \code{\link{theme_bw}};
  \code{\link{theme_classic}}; \code{\link{theme_gray}},
  \code{\link{theme_grey}}; \code{\link{theme_light}};
  \code{\link{theme_minimal}}
}

