\name{theme_minimal}
\alias{theme_minimal}
\title{A minimalistic theme with no background annotations.}
\usage{
theme_minimal(base_size = 12, base_family = "")
}
\arguments{
  \item{base_size}{base font size}

  \item{base_family}{base font family}
}
\description{
A minimalistic theme with no background annotations.
}

