% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{txhousing}
\alias{txhousing}
\title{Housing sales in TX.}
\format{A data frame with 8602 observations and 9 variables:
\itemize{
\item{city}{Name of MLS area}
\item{year,month,date}{Date}
\item{sales}{Number of sales}
\item{volume}{Total value of sales}
\item{median}{Median sale price}
\item{listings}{Total active listings}
\item{inventory}{"Months inventory": amount of time it would take to sell
  all current listings at current pace of sales.}
}}
\usage{
txhousing
}
\description{
Information about the housing market in Texas provided by the TAMU
real estate center, \url{http://recenter.tamu.edu/}.
}
\keyword{datasets}

