% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-defaults.r
\name{update_geom_defaults}
\alias{update_geom_defaults}
\alias{update_stat_defaults}
\title{Modify geom/stat aesthetic defaults for future plots}
\usage{
update_geom_defaults(geom, new)

update_stat_defaults(stat, new)
}
\arguments{
\item{new}{Named list of aesthetics. Alternatively, \code{NULL} to reset the
defaults.}

\item{stat, geom}{Name of geom/stat to modify (like \code{"point"} or
\code{"bin"}), or a Geom/Stat object (like \code{GeomPoint} or
\code{StatBin}).}
}
\description{
Modify geom/stat aesthetic defaults for future plots
}
\examples{
update_geom_defaults("point", list(colour = "darkblue"))
ggplot(mtcars, aes(mpg, wt)) + geom_point()
# Reset defaults by using `new = NULL`
update_geom_defaults("point", NULL)
}
\keyword{internal}
