% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geom-defaults.r
\name{update_geom_defaults}
\alias{update_geom_defaults}
\alias{update_stat_defaults}
\title{Modify geom/stat aesthetic defaults for future plots}
\usage{
update_geom_defaults(geom, new)

update_stat_defaults(stat, new)
}
\arguments{
\item{new}{named list of aesthetics}

\item{stat,geom}{name of geom/stat to modify}
}
\description{
Modify geom/stat aesthetic defaults for future plots
}
\examples{
update_geom_defaults("point", list(colour = "darkblue"))
qplot(mpg, wt, data = mtcars)
update_geom_defaults("point", list(colour = "black"))
}

