\name{vp_path}
\alias{vp_path}
\title{Viewport path}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Calculate viewport path.
}
\usage{vp_path(row, col, type)}
\arguments{
\item{row}{row index}
\item{col}{column index}
\item{type}{viewport type}
}

\details{Convience method for calculating the viewport path to a particular
entry in a matrix viewport.  This helps ensure a common naming scheme throughout
ggplot/}

\examples{}
\keyword{hplot}
\keyword{internal}
