% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{waiver}
\alias{waiver}
\alias{is.waiver}
\title{A waiver object.}
\usage{
waiver()

is.waiver(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
A waiver is a "flag" object, similar to \code{NULL}, that indicates the
calling function should just use the default value.  It is used in certain
functions to distinguish between displaying nothing (\code{NULL}) and
displaying a default value calculated elsewhere (\code{waiver()}).
\code{is.waiver()} reports whether an object is a waiver.
}
\keyword{internal}
