% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{waiver}
\alias{waiver}
\title{A waiver object.}
\usage{
waiver()
}
\description{
A waiver is a "flag" object, similar to \code{NULL}, that indicates the
calling function should just use the default value.  It is used in certain
functions to distinguish between displaying nothing (\code{NULL}) and
displaying a default value calculated elsewhere (\code{waiver()})
}
\keyword{internal}

