% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with-sf.R
\name{with_sf}
\alias{with_sf}
\title{Turn regular layers into sf layers}
\usage{
with_sf(..., crs = 4326)
}
\arguments{
\item{...}{Layer}

\item{crs}{Coordinate reference system of the origin layer. Defaults to
longitude and latitude in WGS84 (EPSG:4326), the World Geodetic System
from 1984.}
}
\description{
Turn regular layers into sf layers
}
