\name{wrap_hmisc}
\alias{wrap_hmisc}
\alias{sum_mean_cl_boot}
\alias{sum_mean_cl_normal}
\alias{sum_mean_sdl}
\alias{sum_median_hilow}
\alias{sum_range}
\alias{mean_cl_boot}
\alias{mean_cl_normal}
\alias{mean_sdl}
\alias{median_hilow}
\title{Wrap HMisc summary functions to make it easy to use them with stat_summary}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{

}
\usage{wrap_hmisc(x, fun, ...)}
\arguments{
\item{x}{}
\item{fun}{}
\item{...}{}
}

\details{}

\examples{}

