\name{wrap_hmisc}
\alias{wrap_hmisc}
\alias{mean_cl_boot}
\alias{mean_cl_normal}
\alias{mean_sdl}
\alias{median_hilow}
\title{Wrap Hmisc summary functions}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Wrap up a selection of Hmisc to make it easy to use with \code{\link{stat_summary}}
}
\usage{wrap_hmisc(fun)}
\arguments{
\item{fun}{}
}

\details{See the Hmisc documentation for details of their options.}
\seealso{\code{\link[Hmisc]{smean.cl.boot}}, \code{\link[Hmisc]{smean.cl.normal}}, \code{\link[Hmisc]{smean.sdl}}, \code{\link[Hmisc]{smedian.hilon}}}
\examples{}
\keyword{internal}
